Bug 1077136 - fix builds4hr tests and improve tasks per feedback
