+++ tests/etl/test_buildapi.py
@@ -12,20 +12,7 @@ def mock_buildapi_pending_url(monkeypatch):
     path = os.path.join(
         tests_folder,
         "sample_data",
-        "builds-pending.js"
-    )
-    monkeypatch.setattr(settings,
-                        'BUILDAPI_PENDING_URL',
-                        "file://{0}".format(path))
-
-
-@pytest.fixture
-def mock_buildapi_pending_missing1_url(monkeypatch):
-    tests_folder = os.path.dirname(os.path.dirname(__file__))
-    path = os.path.join(
-        tests_folder,
-        "sample_data",
-        "builds-pending-missing1.js"
+        "builds-pending.json"
     )
     monkeypatch.setattr(settings,
                         'BUILDAPI_PENDING_URL',
@@ -38,7 +25,7 @@ def mock_buildapi_running_url(monkeypatch):
     path = os.path.join(
         tests_folder,
         "sample_data",
-        "builds-running.js"
+        "builds-running.json"
     )
     monkeypatch.setattr(settings,
                         'BUILDAPI_RUNNING_URL',
@@ -51,7 +38,46 @@ def mock_buildapi_builds4h_url(monkeypatch):
     path = os.path.join(
         tests_folder,
         "sample_data",
-        "builds-4h.js"
+        "builds-4h.json"
+    )
+    monkeypatch.setattr(settings,
+                        'BUILDAPI_BUILDS4H_URL',
+                        "file://{0}".format(path))
+
+
+@pytest.fixture
+def mock_buildapi_pending_missing1_url(monkeypatch):
+    tests_folder = os.path.dirname(os.path.dirname(__file__))
+    path = os.path.join(
+        tests_folder,
+        "sample_data",
+        "builds-pending-missing1.json"
+    )
+    monkeypatch.setattr(settings,
+                        'BUILDAPI_PENDING_URL',
+                        "file://{0}".format(path))
+
+
+@pytest.fixture
+def mock_buildapi_running_missing1_url(monkeypatch):
+    tests_folder = os.path.dirname(os.path.dirname(__file__))
+    path = os.path.join(
+        tests_folder,
+        "sample_data",
+        "builds-running-missing1.json"
+    )
+    monkeypatch.setattr(settings,
+                        'BUILDAPI_RUNNING_URL',
+                        "file://{0}".format(path))
+
+
+@pytest.fixture
+def mock_buildapi_builds4h_missing1_url(monkeypatch):
+    tests_folder = os.path.dirname(os.path.dirname(__file__))
+    path = os.path.join(
+        tests_folder,
+        "sample_data",
+        "builds-4h-missing1.json"
     )
     monkeypatch.setattr(settings,
                         'BUILDAPI_BUILDS4H_URL',
@@ -79,16 +105,112 @@ def test_ingest_pending_jobs(jm, initial_data,
     assert len(stored_obj) == 1
 
 
-def test_ingest_pending_jobs_1_missing_resultset(jm, initial_data, sample_resultset,
-                                test_repository,
-                                mock_buildapi_pending_missing1_url,
+def test_ingest_running_jobs(jm, initial_data,
+                                mock_buildapi_running_url,
                                 mock_post_json_data,
+                                mock_log_parser,
                                 mock_get_resultset,
-                                mock_get_remote_content,
-                                activate_responses):
+                                mock_get_remote_content):
     """
-    Ensure the job with the missing resultset is queued for refetching
+    a new buildapi running job creates a new obj in the job table
     """
+    from treeherder.etl.buildapi import RunningJobsProcess
+    etl_process = RunningJobsProcess()
+    etl_process.run()
+
+    stored_obj = jm.get_jobs_dhub().execute(
+        proc="jobs_test.selects.jobs")
+
+    jm.disconnect()
+
+    assert len(stored_obj) == 1
+
+
+def test_ingest_builds4h_jobs(jm, initial_data,
+                                mock_buildapi_builds4h_url,
+                                mock_post_json_data,
+                                mock_log_parser,
+                                mock_get_resultset,
+                                mock_get_remote_content):
+    """
+    a new buildapi completed job creates a new obj in the job table
+    """
+    from treeherder.etl.buildapi import Builds4hJobsProcess
+    etl_process = Builds4hJobsProcess()
+    etl_process.run()
+    jm.process_objects(20)
+
+    stored_obj = jm.get_jobs_dhub().execute(
+        proc="jobs_test.selects.jobs")
+
+    jm.disconnect()
+
+    assert len(stored_obj) == 20
+
+
+def test_ingest_running_job_fields(jm, initial_data,
+                                mock_buildapi_running_url,
+                                mock_post_json_data,
+                                mock_log_parser,
+                                mock_get_resultset,
+                                mock_get_remote_content):
+    """
+    a new buildapi running job creates a new obj in the job table
+    """
+    from treeherder.etl.buildapi import RunningJobsProcess
+    etl_process = RunningJobsProcess()
+    etl_process.run()
+
+    stored_obj = jm.get_jobs_dhub().execute(
+        proc="jobs_test.selects.jobs")
+
+    jm.disconnect()
+
+    assert len(stored_obj) == 1
+    assert stored_obj[0]["start_timestamp"] is not 0
+
+#####################
+# MISSING RESULTSETS
+#####################
+
+
+def test_ingest_pending_jobs_1_missing_resultset(jm, initial_data,
+        sample_resultset, test_repository, mock_buildapi_pending_missing1_url,
+        mock_post_json_data, mock_get_resultset, mock_get_remote_content,
+        activate_responses):
+    """
+    Ensure the pending job with the missing resultset is queued for refetching
+    """
+    from treeherder.etl.buildapi import PendingJobsProcess
+    etl_process = PendingJobsProcess()
+    _do_missing_resultset_test(jm, etl_process)
+
+
+def test_ingest_running_jobs_1_missing_resultset(jm, initial_data,
+        sample_resultset, test_repository, mock_buildapi_running_missing1_url,
+        mock_post_json_data, mock_get_resultset, mock_get_remote_content,
+        activate_responses):
+    """
+    Ensure the running job with the missing resultset is queued for refetching
+    """
+    from treeherder.etl.buildapi import RunningJobsProcess
+    etl_process = RunningJobsProcess()
+    _do_missing_resultset_test(jm, etl_process)
+
+
+def test_ingest_builds4h_jobs_1_missing_resultset(jm, initial_data,
+        sample_resultset, test_repository, mock_buildapi_builds4h_missing1_url,
+        mock_post_json_data, mock_get_resultset, mock_get_remote_content,
+        activate_responses):
+    """
+    Ensure the builds4h job with the missing resultset is queued for refetching
+    """
+    from treeherder.etl.buildapi import Builds4hJobsProcess
+    etl_process = Builds4hJobsProcess()
+    _do_missing_resultset_test(jm, etl_process)
+
+
+def _do_missing_resultset_test(jm, etl_process):
     new_revision = '222222222222'
     pushlog_content = json.dumps(
         {"33270": {
@@ -114,9 +236,8 @@ def test_ingest_pending_jobs_1_missing_resultset(jm, initial_data, sample_result
                   match_querystring=True,
                   content_type='application/json')
 
-    from treeherder.etl.buildapi import PendingJobsProcess
-    etl_process = PendingJobsProcess()
     etl_process.run()
+    jm.process_objects(2)
 
     stored_obj = jm.get_jobs_dhub().execute(
         proc="jobs_test.selects.jobs")
@@ -134,49 +255,4 @@ def test_ingest_pending_jobs_1_missing_resultset(jm, initial_data, sample_result
         if str(rs['revision']) == new_revision:
             was_stored = True
     assert was_stored
-
-    jm.disconnect()
-
-
-def test_ingest_running_jobs(jm, initial_data,
-                                mock_buildapi_running_url,
-                                mock_post_json_data,
-                                mock_log_parser,
-                                mock_get_resultset,
-                                mock_get_remote_content):
-    """
-    a new buildapi running job creates a new obj in the job table
-    """
-    from treeherder.etl.buildapi import RunningJobsProcess
-    etl_process = RunningJobsProcess()
-    etl_process.run()
-
-    stored_obj = jm.get_jobs_dhub().execute(
-        proc="jobs_test.selects.jobs")
-
-    jm.disconnect()
-
-    assert len(stored_obj) == 1
-
-
-def test_ingest_running_job_fields(jm, initial_data,
-                                mock_buildapi_running_url,
-                                mock_post_json_data,
-                                mock_log_parser,
-                                mock_get_resultset,
-                                mock_get_remote_content):
-    """
-    a new buildapi running job creates a new obj in the job table
-    """
-    from treeherder.etl.buildapi import RunningJobsProcess
-    etl_process = RunningJobsProcess()
-    etl_process.run()
-
-    stored_obj = jm.get_jobs_dhub().execute(
-        proc="jobs_test.selects.jobs")
-
-    jm.disconnect()
-
-    assert len(stored_obj) == 1
-    assert stored_obj[0]["start_timestamp"] is not 0
