--- tests/sample_data/builds-4h.js
@@ -1,5270 +1,0 @@
-{
-  "builds": [
-    {
-      "builder_id": 177027,
-      "buildnumber": 5,
-      "endtime": 1376510970,
-      "id": 26553075,
-      "master_id": 81,
-      "properties": {
-        "basedir": "c:\\slave\\test",
-        "branch": "test_treeherder",
-        "build_url": "http://ftp.mozilla.org/pub/mozilla.org/firefox/try-builds/jcoppeard@mozilla.com-a0a0b3d23040/try-win64/firefox-26.0a1.en-US.win64-x86_64.zip",
-        "buildername": "win64_vm try opt test mochitest-other",
-        "buildid": "20130813092104",
-        "buildnumber": 5,
-        "builduid": "fb381cf2d050416cbf1566db838b1099",
-        "log_url": "http://ftp.mozilla.org/pub/mozilla.org/firefox/try-builds/jcoppeard@mozilla.com-a0a0b3d23040/try-win64/try_win64_vm_test-mochitest-other-bm72-tests1-windows-build5.txt.gz",
-        "master": "http://buildbot-master72.srv.releng.usw2.mozilla.com:8201/",
-        "platform": "win64",
-        "product": "firefox",
-        "project": "",
-        "repo_path": "try",
-        "repository": "",
-        "request_ids": [
-          27949897
-        ],
-        "request_times": {
-          "27949897": 1376416924
-        },
-        "revision": "a0a0b3d23040",
-        "scheduler": "tests-try-win64_vm-opt-unittest",
-        "script_repo_revision": "8d8602c57d3e",
-        "slavebuilddir": "test",
-        "slavename": "tst-w64-ec2-002",
-        "stage_platform": "win64"
-      },
-      "reason": "scheduler",
-      "request_ids": [
-        27949897
-      ],
-      "requesttime": 1376416924,
-      "result": 1,
-      "slave_id": 5021,
-      "starttime": 1376422786
-    },
-    {
-      "builder_id": 144581,
-      "buildnumber": 1467,
-      "endtime": 1376511335,
-      "id": 26553258,
-      "master_id": 75,
-      "properties": {
-        "appName": "Firefox",
-        "appVersion": "26.0a1",
-        "basedir": "/builds/slave/try-osx64-d-000000000000000000",
-        "branch": "test_treeherder",
-        "builddir": "try-osx64-d-000000000000000000",
-        "buildername": "OS X 10.7 64-bit try leak test build",
-        "buildid": "20130814082959",
-        "buildnumber": 1467,
-        "builduid": "caa1cec65c6844c18035ea32a2fa7f7b",
-        "comments": "try: -b d -p linux64,macosx64,win32 -u all[x64] -t none\n#try: -b o -p linux64,macosx64,win32 -u all[6.1] -t none\n#try: -b o -p win32 -u none -t none\n#try: -b do -p win32,linux64,macosx64,android -u all -t none\n#try: -b d -p macosx64 -u all -t none\n#pacing fs 6 try: -b do -p all -u all[win32] -t none\n#nss sync fs nss and psm serial 7 try: -b do -p win32,linux64,macosx64 -u all -t none\n#pacing fs 3 try: -b do -p linux64,win32 -u all -t none\n#try: -b do -p linux64,android,android-armv6,android-noion -u all -t none\n#try: -b do -p win32,linux64,macosx64,android -u all -t none\n#try: -b d -p macosx64 -u all -t none\n#nsitimer fully threadsafe try: -b do -p win32,linux64,macosx64 -u all -t tpn\n#try: -b o -p win32 -u none -t tpn\n#try: -b d -p linux -u mochitest-4 -t none\n#try: -b d -p linux -u all -t none\n#try: -b do -p linux64,win32 -u all -t none\n#try: -b d -p win32,linux64,macosx64 -u all -t none\n#try: -b d -p linux64 -u all -t none\n#try: -b o -p win32,linux,linux64,macosx64 -u none -t none\n#try: -b od -p macosx64 -",
-        "filepath": null,
-        "forced_clobber": false,
-        "got_revision": "982a89285fbd",
-        "hashType": "sha512",
-        "jsshellUrl": "http://ftp.mozilla.org/pub/mozilla.org/firefox/try-builds/mcmanus@ducksong.com-982a89285fbd/try-macosx64-debug/jsshell-mac64.zip",
-        "log_url": "http://ftp.mozilla.org/pub/mozilla.org/firefox/try-builds/mcmanus@ducksong.com-982a89285fbd/try-macosx64-debug/try-macosx64-debug-bm55-try1-build1467.txt.gz",
-        "master": "http://buildbot-master55.srv.releng.use1.mozilla.com:8101/",
-        "packageFilename": "firefox-26.0a1.en-US.mac64.dmg",
-        "packageHash": "a0f78751479b599125981541a982374a598f5ce146d9b169fc23fb5e84ec1470228c14df65d4c6af3353bfb439cebb6a8d771f292e1c29319d0019f13c2301a4",
-        "packageSize": "41701875",
-        "packageUrl": "http://ftp.mozilla.org/pub/mozilla.org/firefox/try-builds/mcmanus@ducksong.com-982a89285fbd/try-macosx64-debug/firefox-26.0a1.en-US.mac64.dmg",
-        "periodic_clobber": false,
-        "platform": "macosx64-debug",
-        "product": "firefox",
-        "project": "",
-        "purge_actual": "534.83GB",
-        "purge_target": "10GB",
-        "purged_clobber": false,
-        "repository": "",
-        "request_ids": [
-          27997895
-        ],
-        "request_times": {
-          "27997895": 1376494162
-        },
-        "revision": "982a89285fbd599786e7bdf2d4ec9ec25e08d330",
-        "scheduler": "try-firefox",
-        "slavebuilddir": "try-osx64-d-000000000000000000",
-        "slavename": "bld-lion-r5-017",
-        "stage_platform": "macosx64-debug",
-        "symbolsUrl": "http://ftp.mozilla.org/pub/mozilla.org/firefox/try-builds/mcmanus@ducksong.com-982a89285fbd/try-macosx64-debug/firefox-26.0a1.en-US.mac64.crashreporter-symbols.zip",
-        "testsUrl": "http://ftp.mozilla.org/pub/mozilla.org/firefox/try-builds/mcmanus@ducksong.com-982a89285fbd/try-macosx64-debug/firefox-26.0a1.en-US.mac64.tests.zip",
-        "toolsdir": "/builds/slave/try-osx64-d-000000000000000000/tools",
-        "who": "mcmanus@ducksong.com"
-      },
-      "reason": "scheduler",
-      "request_ids": [
-        27997895
-      ],
-      "requesttime": 1376494162,
-      "result": 1,
-      "slave_id": 2034,
-      "starttime": 1376494174
-    },
-    {
-      "builder_id": 144793,
-      "buildnumber": 34,
-      "endtime": 1376516293,
-      "id": 26555820,
-      "master_id": 86,
-      "properties": {
-        "appName": "Firefox",
-        "appVersion": "26.0a1",
-        "basedir": "e:/builds/moz2_slave/m-in-w32-pgo-00000000000000000",
-        "branch": "test_treeherder",
-        "builddir": "m-in-w32-pgo-00000000000000000",
-        "buildername": "WINNT 5.2 mozilla-inbound pgo-build",
-        "buildid": "20130814102526",
-        "buildnumber": 34,
-        "builduid": "8d99f0315dec42d89af091bf44b59ed5",
-        "comments": "",
-        "filepath": null,
-        "forced_clobber": false,
-        "got_revision": "355dcff578b6",
-        "hashType": "sha512",
-        "installerFilename": "firefox-26.0a1.en-US.win32.installer.exe",
-        "installerHash": "9d80a3197fd02d26b8b3fb04dc2c6d57e088f71dd98d3bd8ca887a6dc636128e072f70da6516d78fe0daba4c444a04ce591fbb6ed6c82e7f4555b59ba1c64ae2",
-        "installerSize": "26663984",
-        "jsshellUrl": "http://ftp.mozilla.org/pub/mozilla.org/firefox/tinderbox-builds/mozilla-inbound-win32-pgo/1376501126/jsshell-win32.zip",
-        "log_url": "http://ftp.mozilla.org/pub/mozilla.org/firefox/tinderbox-builds/mozilla-inbound-win32-pgo/1376501126/mozilla-inbound-win32-pgo-bm63-build1-build34.txt.gz",
-        "master": "http://buildbot-master63.srv.releng.use1.mozilla.com:8001/",
-        "packageFilename": "firefox-26.0a1.en-US.win32.zip",
-        "packageHash": "1793d2fe3977506cf4f7c74541bbb3744a0c29d00f80382114b5491c3739e5b526792a99995546d7cfc830371c1bea59089f3c050aed85f88511fbf8095ce695",
-        "packageSize": "35345481",
-        "packageUrl": "http://ftp.mozilla.org/pub/mozilla.org/firefox/tinderbox-builds/mozilla-inbound-win32-pgo/1376501126/firefox-26.0a1.en-US.win32.zip",
-        "periodic_clobber": false,
-        "platform": "win32",
-        "product": "firefox",
-        "project": "",
-        "purge_actual": "113.47GB",
-        "purge_target": "12GB",
-        "purged_clobber": true,
-        "repository": "",
-        "request_ids": [
-          28003842
-        ],
-        "request_times": {
-          "28003842": 1376501127
-        },
-        "revision": "355dcff578b6",
-        "slavebuilddir": "m-in-w32-pgo-00000000000000000",
-        "slavename": "w64-ix-slave18",
-        "sourcestamp": "355dcff578b6",
-        "stage_platform": "win32-pgo",
-        "symbolsUrl": "http://ftp.mozilla.org/pub/mozilla.org/firefox/tinderbox-builds/mozilla-inbound-win32-pgo/1376501126/firefox-26.0a1.en-US.win32.crashreporter-symbols.zip",
-        "testresults": [
-          [
-            "libxul_link",
-            "libxul_link",
-            3365036032,
-            "3365036032"
-          ]
-        ],
-        "testsUrl": "http://ftp.mozilla.org/pub/mozilla.org/firefox/tinderbox-builds/mozilla-inbound-win32-pgo/1376501126/firefox-26.0a1.en-US.win32.tests.zip",
-        "toolsdir": "e:/builds/moz2_slave/m-in-w32-pgo-00000000000000000/tools",
-        "vsize": 3365036032
-      },
-      "reason": "Self-serve: Requested by rvandermeulen@mozilla.com",
-      "request_ids": [
-        28003842
-      ],
-      "requesttime": 1376501127,
-      "result": 0,
-      "slave_id": 1323,
-      "starttime": 1376501182
-    },
-    {
-      "builder_id": 143785,
-      "buildnumber": 485,
-      "endtime": 1376511587,
-      "id": 26553385,
-      "master_id": 100,
-      "properties": {
-        "appName": "Firefox",
-        "appVersion": "26.0a1",
-        "basedir": "e:/builds/moz2_slave/m-in-w32-d-0000000000000000000",
-        "branch": "test_treeherder",
-        "builddir": "m-in-w32-d-0000000000000000000",
-        "buildername": "WINNT 5.2 mozilla-inbound leak test build",
-        "buildid": "20130814104121",
-        "buildnumber": 485,
-        "builduid": "56d1370353394fc487c082b6c96af781",
-        "comments": "No bug - Disable timeout-prone B2G crashtests.\nCLOSED TREE",
-        "filepath": null,
-        "forced_clobber": false,
-        "got_revision": "9df04b16a655",
-        "hashType": "sha512",
-        "installerFilename": "firefox-26.0a1.en-US.win32.installer.exe",
-        "installerHash": "1d15e90a3c9233769cde8d2443b6c2643147975a8e98e7c763ebe32fc619a5de4819e59e1ce9f1cf4bc765c85768c071e2357b8a557ef56d548aaf0602a24e9d",
-        "installerSize": "29664768",
-        "jsshellUrl": "http://ftp.mozilla.org/pub/mozilla.org/firefox/tinderbox-builds/mozilla-inbound-win32-debug/1376502081/jsshell-win32.zip",
-        "log_url": "http://ftp.mozilla.org/pub/mozilla.org/firefox/tinderbox-builds/mozilla-inbound-win32-debug/1376502081/mozilla-inbound-win32-debug-bm65-build1-build485.txt.gz",
-        "master": "http://buildbot-master65.srv.releng.usw2.mozilla.com:8001/",
-        "packageFilename": "firefox-26.0a1.en-US.win32.zip",
-        "packageHash": "c54a36027c9a2987602960d08f9693b15dc91e462969cdbbaf76a191856b96e04ad89e46d44f525589520c4f61c0e1282cdf86bc804b317e0d50c2248385a502",
-        "packageSize": "41377368",
-        "packageUrl": "http://ftp.mozilla.org/pub/mozilla.org/firefox/tinderbox-builds/mozilla-inbound-win32-debug/1376502081/firefox-26.0a1.en-US.win32.zip",
-        "periodic_clobber": false,
-        "platform": "win32-debug",
-        "product": "firefox",
-        "project": "",
-        "purge_actual": "93.23GB",
-        "purge_target": "9GB",
-        "purged_clobber": false,
-        "repository": "",
-        "request_ids": [
-          28004256
-        ],
-        "request_times": {
-          "28004256": 1376502082
-        },
-        "revision": "9df04b16a655fda9aa8615eb946a1ab62b697e14",
-        "scheduler": "mozilla-inbound-firefox",
-        "slavebuilddir": "m-in-w32-d-0000000000000000000",
-        "slavename": "w64-ix-slave88",
-        "stage_platform": "win32-debug",
-        "symbolsUrl": "http://ftp.mozilla.org/pub/mozilla.org/firefox/tinderbox-builds/mozilla-inbound-win32-debug/1376502081/firefox-26.0a1.en-US.win32.crashreporter-symbols.zip",
-        "testsUrl": "http://ftp.mozilla.org/pub/mozilla.org/firefox/tinderbox-builds/mozilla-inbound-win32-debug/1376502081/firefox-26.0a1.en-US.win32.tests.zip",
-        "toolsdir": "e:/builds/moz2_slave/m-in-w32-d-0000000000000000000/tools"
-      },
-      "reason": "scheduler",
-      "request_ids": [
-        28004256
-      ],
-      "requesttime": 1376502082,
-      "result": 0,
-      "slave_id": 2455,
-      "starttime": 1376502091
-    },
-    {
-      "builder_id": 146267,
-      "buildnumber": 42,
-      "endtime": 1376517826,
-      "id": 26556531,
-      "master_id": 86,
-      "properties": {
-        "appName": "Firefox",
-        "appVersion": "25.0a2",
-        "basedir": "e:/builds/moz2_slave/m-aurora-w32-00000000000000000",
-        "branch": "test_treeherder",
-        "builddir": "m-aurora-w32-00000000000000000",
-        "buildername": "WINNT 5.2 mozilla-aurora build",
-        "buildid": "20130814104621",
-        "buildnumber": 42,
-        "builduid": "f377570e7b214e2185ad6a995fde8820",
-        "comments": "Bug 899095 - Switch from http://build.mozilla.org/talos to  http://talos-bundles.pvt.b.m.o for talos r=jmaher - a=testing",
-        "filepath": null,
-        "forced_clobber": false,
-        "got_revision": "932868752f82",
-        "hashType": "sha512",
-        "installerFilename": "firefox-25.0a2.en-US.win32.installer.exe",
-        "installerHash": "9cd666c45821d44954e7227cce95e8f612463fd11b7d7b11ba6c2f60245abd109df4ad5a0c1e5230f3388974d819a2c5cdc1262fb5164df0f7b3ab76787a499a",
-        "installerSize": "23609688",
-        "jsshellUrl": "http://ftp.mozilla.org/pub/mozilla.org/firefox/tinderbox-builds/mozilla-aurora-win32/1376502381/jsshell-win32.zip",
-        "log_url": "http://ftp.mozilla.org/pub/mozilla.org/firefox/tinderbox-builds/mozilla-aurora-win32/1376502381/mozilla-aurora-win32-bm63-build1-build42.txt.gz",
-        "master": "http://buildbot-master63.srv.releng.use1.mozilla.com:8001/",
-        "packageFilename": "firefox-25.0a2.en-US.win32.zip",
-        "packageHash": "b897fc1226e4077fb7e3558438f5b220afa54619dfbf15c48c4922949a2ab0becbdbcd18ae36c606d73d44b0fdeb7d47cf2fef2cb6c2728232a5fdc40cf2d7a5",
-        "packageSize": "30159299",
-        "packageUrl": "http://ftp.mozilla.org/pub/mozilla.org/firefox/tinderbox-builds/mozilla-aurora-win32/1376502381/firefox-25.0a2.en-US.win32.zip",
-        "periodic_clobber": false,
-        "platform": "win32",
-        "product": "firefox",
-        "project": "",
-        "purge_actual": "109.48GB",
-        "purge_target": "12GB",
-        "purged_clobber": false,
-        "repository": "",
-        "request_ids": [
-          28004430
-        ],
-        "request_times": {
-          "28004430": 1376502382
-        },
-        "revision": "932868752f82a0256638a0bd0375699b96e346f5",
-        "scheduler": "mozilla-aurora-firefox",
-        "slavebuilddir": "m-aurora-w32-00000000000000000",
-        "slavename": "w64-ix-slave133",
-        "stage_platform": "win32",
-        "symbolsUrl": "http://ftp.mozilla.org/pub/mozilla.org/firefox/tinderbox-builds/mozilla-aurora-win32/1376502381/firefox-25.0a2.en-US.win32.crashreporter-symbols.zip",
-        "testsUrl": "http://ftp.mozilla.org/pub/mozilla.org/firefox/tinderbox-builds/mozilla-aurora-win32/1376502381/firefox-25.0a2.en-US.win32.tests.zip",
-        "toolsdir": "e:/builds/moz2_slave/m-aurora-w32-00000000000000000/tools"
-      },
-      "reason": "scheduler",
-      "request_ids": [
-        28004430
-      ],
-      "requesttime": 1376502382,
-      "result": 0,
-      "slave_id": 4459,
-      "starttime": 1376502386
-    },
-    {
-      "builder_id": 144677,
-      "buildnumber": 1313,
-      "endtime": 1376511244,
-      "id": 26553217,
-      "master_id": 75,
-      "properties": {
-        "appName": "Firefox",
-        "appVersion": "26.0a1",
-        "basedir": "e:/builds/moz2_slave/try-w32-d-00000000000000000000",
-        "branch": "test_treeherder",
-        "builddir": "try-w32-d-00000000000000000000",
-        "buildername": "WINNT 5.2 try leak test build",
-        "buildid": "20130814104704",
-        "buildnumber": 1313,
-        "builduid": "ba54de47161c422db105a19653b4f7ca",
-        "comments": "try: -b do -p linux,macosx64,win32 -u reftest,crashtest,mochitests -t none",
-        "filepath": null,
-        "forced_clobber": false,
-        "got_revision": "1a867785f92b",
-        "hashType": "sha512",
-        "installerFilename": "firefox-26.0a1.en-US.win32.installer.exe",
-        "installerHash": "d650fd4e71b53667eddd6cc5273c43ef7cff6ed7c47b03baa1b39b530e1c29023b8cd1c294d618c9b55b71c8c67d8a50f8e18f8310cec595d6db8d8d07f4185c",
-        "installerSize": "29632000",
-        "jsshellUrl": "http://ftp.mozilla.org/pub/mozilla.org/firefox/try-builds/dholbert@mozilla.com-1a867785f92b/try-win32-debug/jsshell-win32.zip",
-        "log_url": "http://ftp.mozilla.org/pub/mozilla.org/firefox/try-builds/dholbert@mozilla.com-1a867785f92b/try-win32-debug/try-win32-debug-bm55-try1-build1313.txt.gz",
-        "master": "http://buildbot-master55.srv.releng.use1.mozilla.com:8101/",
-        "packageFilename": "firefox-26.0a1.en-US.win32.zip",
-        "packageHash": "2483a7be67a90078de7dac7fe44980946c9cb37c086abd1829d34ab44947175d19d6f160a27062a9fbcefc5b99241e452bd7e23252439062212f4c7d951b84e8",
-        "packageSize": "41382695",
-        "packageUrl": "http://ftp.mozilla.org/pub/mozilla.org/firefox/try-builds/dholbert@mozilla.com-1a867785f92b/try-win32-debug/firefox-26.0a1.en-US.win32.zip",
-        "periodic_clobber": false,
-        "platform": "win32-debug",
-        "product": "firefox",
-        "project": "",
-        "purge_actual": "146.21GB",
-        "purge_target": "9GB",
-        "purged_clobber": false,
-        "repository": "",
-        "request_ids": [
-          28004432
-        ],
-        "request_times": {
-          "28004432": 1376502383
-        },
-        "revision": "1a867785f92b8d29b0bf112a7391684495c851d4",
-        "scheduler": "try-firefox",
-        "slavebuilddir": "try-w32-d-00000000000000000000",
-        "slavename": "w64-ix-slave29",
-        "stage_platform": "win32-debug",
-        "symbolsUrl": "http://ftp.mozilla.org/pub/mozilla.org/firefox/try-builds/dholbert@mozilla.com-1a867785f92b/try-win32-debug/firefox-26.0a1.en-US.win32.crashreporter-symbols.zip",
-        "testsUrl": "http://ftp.mozilla.org/pub/mozilla.org/firefox/try-builds/dholbert@mozilla.com-1a867785f92b/try-win32-debug/firefox-26.0a1.en-US.win32.tests.zip",
-        "toolsdir": "e:/builds/moz2_slave/try-w32-d-00000000000000000000/tools",
-        "who": "dholbert@mozilla.com"
-      },
-      "reason": "scheduler",
-      "request_ids": [
-        28004432
-      ],
-      "requesttime": 1376502383,
-      "result": 0,
-      "slave_id": 1873,
-      "starttime": 1376502397
-    },
-    {
-      "builder_id": 132178,
-      "buildnumber": 63,
-      "endtime": 1376515463,
-      "id": 26555458,
-      "master_id": 86,
-      "properties": {
-        "appName": "Firefox",
-        "appVersion": "25.0a2",
-        "basedir": "/builds/slave/m-aurora-lx-000000000000000000",
-        "branch": "test_treeherder",
-        "builddir": "m-aurora-lx-000000000000000000",
-        "buildername": "Linux mozilla-aurora build",
-        "buildid": "20130814104621",
-        "buildnumber": 63,
-        "builduid": "f377570e7b214e2185ad6a995fde8820",
-        "comments": "Bug 899095 - Switch from http://build.mozilla.org/talos to  http://talos-bundles.pvt.b.m.o for talos r=jmaher - a=testing",
-        "filepath": null,
-        "forced_clobber": false,
-        "got_revision": "932868752f82",
-        "hashType": "sha512",
-        "jsshellUrl": "http://ftp.mozilla.org/pub/mozilla.org/firefox/tinderbox-builds/mozilla-aurora-linux/1376502381/jsshell-linux-i686.zip",
-        "log_url": "http://ftp.mozilla.org/pub/mozilla.org/firefox/tinderbox-builds/mozilla-aurora-linux/1376502381/mozilla-aurora-linux-bm63-build1-build63.txt.gz",
-        "master": "http://buildbot-master63.srv.releng.use1.mozilla.com:8001/",
-        "num_ctors": 223,
-        "packageFilename": "firefox-25.0a2.en-US.linux-i686.tar.bz2",
-        "packageHash": "a87077316d24ba416d4ee6ad7792690835bba1bc67db3fd69d7c7909f0255dfd91f2ede50e554c90717d54c47c1a1c332a1d9db80c84776e8ac3e88a638dfd1f",
-        "packageSize": "34656741",
-        "packageUrl": "http://ftp.mozilla.org/pub/mozilla.org/firefox/tinderbox-builds/mozilla-aurora-linux/1376502381/firefox-25.0a2.en-US.linux-i686.tar.bz2",
-        "periodic_clobber": false,
-        "platform": "linux",
-        "product": "firefox",
-        "project": "",
-        "purge_actual": "112.67GB",
-        "purge_target": "12GB",
-        "purged_clobber": true,
-        "repository": "",
-        "request_ids": [
-          28004424
-        ],
-        "request_times": {
-          "28004424": 1376502382
-        },
-        "revision": "932868752f82a0256638a0bd0375699b96e346f5",
-        "scheduler": "mozilla-aurora-firefox",
-        "slavebuilddir": "m-aurora-lx-000000000000000000",
-        "slavename": "bld-linux64-ec2-048",
-        "sourcestamp": "932868752f82",
-        "stage_platform": "linux",
-        "symbolsUrl": "http://ftp.mozilla.org/pub/mozilla.org/firefox/tinderbox-builds/mozilla-aurora-linux/1376502381/firefox-25.0a2.en-US.linux-i686.crashreporter-symbols.zip",
-        "testresults": [
-          [
-            "num_ctors",
-            "num_ctors",
-            223,
-            "223"
-          ]
-        ],
-        "testsUrl": "http://ftp.mozilla.org/pub/mozilla.org/firefox/tinderbox-builds/mozilla-aurora-linux/1376502381/firefox-25.0a2.en-US.linux-i686.tests.zip",
-        "toolsdir": "/builds/slave/m-aurora-lx-000000000000000000/tools"
-      },
-      "reason": "scheduler",
-      "request_ids": [
-        28004424
-      ],
-      "requesttime": 1376502382,
-      "result": 0,
-      "slave_id": 2303,
-      "starttime": 1376502399
-    },
-    {
-      "builder_id": 137564,
-      "buildnumber": 48,
-      "endtime": 1376514783,
-      "id": 26555110,
-      "master_id": 77,
-      "properties": {
-        "appName": "Firefox",
-        "appVersion": "25.0a2",
-        "basedir": "/builds/slave/m-aurora-l64-00000000000000000",
-        "branch": "test_treeherder",
-        "builddir": "m-aurora-l64-00000000000000000",
-        "buildername": "Linux x86-64 mozilla-aurora build",
-        "buildid": "20130814104621",
-        "buildnumber": 48,
-        "builduid": "f377570e7b214e2185ad6a995fde8820",
-        "comments": "Bug 899095 - Switch from http://build.mozilla.org/talos to  http://talos-bundles.pvt.b.m.o for talos r=jmaher - a=testing",
-        "filepath": null,
-        "forced_clobber": false,
-        "got_revision": "932868752f82",
-        "hashType": "sha512",
-        "jsshellUrl": "http://ftp.mozilla.org/pub/mozilla.org/firefox/tinderbox-builds/mozilla-aurora-linux64/1376502381/jsshell-linux-x86_64.zip",
-        "log_url": "http://ftp.mozilla.org/pub/mozilla.org/firefox/tinderbox-builds/mozilla-aurora-linux64/1376502381/mozilla-aurora-linux64-bm58-build1-build48.txt.gz",
-        "master": "http://buildbot-master58.srv.releng.usw2.mozilla.com:8001/",
-        "num_ctors": 223,
-        "packageFilename": "firefox-25.0a2.en-US.linux-x86_64.tar.bz2",
-        "packageHash": "4c5754e82baabb0833fb9fe061dd11e95fa858cd166a731d259acd00b33b4600e05466b9c8af613784e376f5a4ca1c2e15de4a8e99215482695fecba90126076",
-        "packageSize": "34834570",
-        "packageUrl": "http://ftp.mozilla.org/pub/mozilla.org/firefox/tinderbox-builds/mozilla-aurora-linux64/1376502381/firefox-25.0a2.en-US.linux-x86_64.tar.bz2",
-        "periodic_clobber": false,
-        "platform": "linux64",
-        "product": "firefox",
-        "project": "",
-        "purge_actual": "145.23GB",
-        "purge_target": "14GB",
-        "purged_clobber": true,
-        "repository": "",
-        "request_ids": [
-          28004426
-        ],
-        "request_times": {
-          "28004426": 1376502382
-        },
-        "revision": "932868752f82a0256638a0bd0375699b96e346f5",
-        "scheduler": "mozilla-aurora-firefox",
-        "slavebuilddir": "m-aurora-l64-00000000000000000",
-        "slavename": "bld-linux64-ec2-479",
-        "sourcestamp": "932868752f82",
-        "stage_platform": "linux64",
-        "symbolsUrl": "http://ftp.mozilla.org/pub/mozilla.org/firefox/tinderbox-builds/mozilla-aurora-linux64/1376502381/firefox-25.0a2.en-US.linux-x86_64.crashreporter-symbols.zip",
-        "testresults": [
-          [
-            "num_ctors",
-            "num_ctors",
-            223,
-            "223"
-          ]
-        ],
-        "testsUrl": "http://ftp.mozilla.org/pub/mozilla.org/firefox/tinderbox-builds/mozilla-aurora-linux64/1376502381/firefox-25.0a2.en-US.linux-x86_64.tests.zip",
-        "toolsdir": "/builds/slave/m-aurora-l64-00000000000000000/tools"
-      },
-      "reason": "scheduler",
-      "request_ids": [
-        28004426
-      ],
-      "requesttime": 1376502382,
-      "result": 0,
-      "slave_id": 4355,
-      "starttime": 1376502471
-    },
-    {
-      "builder_id": 132814,
-      "buildnumber": 166,
-      "endtime": 1376510965,
-      "id": 26553070,
-      "master_id": 90,
-      "properties": {
-        "basedir": "/builds/slave/talos-slave/test",
-        "branch": "test_treeherder",
-        "build_url": "http://ftp.mozilla.org/pub/mozilla.org/firefox/tinderbox-builds/fx-team-macosx64-debug/1376499083/firefox-26.0a1.en-US.mac64.dmg",
-        "buildername": "Rev4 MacOSX Lion 10.7 fx-team debug test mochitest-browser-chrome",
-        "buildid": "20130814095123",
-        "buildnumber": 166,
-        "builduid": "e8c4556f1c404e6786b4d1e975cdf9c5",
-        "log_url": "http://ftp.mozilla.org/pub/mozilla.org/firefox/tinderbox-builds/fx-team-macosx64-debug/1376499083/fx-team_lion-debug_test-mochitest-browser-chrome-bm79-tests1-macosx-build166.txt.gz",
-        "master": "http://buildbot-master79.srv.releng.usw2.mozilla.com:8201/",
-        "pgo_build": "False",
-        "platform": "macosx64",
-        "product": "firefox",
-        "project": "",
-        "repo_path": "integration/fx-team",
-        "repository": "",
-        "request_ids": [
-          28005070
-        ],
-        "request_times": {
-          "28005070": 1376502662
-        },
-        "revision": "15aaeae89597",
-        "scheduler": "tests-fx-team-lion-debug-unittest",
-        "script_repo_revision": "2f85052176f2",
-        "slavebuilddir": "test",
-        "slavename": "talos-r4-lion-020",
-        "stage_platform": "macosx64",
-        "symbols_url": "http://ftp.mozilla.org/pub/mozilla.org/firefox/tinderbox-builds/fx-team-macosx64-debug/1376499083/firefox-26.0a1.en-US.mac64.crashreporter-symbols.zip"
-      },
-      "reason": "scheduler",
-      "request_ids": [
-        28005070
-      ],
-      "requesttime": 1376502662,
-      "result": 0,
-      "slave_id": 1627,
-      "starttime": 1376503174
-    },
-    {
-      "builder_id": 129034,
-      "buildnumber": 161,
-      "endtime": 1376511090,
-      "id": 26553130,
-      "master_id": 90,
-      "properties": {
-        "basedir": "/Users/cltbld/talos-slave/test",
-        "branch": "test_treeherder",
-        "build_url": "http://ftp.mozilla.org/pub/mozilla.org/firefox/tinderbox-builds/fx-team-macosx64-debug/1376499083/firefox-26.0a1.en-US.mac64.dmg",
-        "buildername": "Rev4 MacOSX Snow Leopard 10.6 fx-team debug test mochitest-browser-chrome",
-        "buildid": "20130814095123",
-        "buildnumber": 161,
-        "builduid": "e8c4556f1c404e6786b4d1e975cdf9c5",
-        "log_url": "http://ftp.mozilla.org/pub/mozilla.org/firefox/tinderbox-builds/fx-team-macosx64-debug/1376499083/fx-team_snowleopard-debug_test-mochitest-browser-chrome-bm79-tests1-macosx-build161.txt.gz",
-        "master": "http://buildbot-master79.srv.releng.usw2.mozilla.com:8201/",
-        "pgo_build": "False",
-        "platform": "macosx64",
-        "product": "firefox",
-        "project": "",
-        "repo_path": "integration/fx-team",
-        "repository": "",
-        "request_ids": [
-          28004999
-        ],
-        "request_times": {
-          "28004999": 1376502655
-        },
-        "revision": "15aaeae89597",
-        "scheduler": "tests-fx-team-snowleopard-debug-unittest",
-        "script_repo_revision": "2f85052176f2",
-        "slavebuilddir": "test",
-        "slavename": "talos-r4-snow-024",
-        "stage_platform": "macosx64",
-        "symbols_url": "http://ftp.mozilla.org/pub/mozilla.org/firefox/tinderbox-builds/fx-team-macosx64-debug/1376499083/firefox-26.0a1.en-US.mac64.crashreporter-symbols.zip"
-      },
-      "reason": "scheduler",
-      "request_ids": [
-        28004999
-      ],
-      "requesttime": 1376502655,
-      "result": 0,
-      "slave_id": 1524,
-      "starttime": 1376503489
-    },
-    {
-      "builder_id": 140326,
-      "buildnumber": 439,
-      "endtime": 1376511019,
-      "id": 26553102,
-      "master_id": 71,
-      "properties": {
-        "basedir": "/home/cltbld/talos-slave/test",
-        "branch": "test_treeherder",
-        "build_url": "http://ftp.mozilla.org/pub/mozilla.org/firefox/try-builds/hurley@mozilla.com-f55b1a3d65a4/try-linux-debug/firefox-26.0a1.en-US.linux-i686.tar.bz2",
-        "buildername": "Rev3 Fedora 12 try debug test mochitest-browser-chrome",
-        "buildid": "20130814092605",
-        "buildnumber": 439,
-        "builduid": "738259f2cb844166af2e37e8e7e07bda",
-        "log_url": "http://ftp.mozilla.org/pub/mozilla.org/firefox/try-builds/hurley@mozilla.com-f55b1a3d65a4/try-linux-debug/try_fedora-debug_test-mochitest-browser-chrome-bm51-tests1-linux-build439.txt.gz",
-        "master": "http://buildbot-master51.srv.releng.use1.mozilla.com:8201/",
-        "platform": "linux",
-        "product": "firefox",
-        "project": "",
-        "repo_path": "try",
-        "repository": "",
-        "request_ids": [
-          28006073
-        ],
-        "request_times": {
-          "28006073": 1376503805
-        },
-        "revision": "f55b1a3d65a4",
-        "scheduler": "tests-try-fedora-debug-unittest",
-        "script_repo_revision": "2f85052176f2",
-        "slavebuilddir": "test",
-        "slavename": "talos-r3-fed-028",
-        "stage_platform": "linux",
-        "symbols_url": "http://ftp.mozilla.org/pub/mozilla.org/firefox/try-builds/hurley@mozilla.com-f55b1a3d65a4/try-linux-debug/firefox-26.0a1.en-US.linux-i686.crashreporter-symbols.zip"
-      },
-      "reason": "scheduler",
-      "request_ids": [
-        28006073
-      ],
-      "requesttime": 1376503805,
-      "result": 1,
-      "slave_id": 440,
-      "starttime": 1376503812
-    },
-    {
-      "builder_id": 116117,
-      "buildnumber": 435,
-      "endtime": 1376510916,
-      "id": 26553044,
-      "master_id": 73,
-      "properties": {
-        "basedir": "/builds/slave/test",
-        "branch": "test_treeherder",
-        "build_url": "http://ftp.mozilla.org/pub/mozilla.org/firefox/try-builds/hurley@mozilla.com-f55b1a3d65a4/try-linux-debug/firefox-26.0a1.en-US.linux-i686.tar.bz2",
-        "buildername": "Ubuntu VM 12.04 try debug test reftest",
-        "buildid": "20130814092605",
-        "buildnumber": 435,
-        "builduid": "738259f2cb844166af2e37e8e7e07bda",
-        "log_url": "http://ftp.mozilla.org/pub/mozilla.org/firefox/try-builds/hurley@mozilla.com-f55b1a3d65a4/try-linux-debug/try_ubuntu32_vm-debug_test-reftest-bm54-tests1-linux-build435.txt.gz",
-        "master": "http://buildbot-master54.srv.releng.usw2.mozilla.com:8201/",
-        "platform": "linux",
-        "product": "firefox",
-        "project": "",
-        "repo_path": "try",
-        "repository": "",
-        "request_ids": [
-          28006074
-        ],
-        "request_times": {
-          "28006074": 1376503805
-        },
-        "revision": "f55b1a3d65a4",
-        "scheduler": "tests-try-ubuntu32_vm-debug-unittest",
-        "script_repo_revision": "2f85052176f2",
-        "slavebuilddir": "test",
-        "slavename": "tst-linux32-ec2-315",
-        "stage_platform": "linux",
-        "symbols_url": "http://ftp.mozilla.org/pub/mozilla.org/firefox/try-builds/hurley@mozilla.com-f55b1a3d65a4/try-linux-debug/firefox-26.0a1.en-US.linux-i686.crashreporter-symbols.zip"
-      },
-      "reason": "scheduler",
-      "request_ids": [
-        28006074
-      ],
-      "requesttime": 1376503805,
-      "result": 1,
-      "slave_id": 3586,
-      "starttime": 1376503864
-    },
-    {
-      "builder_id": 119956,
-      "buildnumber": 1012,
-      "endtime": 1376511677,
-      "id": 26553425,
-      "master_id": 82,
-      "properties": {
-        "basedir": "/builds/slave/talos-slave/test",
-        "branch": "test_treeherder",
-        "build_url": "http://ftp.mozilla.org/pub/mozilla.org/firefox/tinderbox-builds/mozilla-inbound-macosx64-debug/1376500461/firefox-26.0a1.en-US.mac64.dmg",
-        "buildername": "Rev4 MacOSX Lion 10.7 mozilla-inbound debug test mochitest-browser-chrome",
-        "buildid": "20130814101421",
-        "buildnumber": 1012,
-        "builduid": "5e66a9f140074af7b494fe2c7c1d54af",
-        "log_url": "http://ftp.mozilla.org/pub/mozilla.org/firefox/tinderbox-builds/mozilla-inbound-macosx64-debug/1376500461/mozilla-inbound_lion-debug_test-mochitest-browser-chrome-bm75-tests1-macosx-build1012.txt.gz",
-        "master": "http://buildbot-master75.srv.releng.use1.mozilla.com:8201/",
-        "pgo_build": "False",
-        "platform": "macosx64",
-        "product": "firefox",
-        "project": "",
-        "repo_path": "integration/mozilla-inbound",
-        "repository": "",
-        "request_ids": [
-          28005220,
-          28006065
-        ],
-        "request_times": {
-          "28005220": 1376502844,
-          "28006065": 1376503803
-        },
-        "revision": "1fa1f28c6edf",
-        "scheduler": "tests-mozilla-inbound-lion-debug-unittest",
-        "script_repo_revision": "2f85052176f2",
-        "slavebuilddir": "test",
-        "slavename": "talos-r4-lion-046",
-        "stage_platform": "macosx64",
-        "symbols_url": "http://ftp.mozilla.org/pub/mozilla.org/firefox/tinderbox-builds/mozilla-inbound-macosx64-debug/1376500461/firefox-26.0a1.en-US.mac64.crashreporter-symbols.zip"
-      },
-      "reason": "scheduler",
-      "request_ids": [
-        28005220,
-        28006065
-      ],
-      "requesttime": 1376502844,
-      "result": 0,
-      "slave_id": 1632,
-      "starttime": 1376503892
-    },
-    {
-      "builder_id": 119896,
-      "buildnumber": 1100,
-      "endtime": 1376511518,
-      "id": 26553358,
-      "master_id": 82,
-      "properties": {
-        "basedir": "/Users/cltbld/talos-slave/test",
-        "branch": "test_treeherder",
-        "build_url": "http://ftp.mozilla.org/pub/mozilla.org/firefox/tinderbox-builds/mozilla-inbound-macosx64-debug/1376500461/firefox-26.0a1.en-US.mac64.dmg",
-        "buildername": "Rev4 MacOSX Snow Leopard 10.6 mozilla-inbound debug test mochitest-browser-chrome",
-        "buildid": "20130814101421",
-        "buildnumber": 1100,
-        "builduid": "5e66a9f140074af7b494fe2c7c1d54af",
-        "log_url": "http://ftp.mozilla.org/pub/mozilla.org/firefox/tinderbox-builds/mozilla-inbound-macosx64-debug/1376500461/mozilla-inbound_snowleopard-debug_test-mochitest-browser-chrome-bm75-tests1-macosx-build1100.txt.gz",
-        "master": "http://buildbot-master75.srv.releng.use1.mozilla.com:8201/",
-        "pgo_build": "False",
-        "platform": "macosx64",
-        "product": "firefox",
-        "project": "",
-        "repo_path": "integration/mozilla-inbound",
-        "repository": "",
-        "request_ids": [
-          28005207,
-          28006052
-        ],
-        "request_times": {
-          "28005207": 1376502844,
-          "28006052": 1376503803
-        },
-        "revision": "1fa1f28c6edf",
-        "scheduler": "tests-mozilla-inbound-snowleopard-debug-unittest",
-        "script_repo_revision": "2f85052176f2",
-        "slavebuilddir": "test",
-        "slavename": "talos-r4-snow-082",
-        "stage_platform": "macosx64",
-        "symbols_url": "http://ftp.mozilla.org/pub/mozilla.org/firefox/tinderbox-builds/mozilla-inbound-macosx64-debug/1376500461/firefox-26.0a1.en-US.mac64.crashreporter-symbols.zip"
-      },
-      "reason": "scheduler",
-      "request_ids": [
-        28005207,
-        28006052
-      ],
-      "requesttime": 1376502844,
-      "result": 1,
-      "slave_id": 1846,
-      "starttime": 1376503927
-    },
-    {
-      "builder_id": 141353,
-      "buildnumber": 373,
-      "endtime": 1376511001,
-      "id": 26553092,
-      "master_id": 80,
-      "properties": {
-        "basedir": "/home/cltbld/talos-slave/test",
-        "branch": "test_treeherder",
-        "build_url": "http://ftp.mozilla.org/pub/mozilla.org/firefox/try-builds/jcoppeard@mozilla.com-7cd16ff70711/try-linux-debug/firefox-26.0a1.en-US.linux-i686.tar.bz2",
-        "buildername": "Rev3 Fedora 12 try debug test mochitest-browser-chrome",
-        "buildid": "20130814093237",
-        "buildnumber": 373,
-        "builduid": "6d441a71ff4d49809832c35bf9613c82",
-        "log_url": "http://ftp.mozilla.org/pub/mozilla.org/firefox/try-builds/jcoppeard@mozilla.com-7cd16ff70711/try-linux-debug/try_fedora-debug_test-mochitest-browser-chrome-bm67-tests1-linux-build373.txt.gz",
-        "master": "http://buildbot-master67.srv.releng.use1.mozilla.com:8201/",
-        "platform": "linux",
-        "product": "firefox",
-        "project": "",
-        "repo_path": "try",
-        "repository": "",
-        "request_ids": [
-          28006094
-        ],
-        "request_times": {
-          "28006094": 1376503926
-        },
-        "revision": "7cd16ff70711",
-        "scheduler": "tests-try-fedora-debug-unittest",
-        "script_repo_revision": "2f85052176f2",
-        "slavebuilddir": "test",
-        "slavename": "talos-r3-fed-011",
-        "stage_platform": "linux",
-        "symbols_url": "http://ftp.mozilla.org/pub/mozilla.org/firefox/try-builds/jcoppeard@mozilla.com-7cd16ff70711/try-linux-debug/firefox-26.0a1.en-US.linux-i686.crashreporter-symbols.zip"
-      },
-      "reason": "scheduler",
-      "request_ids": [
-        28006094
-      ],
-      "requesttime": 1376503926,
-      "result": 0,
-      "slave_id": 461,
-      "starttime": 1376503936
-    },
-    {
-      "builder_id": 150974,
-      "buildnumber": 1,
-      "endtime": 1376514420,
-      "id": 26554872,
-      "master_id": 100,
-      "properties": {
-        "basedir": "/e/builds/moz2_slave/rel-m-rel-w32_rpk_3-0000000000",
-        "branch": "test_treeherder",
-        "build_number": 1,
-        "builddir": "release-mozilla-release-win32_repack_3",
-        "buildername": "release-mozilla-release-win32_repack_3/10",
-        "buildid": "20130814063812",
-        "buildnumber": 1,
-        "builduid": "df292bc2d0934dd69bb8182448919107",
-        "locales": "el,en-GB,en-ZA,eo,es-AR,es-CL,es-ES,es-MX,et",
-        "log_url": "http://stage.mozilla.org/pub/mozilla.org/firefox/nightly/23.0.1-candidates/build1/logs/release-mozilla-release-win32_repack_3-bm65-build1-build1.txt.gz",
-        "master": "http://buildbot-master65.srv.releng.usw2.mozilla.com:8001/",
-        "platform": "win32",
-        "product": "firefox",
-        "products": "firefox",
-        "project": "",
-        "release_config": "mozilla/release-firefox-mozilla-release.py",
-        "repository": "",
-        "request_ids": [
-          28006144
-        ],
-        "request_times": {
-          "28006144": 1376503997
-        },
-        "revision": "15aaeae89597",
-        "scheduler": "release-mozilla-release-win32_repack",
-        "script_repo_revision": "97080c57a9ab",
-        "slavebuilddir": "rel-m-rel-w32_rpk_3-0000000000",
-        "slavename": "w64-ix-slave73",
-        "toolsdir": "/e/builds/moz2_slave/rel-m-rel-w32_rpk_3-0000000000/scripts",
-        "version": "23.0.1"
-      },
-      "reason": "Triggerable(release-mozilla-release-win32_repack)",
-      "request_ids": [
-        28006144
-      ],
-      "requesttime": 1376503997,
-      "result": 0,
-      "slave_id": 2147,
-      "starttime": 1376503997
-    },
-    {
-      "builder_id": 154329,
-      "buildnumber": 2,
-      "endtime": 1376513734,
-      "id": 26554459,
-      "master_id": 100,
-      "properties": {
-        "basedir": "/e/builds/moz2_slave/rel-m-rel-w32_rpk_2-0000000000",
-        "branch": "test_treeherder",
-        "build_number": 1,
-        "builddir": "release-mozilla-release-win32_repack_2",
-        "buildername": "release-mozilla-release-win32_repack_2/10",
-        "buildid": "20130814063812",
-        "buildnumber": 2,
-        "builduid": "df292bc2d0934dd69bb8182448919107",
-        "locales": "bn-IN,br,bs,ca,cs,csb,cy,da,de",
-        "log_url": "http://stage.mozilla.org/pub/mozilla.org/firefox/nightly/23.0.1-candidates/build1/logs/release-mozilla-release-win32_repack_2-bm65-build1-build2.txt.gz",
-        "master": "http://buildbot-master65.srv.releng.usw2.mozilla.com:8001/",
-        "platform": "win32",
-        "product": "firefox",
-        "products": "firefox",
-        "project": "",
-        "release_config": "mozilla/release-firefox-mozilla-release.py",
-        "repository": "",
-        "request_ids": [
-          28006143
-        ],
-        "request_times": {
-          "28006143": 1376503997
-        },
-        "revision": "15aaeae89597",
-        "scheduler": "release-mozilla-release-win32_repack",
-        "script_repo_revision": "97080c57a9ab",
-        "slavebuilddir": "rel-m-rel-w32_rpk_2-0000000000",
-        "slavename": "w64-ix-slave122",
-        "toolsdir": "/e/builds/moz2_slave/rel-m-rel-w32_rpk_2-0000000000/scripts",
-        "version": "23.0.1"
-      },
-      "reason": "Triggerable(release-mozilla-release-win32_repack)",
-      "request_ids": [
-        28006143
-      ],
-      "requesttime": 1376503997,
-      "result": 2,
-      "slave_id": 4467,
-      "starttime": 1376503997
-    },
-    {
-      "builder_id": 154470,
-      "buildnumber": 1,
-      "endtime": 1376514615,
-      "id": 26554997,
-      "master_id": 100,
-      "properties": {
-        "basedir": "/e/builds/moz2_slave/rel-m-rel-w32_rpk_1-0000000000",
-        "branch": "test_treeherder",
-        "build_number": 1,
-        "builddir": "release-mozilla-release-win32_repack_1",
-        "buildername": "release-mozilla-release-win32_repack_1/10",
-        "buildid": "20130814063812",
-        "buildnumber": 1,
-        "builduid": "df292bc2d0934dd69bb8182448919107",
-        "locales": "ach,af,ak,ar,as,ast,be,bg,bn-BD",
-        "log_url": "http://stage.mozilla.org/pub/mozilla.org/firefox/nightly/23.0.1-candidates/build1/logs/release-mozilla-release-win32_repack_1-bm65-build1-build1.txt.gz",
-        "master": "http://buildbot-master65.srv.releng.usw2.mozilla.com:8001/",
-        "platform": "win32",
-        "product": "firefox",
-        "products": "firefox",
-        "project": "",
-        "release_config": "mozilla/release-firefox-mozilla-release.py",
-        "repository": "",
-        "request_ids": [
-          28006142
-        ],
-        "request_times": {
-          "28006142": 1376503997
-        },
-        "revision": "15aaeae89597",
-        "scheduler": "release-mozilla-release-win32_repack",
-        "script_repo_revision": "97080c57a9ab",
-        "slavebuilddir": "rel-m-rel-w32_rpk_1-0000000000",
-        "slavename": "w64-ix-slave126",
-        "toolsdir": "/e/builds/moz2_slave/rel-m-rel-w32_rpk_1-0000000000/scripts",
-        "version": "23.0.1"
-      },
-      "reason": "Triggerable(release-mozilla-release-win32_repack)",
-      "request_ids": [
-        28006142
-      ],
-      "requesttime": 1376503997,
-      "result": 0,
-      "slave_id": 4460,
-      "starttime": 1376503997
-    },
-    {
-      "builder_id": 150967,
-      "buildnumber": 2,
-      "endtime": 1376514745,
-      "id": 26555079,
-      "master_id": 95,
-      "properties": {
-        "basedir": "/e/builds/moz2_slave/rel-m-rel-w32_rpk_6-0000000000",
-        "branch": "test_treeherder",
-        "build_number": 1,
-        "builddir": "release-mozilla-release-win32_repack_6",
-        "buildername": "release-mozilla-release-win32_repack_6/10",
-        "buildid": "20130814063812",
-        "buildnumber": 2,
-        "builduid": "df292bc2d0934dd69bb8182448919107",
-        "locales": "ja,kk,km,kn,ko,ku,lg,lij,lt",
-        "log_url": "http://stage.mozilla.org/pub/mozilla.org/firefox/nightly/23.0.1-candidates/build1/logs/release-mozilla-release-win32_repack_6-bm66-build1-build2.txt.gz",
-        "master": "http://buildbot-master66.srv.releng.usw2.mozilla.com:8001/",
-        "platform": "win32",
-        "product": "firefox",
-        "products": "firefox",
-        "project": "",
-        "release_config": "mozilla/release-firefox-mozilla-release.py",
-        "repository": "",
-        "request_ids": [
-          28006147
-        ],
-        "request_times": {
-          "28006147": 1376503997
-        },
-        "revision": "15aaeae89597",
-        "scheduler": "release-mozilla-release-win32_repack",
-        "script_repo_revision": "97080c57a9ab",
-        "slavebuilddir": "rel-m-rel-w32_rpk_6-0000000000",
-        "slavename": "w64-ix-slave116",
-        "toolsdir": "/e/builds/moz2_slave/rel-m-rel-w32_rpk_6-0000000000/scripts",
-        "version": "23.0.1"
-      },
-      "reason": "Triggerable(release-mozilla-release-win32_repack)",
-      "request_ids": [
-        28006147
-      ],
-      "requesttime": 1376503997,
-      "result": 0,
-      "slave_id": 4164,
-      "starttime": 1376504000
-    },
-    {
-      "builder_id": 181859,
-      "buildnumber": 0,
-      "endtime": 1376514500,
-      "id": 26554926,
-      "master_id": 95,
-      "properties": {
-        "basedir": "/e/builds/moz2_slave/rel-m-rel-w32_rpk_7-0000000000",
-        "branch": "test_treeherder",
-        "build_number": 1,
-        "builddir": "release-mozilla-release-win32_repack_7",
-        "buildername": "release-mozilla-release-win32_repack_7/10",
-        "buildid": "20130814063812",
-        "buildnumber": 0,
-        "builduid": "df292bc2d0934dd69bb8182448919107",
-        "locales": "lv,mai,mk,ml,mr,nb-NO,nl,nn-NO,nso",
-        "log_url": "http://stage.mozilla.org/pub/mozilla.org/firefox/nightly/23.0.1-candidates/build1/logs/release-mozilla-release-win32_repack_7-bm66-build1-build0.txt.gz",
-        "master": "http://buildbot-master66.srv.releng.usw2.mozilla.com:8001/",
-        "platform": "win32",
-        "product": "firefox",
-        "products": "firefox",
-        "project": "",
-        "release_config": "mozilla/release-firefox-mozilla-release.py",
-        "repository": "",
-        "request_ids": [
-          28006148
-        ],
-        "request_times": {
-          "28006148": 1376503997
-        },
-        "revision": "15aaeae89597",
-        "scheduler": "release-mozilla-release-win32_repack",
-        "script_repo_revision": "97080c57a9ab",
-        "slavebuilddir": "rel-m-rel-w32_rpk_7-0000000000",
-        "slavename": "w64-ix-slave112",
-        "toolsdir": "/e/builds/moz2_slave/rel-m-rel-w32_rpk_7-0000000000/scripts",
-        "version": "23.0.1"
-      },
-      "reason": "Triggerable(release-mozilla-release-win32_repack)",
-      "request_ids": [
-        28006148
-      ],
-      "requesttime": 1376503997,
-      "result": 0,
-      "slave_id": 4157,
-      "starttime": 1376504000
-    },
-    {
-      "builder_id": 150971,
-      "buildnumber": 1,
-      "endtime": 1376514161,
-      "id": 26554705,
-      "master_id": 99,
-      "properties": {
-        "basedir": "/e/builds/moz2_slave/rel-m-rel-w32_rpk_5-0000000000",
-        "branch": "test_treeherder",
-        "build_number": 1,
-        "builddir": "release-mozilla-release-win32_repack_5",
-        "buildername": "release-mozilla-release-win32_repack_5/10",
-        "buildid": "20130814063812",
-        "buildnumber": 1,
-        "builduid": "df292bc2d0934dd69bb8182448919107",
-        "locales": "gu-IN,he,hi-IN,hr,hu,hy-AM,id,is,it",
-        "log_url": "http://stage.mozilla.org/pub/mozilla.org/firefox/nightly/23.0.1-candidates/build1/logs/release-mozilla-release-win32_repack_5-bm62-build1-build1.txt.gz",
-        "master": "http://buildbot-master62.srv.releng.use1.mozilla.com:8001/",
-        "platform": "win32",
-        "product": "firefox",
-        "products": "firefox",
-        "project": "",
-        "release_config": "mozilla/release-firefox-mozilla-release.py",
-        "repository": "",
-        "request_ids": [
-          28006146
-        ],
-        "request_times": {
-          "28006146": 1376503997
-        },
-        "revision": "15aaeae89597",
-        "scheduler": "release-mozilla-release-win32_repack",
-        "script_repo_revision": "97080c57a9ab",
-        "slavebuilddir": "rel-m-rel-w32_rpk_5-0000000000",
-        "slavename": "w64-ix-slave13",
-        "toolsdir": "/e/builds/moz2_slave/rel-m-rel-w32_rpk_5-0000000000/scripts",
-        "version": "23.0.1"
-      },
-      "reason": "Triggerable(release-mozilla-release-win32_repack)",
-      "request_ids": [
-        28006146
-      ],
-      "requesttime": 1376503997,
-      "result": 0,
-      "slave_id": 1318,
-      "starttime": 1376504000
-    },
-    {
-      "builder_id": 181856,
-      "buildnumber": 0,
-      "endtime": 1376513757,
-      "id": 26554471,
-      "master_id": 95,
-      "properties": {
-        "basedir": "/e/builds/moz2_slave/rel-m-rel-w32_rpk_9-0000000000",
-        "branch": "test_treeherder",
-        "build_number": 1,
-        "builddir": "release-mozilla-release-win32_repack_9",
-        "buildername": "release-mozilla-release-win32_repack_9/10",
-        "buildid": "20130814063812",
-        "buildnumber": 0,
-        "builduid": "df292bc2d0934dd69bb8182448919107",
-        "locales": "sk,sl,son,sq,sr,sv-SE,ta,ta-LK",
-        "log_url": "http://stage.mozilla.org/pub/mozilla.org/firefox/nightly/23.0.1-candidates/build1/logs/release-mozilla-release-win32_repack_9-bm66-build1-build0.txt.gz",
-        "master": "http://buildbot-master66.srv.releng.usw2.mozilla.com:8001/",
-        "platform": "win32",
-        "product": "firefox",
-        "products": "firefox",
-        "project": "",
-        "release_config": "mozilla/release-firefox-mozilla-release.py",
-        "repository": "",
-        "request_ids": [
-          28006150
-        ],
-        "request_times": {
-          "28006150": 1376503997
-        },
-        "revision": "15aaeae89597",
-        "scheduler": "release-mozilla-release-win32_repack",
-        "script_repo_revision": "97080c57a9ab",
-        "slavebuilddir": "rel-m-rel-w32_rpk_9-0000000000",
-        "slavename": "w64-ix-slave75",
-        "toolsdir": "/e/builds/moz2_slave/rel-m-rel-w32_rpk_9-0000000000/scripts",
-        "version": "23.0.1"
-      },
-      "reason": "Triggerable(release-mozilla-release-win32_repack)",
-      "request_ids": [
-        28006150
-      ],
-      "requesttime": 1376503997,
-      "result": 0,
-      "slave_id": 2118,
-      "starttime": 1376504000
-    },
-    {
-      "builder_id": 181857,
-      "buildnumber": 0,
-      "endtime": 1376514096,
-      "id": 26554666,
-      "master_id": 95,
-      "properties": {
-        "basedir": "/e/builds/moz2_slave/rel-m-rel-w32_rpk_8-0000000000",
-        "branch": "test_treeherder",
-        "build_number": 1,
-        "builddir": "release-mozilla-release-win32_repack_8",
-        "buildername": "release-mozilla-release-win32_repack_8/10",
-        "buildid": "20130814063812",
-        "buildnumber": 0,
-        "builduid": "df292bc2d0934dd69bb8182448919107",
-        "locales": "or,pa-IN,pl,pt-BR,pt-PT,rm,ro,ru,si",
-        "log_url": "http://stage.mozilla.org/pub/mozilla.org/firefox/nightly/23.0.1-candidates/build1/logs/release-mozilla-release-win32_repack_8-bm66-build1-build0.txt.gz",
-        "master": "http://buildbot-master66.srv.releng.usw2.mozilla.com:8001/",
-        "platform": "win32",
-        "product": "firefox",
-        "products": "firefox",
-        "project": "",
-        "release_config": "mozilla/release-firefox-mozilla-release.py",
-        "repository": "",
-        "request_ids": [
-          28006149
-        ],
-        "request_times": {
-          "28006149": 1376503997
-        },
-        "revision": "15aaeae89597",
-        "scheduler": "release-mozilla-release-win32_repack",
-        "script_repo_revision": "97080c57a9ab",
-        "slavebuilddir": "rel-m-rel-w32_rpk_8-0000000000",
-        "slavename": "w64-ix-slave90",
-        "toolsdir": "/e/builds/moz2_slave/rel-m-rel-w32_rpk_8-0000000000/scripts",
-        "version": "23.0.1"
-      },
-      "reason": "Triggerable(release-mozilla-release-win32_repack)",
-      "request_ids": [
-        28006149
-      ],
-      "requesttime": 1376503997,
-      "result": 0,
-      "slave_id": 2459,
-      "starttime": 1376504001
-    },
-    {
-      "builder_id": 181855,
-      "buildnumber": 0,
-      "endtime": 1376513407,
-      "id": 26554270,
-      "master_id": 95,
-      "properties": {
-        "basedir": "/e/builds/moz2_slave/rel-m-rel-w32_rpk_10-000000000",
-        "branch": "test_treeherder",
-        "build_number": 1,
-        "builddir": "release-mozilla-release-win32_repack_10",
-        "buildername": "release-mozilla-release-win32_repack_10/10",
-        "buildid": "20130814063812",
-        "buildnumber": 0,
-        "builduid": "df292bc2d0934dd69bb8182448919107",
-        "locales": "te,th,tr,uk,vi,zh-CN,zh-TW,zu",
-        "log_url": "http://stage.mozilla.org/pub/mozilla.org/firefox/nightly/23.0.1-candidates/build1/logs/release-mozilla-release-win32_repack_10-bm66-build1-build0.txt.gz",
-        "master": "http://buildbot-master66.srv.releng.usw2.mozilla.com:8001/",
-        "platform": "win32",
-        "product": "firefox",
-        "products": "firefox",
-        "project": "",
-        "release_config": "mozilla/release-firefox-mozilla-release.py",
-        "repository": "",
-        "request_ids": [
-          28006151
-        ],
-        "request_times": {
-          "28006151": 1376503997
-        },
-        "revision": "15aaeae89597",
-        "scheduler": "release-mozilla-release-win32_repack",
-        "script_repo_revision": "97080c57a9ab",
-        "slavebuilddir": "rel-m-rel-w32_rpk_10-000000000",
-        "slavename": "w64-ix-slave07",
-        "toolsdir": "/e/builds/moz2_slave/rel-m-rel-w32_rpk_10-000000000/scripts",
-        "version": "23.0.1"
-      },
-      "reason": "Triggerable(release-mozilla-release-win32_repack)",
-      "request_ids": [
-        28006151
-      ],
-      "requesttime": 1376503997,
-      "result": 0,
-      "slave_id": 1330,
-      "starttime": 1376504001
-    },
-    {
-      "builder_id": 181858,
-      "buildnumber": 0,
-      "endtime": 1376514390,
-      "id": 26554852,
-      "master_id": 100,
-      "properties": {
-        "basedir": "/e/builds/moz2_slave/rel-m-rel-w32_rpk_4-0000000000",
-        "branch": "test_treeherder",
-        "build_number": 1,
-        "builddir": "release-mozilla-release-win32_repack_4",
-        "buildername": "release-mozilla-release-win32_repack_4/10",
-        "buildid": "20130814063812",
-        "buildnumber": 0,
-        "builduid": "df292bc2d0934dd69bb8182448919107",
-        "locales": "eu,fa,ff,fi,fr,fy-NL,ga-IE,gd,gl",
-        "log_url": "http://stage.mozilla.org/pub/mozilla.org/firefox/nightly/23.0.1-candidates/build1/logs/release-mozilla-release-win32_repack_4-bm65-build1-build0.txt.gz",
-        "master": "http://buildbot-master65.srv.releng.usw2.mozilla.com:8001/",
-        "platform": "win32",
-        "product": "firefox",
-        "products": "firefox",
-        "project": "",
-        "release_config": "mozilla/release-firefox-mozilla-release.py",
-        "repository": "",
-        "request_ids": [
-          28006145
-        ],
-        "request_times": {
-          "28006145": 1376503997
-        },
-        "revision": "15aaeae89597",
-        "scheduler": "release-mozilla-release-win32_repack",
-        "script_repo_revision": "97080c57a9ab",
-        "slavebuilddir": "rel-m-rel-w32_rpk_4-0000000000",
-        "slavename": "w64-ix-slave118",
-        "toolsdir": "/e/builds/moz2_slave/rel-m-rel-w32_rpk_4-0000000000/scripts",
-        "version": "23.0.1"
-      },
-      "reason": "Triggerable(release-mozilla-release-win32_repack)",
-      "request_ids": [
-        28006145
-      ],
-      "requesttime": 1376503997,
-      "result": 0,
-      "slave_id": 4151,
-      "starttime": 1376504004
-    },
-    {
-      "builder_id": 142056,
-      "buildnumber": 172,
-      "endtime": 1376511144,
-      "id": 26553157,
-      "master_id": 73,
-      "properties": {
-        "basedir": "/home/cltbld/talos-slave/test",
-        "branch": "test_treeherder",
-        "build_url": "http://ftp.mozilla.org/pub/mozilla.org/firefox/tinderbox-builds/fx-team-linux-debug/1376499083/firefox-26.0a1.en-US.linux-i686.tar.bz2",
-        "buildername": "Rev3 Fedora 12 fx-team debug test mochitest-browser-chrome",
-        "buildid": "20130814095123",
-        "buildnumber": 172,
-        "builduid": "e8c4556f1c404e6786b4d1e975cdf9c5",
-        "log_url": "http://ftp.mozilla.org/pub/mozilla.org/firefox/tinderbox-builds/fx-team-linux-debug/1376499083/fx-team_fedora-debug_test-mochitest-browser-chrome-bm54-tests1-linux-build172.txt.gz",
-        "master": "http://buildbot-master54.srv.releng.usw2.mozilla.com:8201/",
-        "pgo_build": "False",
-        "platform": "linux",
-        "product": "firefox",
-        "project": "",
-        "repo_path": "integration/fx-team",
-        "repository": "",
-        "request_ids": [
-          28006196
-        ],
-        "request_times": {
-          "28006196": 1376504041
-        },
-        "revision": "15aaeae89597",
-        "scheduler": "tests-fx-team-fedora-debug-unittest",
-        "script_repo_revision": "2f85052176f2",
-        "slavebuilddir": "test",
-        "slavename": "talos-r3-fed-074",
-        "stage_platform": "linux",
-        "symbols_url": "http://ftp.mozilla.org/pub/mozilla.org/firefox/tinderbox-builds/fx-team-linux-debug/1376499083/firefox-26.0a1.en-US.linux-i686.crashreporter-symbols.zip"
-      },
-      "reason": "scheduler",
-      "request_ids": [
-        28006196
-      ],
-      "requesttime": 1376504041,
-      "result": 0,
-      "slave_id": 1734,
-      "starttime": 1376504047
-    },
-    {
-      "builder_id": 118025,
-      "buildnumber": 120,
-      "endtime": 1376510984,
-      "id": 26553082,
-      "master_id": 71,
-      "properties": {
-        "basedir": "/builds/slave/test",
-        "branch": "test_treeherder",
-        "build_url": "http://ftp.mozilla.org/pub/mozilla.org/firefox/tinderbox-builds/fx-team-linux-debug/1376499083/firefox-26.0a1.en-US.linux-i686.tar.bz2",
-        "buildername": "Ubuntu VM 12.04 fx-team debug test reftest",
-        "buildid": "20130814095123",
-        "buildnumber": 120,
-        "builduid": "e8c4556f1c404e6786b4d1e975cdf9c5",
-        "log_url": "http://ftp.mozilla.org/pub/mozilla.org/firefox/tinderbox-builds/fx-team-linux-debug/1376499083/fx-team_ubuntu32_vm-debug_test-reftest-bm51-tests1-linux-build120.txt.gz",
-        "master": "http://buildbot-master51.srv.releng.use1.mozilla.com:8201/",
-        "pgo_build": "False",
-        "platform": "linux",
-        "product": "firefox",
-        "project": "",
-        "repo_path": "integration/fx-team",
-        "repository": "",
-        "request_ids": [
-          28006189
-        ],
-        "request_times": {
-          "28006189": 1376504039
-        },
-        "revision": "15aaeae89597",
-        "scheduler": "tests-fx-team-ubuntu32_vm-debug-unittest",
-        "script_repo_revision": "2f85052176f2",
-        "slavebuilddir": "test",
-        "slavename": "tst-linux32-ec2-045",
-        "stage_platform": "linux",
-        "symbols_url": "http://ftp.mozilla.org/pub/mozilla.org/firefox/tinderbox-builds/fx-team-linux-debug/1376499083/firefox-26.0a1.en-US.linux-i686.crashreporter-symbols.zip"
-      },
-      "reason": "scheduler",
-      "request_ids": [
-        28006189
-      ],
-      "requesttime": 1376504039,
-      "result": 0,
-      "slave_id": 3405,
-      "starttime": 1376504185
-    },
-    {
-      "builder_id": 141237,
-      "buildnumber": 1025,
-      "endtime": 1376511430,
-      "id": 26553313,
-      "master_id": 73,
-      "properties": {
-        "basedir": "/home/cltbld/talos-slave/test",
-        "branch": "test_treeherder",
-        "build_url": "http://ftp.mozilla.org/pub/mozilla.org/firefox/tinderbox-builds/mozilla-inbound-linux-debug/1376498063/firefox-26.0a1.en-US.linux-i686.tar.bz2",
-        "buildername": "Rev3 Fedora 12 mozilla-inbound debug test mochitest-browser-chrome",
-        "buildid": "20130814093423",
-        "buildnumber": 1025,
-        "builduid": "3abbb25a8451415ea08761a30789e73d",
-        "log_url": "http://ftp.mozilla.org/pub/mozilla.org/firefox/tinderbox-builds/mozilla-inbound-linux-debug/1376498063/mozilla-inbound_fedora-debug_test-mochitest-browser-chrome-bm54-tests1-linux-build1025.txt.gz",
-        "master": "http://buildbot-master54.srv.releng.usw2.mozilla.com:8201/",
-        "pgo_build": "False",
-        "platform": "linux",
-        "product": "firefox",
-        "project": "",
-        "repo_path": "integration/mozilla-inbound",
-        "repository": "",
-        "request_ids": [
-          28006530
-        ],
-        "request_times": {
-          "28006530": 1376504335
-        },
-        "revision": "8bf9c217ddd9",
-        "scheduler": "tests-mozilla-inbound-fedora-debug-unittest",
-        "script_repo_revision": "2f85052176f2",
-        "slavebuilddir": "test",
-        "slavename": "talos-r3-fed-062",
-        "stage_platform": "linux",
-        "symbols_url": "http://ftp.mozilla.org/pub/mozilla.org/firefox/tinderbox-builds/mozilla-inbound-linux-debug/1376498063/firefox-26.0a1.en-US.linux-i686.crashreporter-symbols.zip"
-      },
-      "reason": "scheduler",
-      "request_ids": [
-        28006530
-      ],
-      "requesttime": 1376504335,
-      "result": 0,
-      "slave_id": 1267,
-      "starttime": 1376504341
-    },
-    {
-      "builder_id": 141353,
-      "buildnumber": 374,
-      "endtime": 1376511648,
-      "id": 26553414,
-      "master_id": 80,
-      "properties": {
-        "basedir": "/home/cltbld/talos-slave/test",
-        "branch": "test_treeherder",
-        "build_url": "http://ftp.mozilla.org/pub/mozilla.org/firefox/try-builds/hurley@mozilla.com-3beb9ebd1dd7/try-linux-debug/firefox-26.0a1.en-US.linux-i686.tar.bz2",
-        "buildername": "Rev3 Fedora 12 try debug test mochitest-browser-chrome",
-        "buildid": "20130814093938",
-        "buildnumber": 374,
-        "builduid": "0c05c1b2af1d4a39868d5d5e2076b60e",
-        "log_url": "http://ftp.mozilla.org/pub/mozilla.org/firefox/try-builds/hurley@mozilla.com-3beb9ebd1dd7/try-linux-debug/try_fedora-debug_test-mochitest-browser-chrome-bm67-tests1-linux-build374.txt.gz",
-        "master": "http://buildbot-master67.srv.releng.use1.mozilla.com:8201/",
-        "platform": "linux",
-        "product": "firefox",
-        "project": "",
-        "repo_path": "try",
-        "repository": "",
-        "request_ids": [
-          28006558
-        ],
-        "request_times": {
-          "28006558": 1376504346
-        },
-        "revision": "3beb9ebd1dd7",
-        "scheduler": "tests-try-fedora-debug-unittest",
-        "script_repo_revision": "2f85052176f2",
-        "slavebuilddir": "test",
-        "slavename": "talos-r3-fed-024",
-        "stage_platform": "linux",
-        "symbols_url": "http://ftp.mozilla.org/pub/mozilla.org/firefox/try-builds/hurley@mozilla.com-3beb9ebd1dd7/try-linux-debug/firefox-26.0a1.en-US.linux-i686.crashreporter-symbols.zip"
-      },
-      "reason": "scheduler",
-      "request_ids": [
-        28006558
-      ],
-      "requesttime": 1376504346,
-      "result": 1,
-      "slave_id": 448,
-      "starttime": 1376504348
-    },
-    {
-      "builder_id": 116007,
-      "buildnumber": 1612,
-      "endtime": 1376511215,
-      "id": 26553199,
-      "master_id": 75,
-      "properties": {
-        "appName": "Firefox",
-        "appVersion": "26.0a1",
-        "basedir": "/builds/slave/try-lx-d-000000000000000000000",
-        "branch": "test_treeherder",
-        "builddir": "try-lx-d-000000000000000000000",
-        "buildername": "Linux try leak test build",
-        "buildid": "20130814112248",
-        "buildnumber": 1612,
-        "builduid": "c0b481e67f354f998daa81617ac025e4",
-        "comments": "try: -b do -p linux,linux64,macosx64,win32,android,android-armv6,android-noion,android-x86,panda,unagi -u crashtest,crashtest-ipc,xpcshell -t none",
-        "filepath": null,
-        "forced_clobber": false,
-        "got_revision": "68d9343692ac",
-        "hashType": "sha512",
-        "jsshellUrl": "http://ftp.mozilla.org/pub/mozilla.org/firefox/try-builds/gpascutto@mozilla.com-68d9343692ac/try-linux-debug/jsshell-linux-i686.zip",
-        "log_url": "http://ftp.mozilla.org/pub/mozilla.org/firefox/try-builds/gpascutto@mozilla.com-68d9343692ac/try-linux-debug/try-linux-debug-bm55-try1-build1612.txt.gz",
-        "master": "http://buildbot-master55.srv.releng.use1.mozilla.com:8101/",
-        "packageFilename": "firefox-26.0a1.en-US.linux-i686.tar.bz2",
-        "packageHash": "dc87e88cfd298f3680e7100a56e378d464575b8c5a9ecfafcd01de4984e538b58c8bfa1d0f3f22d9af515b28dec2cc2a676d6bf7d1d44a695a3a2ef7412f1b8d",
-        "packageSize": "36540660",
-        "packageUrl": "http://ftp.mozilla.org/pub/mozilla.org/firefox/try-builds/gpascutto@mozilla.com-68d9343692ac/try-linux-debug/firefox-26.0a1.en-US.linux-i686.tar.bz2",
-        "periodic_clobber": false,
-        "platform": "linux-debug",
-        "product": "firefox",
-        "project": "",
-        "purge_actual": "19.90GB",
-        "purge_target": "14GB",
-        "purged_clobber": true,
-        "repository": "",
-        "request_ids": [
-          28006029
-        ],
-        "request_times": {
-          "28006029": 1376503770
-        },
-        "revision": "68d9343692ac5cc1f3e4e3effc9b5a06da7f5e0d",
-        "scheduler": "try-firefox",
-        "slavebuilddir": "try-lx-d-000000000000000000000",
-        "slavename": "try-linux64-ec2-003",
-        "stage_platform": "linux-debug",
-        "symbolsUrl": "http://ftp.mozilla.org/pub/mozilla.org/firefox/try-builds/gpascutto@mozilla.com-68d9343692ac/try-linux-debug/firefox-26.0a1.en-US.linux-i686.crashreporter-symbols.zip",
-        "testsUrl": "http://ftp.mozilla.org/pub/mozilla.org/firefox/try-builds/gpascutto@mozilla.com-68d9343692ac/try-linux-debug/firefox-26.0a1.en-US.linux-i686.tests.zip",
-        "toolsdir": "/builds/slave/try-lx-d-000000000000000000000/tools",
-        "who": "gpascutto@mozilla.com"
-      },
-      "reason": "scheduler",
-      "request_ids": [
-        28006029
-      ],
-      "requesttime": 1376503770,
-      "result": 1,
-      "slave_id": 2169,
-      "starttime": 1376504371
-    },
-    {
-      "builder_id": 120700,
-      "buildnumber": 915,
-      "endtime": 1376512220,
-      "id": 26553684,
-      "master_id": 87,
-      "properties": {
-        "basedir": "/builds/slave/talos-slave/test",
-        "branch": "test_treeherder",
-        "build_url": "http://ftp.mozilla.org/pub/mozilla.org/firefox/tinderbox-builds/mozilla-inbound-macosx64-debug/1376500942/firefox-26.0a1.en-US.mac64.dmg",
-        "buildername": "Rev4 MacOSX Lion 10.7 mozilla-inbound debug test mochitest-browser-chrome",
-        "buildid": "20130814102222",
-        "buildnumber": 915,
-        "builduid": "1b94f824176a48a59aeaf61252f53666",
-        "log_url": "http://ftp.mozilla.org/pub/mozilla.org/firefox/tinderbox-builds/mozilla-inbound-macosx64-debug/1376500942/mozilla-inbound_lion-debug_test-mochitest-browser-chrome-bm78-tests1-macosx-build915.txt.gz",
-        "master": "http://buildbot-master78.srv.releng.usw2.mozilla.com:8201/",
-        "pgo_build": "False",
-        "platform": "macosx64",
-        "product": "firefox",
-        "project": "",
-        "repo_path": "integration/mozilla-inbound",
-        "repository": "",
-        "request_ids": [
-          28006311
-        ],
-        "request_times": {
-          "28006311": 1376504104
-        },
-        "revision": "355dcff578b6",
-        "scheduler": "tests-mozilla-inbound-lion-debug-unittest",
-        "script_repo_revision": "2f85052176f2",
-        "slavebuilddir": "test",
-        "slavename": "talos-r4-lion-001",
-        "stage_platform": "macosx64",
-        "symbols_url": "http://ftp.mozilla.org/pub/mozilla.org/firefox/tinderbox-builds/mozilla-inbound-macosx64-debug/1376500942/firefox-26.0a1.en-US.mac64.crashreporter-symbols.zip"
-      },
-      "reason": "scheduler",
-      "request_ids": [
-        28006311
-      ],
-      "requesttime": 1376504104,
-      "result": 0,
-      "slave_id": 4231,
-      "starttime": 1376504428
-    },
-    {
-      "builder_id": 178191,
-      "buildnumber": 4,
-      "endtime": 1376523667,
-      "id": 26560898,
-      "master_id": 89,
-      "properties": {
-        "basedir": "c:\\slave\\test",
-        "branch": "test_treeherder",
-        "build_url": "http://ftp.mozilla.org/pub/mozilla.org/firefox/tinderbox-builds/date-win64/1376488843/firefox-26.0a1.en-US.win64-x86_64.zip",
-        "buildername": "win64_vm date opt test mochitest-browser-chrome",
-        "buildid": "20130814070043",
-        "buildnumber": 4,
-        "builduid": "6cf450df1e4940298e09e24a1cc9a698",
-        "log_url": "http://ftp.mozilla.org/pub/mozilla.org/firefox/tinderbox-builds/date-win64/1376488843/date_win64_vm_test-mochitest-browser-chrome-bm71-tests1-windows-build4.txt.gz",
-        "master": "http://buildbot-master71.srv.releng.use1.mozilla.com:8201/",
-        "pgo_build": "False",
-        "platform": "win64",
-        "product": "firefox",
-        "project": "",
-        "repo_path": "projects/date",
-        "repository": "",
-        "request_ids": [
-          27999325
-        ],
-        "request_times": {
-          "27999325": 1376495343
-        },
-        "revision": "4930fdea3efa",
-        "scheduler": "tests-date-win64_vm-opt-unittest",
-        "script_repo_revision": "2f85052176f2",
-        "slavebuilddir": "test",
-        "slavename": "tst-w64-ec2-005",
-        "stage_platform": "win64"
-      },
-      "reason": "scheduler",
-      "request_ids": [
-        27999325
-      ],
-      "requesttime": 1376495343,
-      "result": 4,
-      "slave_id": 5023,
-      "starttime": 1376504436
-    },
-    {
-      "builder_id": 121197,
-      "buildnumber": 908,
-      "endtime": 1376512088,
-      "id": 26553633,
-      "master_id": 87,
-      "properties": {
-        "basedir": "/Users/cltbld/talos-slave/test",
-        "branch": "test_treeherder",
-        "build_url": "http://ftp.mozilla.org/pub/mozilla.org/firefox/tinderbox-builds/mozilla-inbound-macosx64-debug/1376500942/firefox-26.0a1.en-US.mac64.dmg",
-        "buildername": "Rev4 MacOSX Snow Leopard 10.6 mozilla-inbound debug test mochitest-browser-chrome",
-        "buildid": "20130814102222",
-        "buildnumber": 908,
-        "builduid": "1b94f824176a48a59aeaf61252f53666",
-        "log_url": "http://ftp.mozilla.org/pub/mozilla.org/firefox/tinderbox-builds/mozilla-inbound-macosx64-debug/1376500942/mozilla-inbound_snowleopard-debug_test-mochitest-browser-chrome-bm78-tests1-macosx-build908.txt.gz",
-        "master": "http://buildbot-master78.srv.releng.usw2.mozilla.com:8201/",
-        "pgo_build": "False",
-        "platform": "macosx64",
-        "product": "firefox",
-        "project": "",
-        "repo_path": "integration/mozilla-inbound",
-        "repository": "",
-        "request_ids": [
-          28006298
-        ],
-        "request_times": {
-          "28006298": 1376504104
-        },
-        "revision": "355dcff578b6",
-        "scheduler": "tests-mozilla-inbound-snowleopard-debug-unittest",
-        "script_repo_revision": "2f85052176f2",
-        "slavebuilddir": "test",
-        "slavename": "talos-r4-snow-075",
-        "stage_platform": "macosx64",
-        "symbols_url": "http://ftp.mozilla.org/pub/mozilla.org/firefox/tinderbox-builds/mozilla-inbound-macosx64-debug/1376500942/firefox-26.0a1.en-US.mac64.crashreporter-symbols.zip"
-      },
-      "reason": "scheduler",
-      "request_ids": [
-        28006298
-      ],
-      "requesttime": 1376504104,
-      "result": 0,
-      "slave_id": 1413,
-      "starttime": 1376504444
-    },
-    {
-      "builder_id": 141237,
-      "buildnumber": 1026,
-      "endtime": 1376511536,
-      "id": 26553362,
-      "master_id": 73,
-      "properties": {
-        "basedir": "/home/cltbld/talos-slave/test",
-        "branch": "test_treeherder",
-        "build_url": "http://ftp.mozilla.org/pub/mozilla.org/firefox/tinderbox-builds/mozilla-inbound-linux-debug/1376498842/firefox-26.0a1.en-US.linux-i686.tar.bz2",
-        "buildername": "Rev3 Fedora 12 mozilla-inbound debug test mochitest-browser-chrome",
-        "buildid": "20130814094722",
-        "buildnumber": 1026,
-        "builduid": "352ec5167ac94127b8d954c694180925",
-        "log_url": "http://ftp.mozilla.org/pub/mozilla.org/firefox/tinderbox-builds/mozilla-inbound-linux-debug/1376498842/mozilla-inbound_fedora-debug_test-mochitest-browser-chrome-bm54-tests1-linux-build1026.txt.gz",
-        "master": "http://buildbot-master54.srv.releng.usw2.mozilla.com:8201/",
-        "pgo_build": "False",
-        "platform": "linux",
-        "product": "firefox",
-        "project": "",
-        "repo_path": "integration/mozilla-inbound",
-        "repository": "",
-        "request_ids": [
-          28006694
-        ],
-        "request_times": {
-          "28006694": 1376504454
-        },
-        "revision": "135c0ebda47b",
-        "scheduler": "tests-mozilla-inbound-fedora-debug-unittest",
-        "script_repo_revision": "2f85052176f2",
-        "slavebuilddir": "test",
-        "slavename": "talos-r3-fed-057",
-        "stage_platform": "linux",
-        "symbols_url": "http://ftp.mozilla.org/pub/mozilla.org/firefox/tinderbox-builds/mozilla-inbound-linux-debug/1376498842/firefox-26.0a1.en-US.linux-i686.crashreporter-symbols.zip"
-      },
-      "reason": "scheduler",
-      "request_ids": [
-        28006694
-      ],
-      "requesttime": 1376504454,
-      "result": 0,
-      "slave_id": 1271,
-      "starttime": 1376504456
-    },
-    {
-      "builder_id": 116679,
-      "buildnumber": 1892,
-      "endtime": 1376511593,
-      "id": 26553393,
-      "master_id": 75,
-      "properties": {
-        "appName": "Firefox",
-        "appVersion": "26.0a1",
-        "basedir": "/builds/slave/try-lx-00000000000000000000000",
-        "branch": "test_treeherder",
-        "builddir": "try-lx-00000000000000000000000",
-        "buildername": "Linux try build",
-        "buildid": "20130814112600",
-        "buildnumber": 1892,
-        "builduid": "37865e72c32c4a0c86469c1f7ac5bae4",
-        "comments": "SDK Test: try: -b do -p linux -u jetpack -t none",
-        "filepath": null,
-        "forced_clobber": false,
-        "got_revision": "8cfe7c02a0ed",
-        "hashType": "sha512",
-        "jsshellUrl": "http://ftp.mozilla.org/pub/mozilla.org/firefox/try-builds/dtownsend@mozilla.com-8cfe7c02a0ed/try-linux/jsshell-linux-i686.zip",
-        "log_url": "http://ftp.mozilla.org/pub/mozilla.org/firefox/try-builds/dtownsend@mozilla.com-8cfe7c02a0ed/try-linux/try-linux-bm55-try1-build1892.txt.gz",
-        "master": "http://buildbot-master55.srv.releng.use1.mozilla.com:8101/",
-        "num_ctors": 168,
-        "packageFilename": "firefox-26.0a1.en-US.linux-i686.tar.bz2",
-        "packageHash": "439b330d1ae0c0f7f3ed2e6d51118fffe093bd71d5c773e4caf6007bcd06cd89f6531208f3568648b15f712652c7c5034ac710a14be1d9eb973e9e803c6c506c",
-        "packageSize": "36582665",
-        "packageUrl": "http://ftp.mozilla.org/pub/mozilla.org/firefox/try-builds/dtownsend@mozilla.com-8cfe7c02a0ed/try-linux/firefox-26.0a1.en-US.linux-i686.tar.bz2",
-        "periodic_clobber": false,
-        "platform": "linux",
-        "product": "firefox",
-        "project": "",
-        "purge_actual": "16.37GB",
-        "purge_target": "12GB",
-        "purged_clobber": true,
-        "repository": "",
-        "request_ids": [
-          28006771
-        ],
-        "request_times": {
-          "28006771": 1376504542
-        },
-        "revision": "8cfe7c02a0ed4f8662df2fb5a13929b09b483dd6",
-        "scheduler": "try-firefox",
-        "slavebuilddir": "try-lx-00000000000000000000000",
-        "slavename": "try-linux64-ec2-070",
-        "sourcestamp": "8cfe7c02a0ed",
-        "stage_platform": "linux",
-        "symbolsUrl": "http://ftp.mozilla.org/pub/mozilla.org/firefox/try-builds/dtownsend@mozilla.com-8cfe7c02a0ed/try-linux/firefox-26.0a1.en-US.linux-i686.crashreporter-symbols.zip",
-        "testresults": [
-          [
-            "num_ctors",
-            "num_ctors",
-            168,
-            "168"
-          ]
-        ],
-        "testsUrl": "http://ftp.mozilla.org/pub/mozilla.org/firefox/try-builds/dtownsend@mozilla.com-8cfe7c02a0ed/try-linux/firefox-26.0a1.en-US.linux-i686.tests.zip",
-        "toolsdir": "/builds/slave/try-lx-00000000000000000000000/tools",
-        "who": "dtownsend@mozilla.com"
-      },
-      "reason": "scheduler",
-      "request_ids": [
-        28006771
-      ],
-      "requesttime": 1376504542,
-      "result": 0,
-      "slave_id": 2514,
-      "starttime": 1376504555
-    },
-    {
-      "builder_id": 141256,
-      "buildnumber": 593,
-      "endtime": 1376511779,
-      "id": 26553476,
-      "master_id": 79,
-      "properties": {
-        "basedir": "/home/cltbld/talos-slave/test",
-        "branch": "test_treeherder",
-        "build_url": "http://ftp.mozilla.org/pub/mozilla.org/firefox/try-builds/jandemooij@gmail.com-5974016d7c49/try-linux-debug/firefox-26.0a1.en-US.linux-i686.tar.bz2",
-        "buildername": "Rev3 Fedora 12 try debug test mochitest-browser-chrome",
-        "buildid": "20130814095056",
-        "buildnumber": 593,
-        "builduid": "df39c7ff9efc416d93cd988e6c9503e8",
-        "log_url": "http://ftp.mozilla.org/pub/mozilla.org/firefox/try-builds/jandemooij@gmail.com-5974016d7c49/try-linux-debug/try_fedora-debug_test-mochitest-browser-chrome-bm68-tests1-linux-build593.txt.gz",
-        "master": "http://buildbot-master68.srv.releng.usw2.mozilla.com:8201/",
-        "platform": "linux",
-        "product": "firefox",
-        "project": "",
-        "repo_path": "try",
-        "repository": "",
-        "request_ids": [
-          28006995
-        ],
-        "request_times": {
-          "28006995": 1376504655
-        },
-        "revision": "5974016d7c49",
-        "scheduler": "tests-try-fedora-debug-unittest",
-        "script_repo_revision": "2f85052176f2",
-        "slavebuilddir": "test",
-        "slavename": "talos-r3-fed-098",
-        "stage_platform": "linux",
-        "symbols_url": "http://ftp.mozilla.org/pub/mozilla.org/firefox/try-builds/jandemooij@gmail.com-5974016d7c49/try-linux-debug/firefox-26.0a1.en-US.linux-i686.crashreporter-symbols.zip"
-      },
-      "reason": "scheduler",
-      "request_ids": [
-        28006995
-      ],
-      "requesttime": 1376504655,
-      "result": 1,
-      "slave_id": 3646,
-      "starttime": 1376504657
-    },
-    {
-      "builder_id": 115926,
-      "buildnumber": 556,
-      "endtime": 1376511522,
-      "id": 26553360,
-      "master_id": 72,
-      "properties": {
-        "basedir": "/builds/slave/test",
-        "branch": "test_treeherder",
-        "build_url": "http://ftp.mozilla.org/pub/mozilla.org/firefox/try-builds/hurley@mozilla.com-3beb9ebd1dd7/try-linux-debug/firefox-26.0a1.en-US.linux-i686.tar.bz2",
-        "buildername": "Ubuntu VM 12.04 try debug test reftest",
-        "buildid": "20130814093938",
-        "buildnumber": 556,
-        "builduid": "0c05c1b2af1d4a39868d5d5e2076b60e",
-        "log_url": "http://ftp.mozilla.org/pub/mozilla.org/firefox/try-builds/hurley@mozilla.com-3beb9ebd1dd7/try-linux-debug/try_ubuntu32_vm-debug_test-reftest-bm52-tests1-linux-build556.txt.gz",
-        "master": "http://buildbot-master52.srv.releng.use1.mozilla.com:8201/",
-        "platform": "linux",
-        "product": "firefox",
-        "project": "",
-        "repo_path": "try",
-        "repository": "",
-        "request_ids": [
-          28006571
-        ],
-        "request_times": {
-          "28006571": 1376504347
-        },
-        "revision": "3beb9ebd1dd7",
-        "scheduler": "tests-try-ubuntu32_vm-debug-unittest",
-        "script_repo_revision": "2f85052176f2",
-        "slavebuilddir": "test",
-        "slavename": "tst-linux32-ec2-070",
-        "stage_platform": "linux",
-        "symbols_url": "http://ftp.mozilla.org/pub/mozilla.org/firefox/try-builds/hurley@mozilla.com-3beb9ebd1dd7/try-linux-debug/firefox-26.0a1.en-US.linux-i686.crashreporter-symbols.zip"
-      },
-      "reason": "scheduler",
-      "request_ids": [
-        28006571
-      ],
-      "requesttime": 1376504347,
-      "result": 0,
-      "slave_id": 3524,
-      "starttime": 1376504733
-    },
-    {
-      "builder_id": 173109,
-      "buildnumber": 5,
-      "endtime": 1376511647,
-      "id": 26553415,
-      "master_id": 86,
-      "properties": {
-        "appName": "Firefox",
-        "appVersion": "26.0a1",
-        "basedir": "/builds/slave/ced-lx-d-000000000000000000000",
-        "branch": "test_treeherder",
-        "builddir": "ced-lx-d-000000000000000000000",
-        "buildername": "Linux cedar leak test build",
-        "buildid": "20130814112532",
-        "buildnumber": 5,
-        "builduid": "f3984a6bbba449f3b35ec692a783fd62",
-        "comments": "Merge m-c -> cedar",
-        "filepath": null,
-        "forced_clobber": false,
-        "got_revision": "f83f96c52b11",
-        "hashType": "sha512",
-        "jsshellUrl": "http://ftp.mozilla.org/pub/mozilla.org/firefox/tinderbox-builds/cedar-linux-debug/1376504732/jsshell-linux-i686.zip",
-        "log_url": "http://ftp.mozilla.org/pub/mozilla.org/firefox/tinderbox-builds/cedar-linux-debug/1376504732/cedar-linux-debug-bm63-build1-build5.txt.gz",
-        "master": "http://buildbot-master63.srv.releng.use1.mozilla.com:8001/",
-        "packageFilename": "firefox-26.0a1.en-US.linux-i686.tar.bz2",
-        "packageHash": "13873b93e8f8552faf8927164666378b7a6d91a9814f82d5233492534321e05b4f7547db351f94d95409b1ffa3ce584b38d5575c794d78fbf69bd98715f92795",
-        "packageSize": "36511936",
-        "packageUrl": "http://ftp.mozilla.org/pub/mozilla.org/firefox/tinderbox-builds/cedar-linux-debug/1376504732/firefox-26.0a1.en-US.linux-i686.tar.bz2",
-        "periodic_clobber": false,
-        "platform": "linux-debug",
-        "product": "firefox",
-        "project": "",
-        "purge_actual": "41.51GB",
-        "purge_target": "14GB",
-        "purged_clobber": true,
-        "repository": "",
-        "request_ids": [
-          28007014
-        ],
-        "request_times": {
-          "28007014": 1376504732
-        },
-        "revision": "f83f96c52b11820b317f21d138b5ce3452334082",
-        "scheduler": "cedar-firefox",
-        "slavebuilddir": "ced-lx-d-000000000000000000000",
-        "slavename": "bld-linux64-ix-037",
-        "stage_platform": "linux-debug",
-        "symbolsUrl": "http://ftp.mozilla.org/pub/mozilla.org/firefox/tinderbox-builds/cedar-linux-debug/1376504732/firefox-26.0a1.en-US.linux-i686.crashreporter-symbols.zip",
-        "testsUrl": "http://ftp.mozilla.org/pub/mozilla.org/firefox/tinderbox-builds/cedar-linux-debug/1376504732/firefox-26.0a1.en-US.linux-i686.tests.zip",
-        "toolsdir": "/builds/slave/ced-lx-d-000000000000000000000/tools"
-      },
-      "reason": "scheduler",
-      "request_ids": [
-        28007014
-      ],
-      "requesttime": 1376504732,
-      "result": 0,
-      "slave_id": 3781,
-      "starttime": 1376504737
-    },
-    {
-      "builder_id": 136604,
-      "buildnumber": 14,
-      "endtime": 1376511930,
-      "id": 26553541,
-      "master_id": 76,
-      "properties": {
-        "appName": "Firefox",
-        "appVersion": "26.0a1",
-        "basedir": "/builds/slave/ced-lx-00000000000000000000000",
-        "branch": "test_treeherder",
-        "builddir": "ced-lx-00000000000000000000000",
-        "buildername": "Linux cedar build",
-        "buildid": "20130814112532",
-        "buildnumber": 14,
-        "builduid": "f3984a6bbba449f3b35ec692a783fd62",
-        "comments": "Merge m-c -> cedar",
-        "filepath": null,
-        "forced_clobber": false,
-        "got_revision": "f83f96c52b11",
-        "hashType": "sha512",
-        "jsshellUrl": "http://ftp.mozilla.org/pub/mozilla.org/firefox/tinderbox-builds/cedar-linux/1376504732/jsshell-linux-i686.zip",
-        "log_url": "http://ftp.mozilla.org/pub/mozilla.org/firefox/tinderbox-builds/cedar-linux/1376504732/cedar-linux-bm57-build1-build14.txt.gz",
-        "master": "http://buildbot-master57.srv.releng.use1.mozilla.com:8001/",
-        "num_ctors": 168,
-        "packageFilename": "firefox-26.0a1.en-US.linux-i686.tar.bz2",
-        "packageHash": "a804e97260212554d0f20aedb9a500071970654ab409c0d8dbf30b226d085e98807aa13d0231260049b92f4f9e97fbab20348c9d010a831f6df07ac11891abca",
-        "packageSize": "36581316",
-        "packageUrl": "http://ftp.mozilla.org/pub/mozilla.org/firefox/tinderbox-builds/cedar-linux/1376504732/firefox-26.0a1.en-US.linux-i686.tar.bz2",
-        "periodic_clobber": false,
-        "platform": "linux",
-        "product": "firefox",
-        "project": "",
-        "purge_actual": "85.97GB",
-        "purge_target": "12GB",
-        "purged_clobber": true,
-        "repository": "",
-        "request_ids": [
-          28007013
-        ],
-        "request_times": {
-          "28007013": 1376504732
-        },
-        "revision": "f83f96c52b11820b317f21d138b5ce3452334082",
-        "scheduler": "cedar-firefox",
-        "slavebuilddir": "ced-lx-00000000000000000000000",
-        "slavename": "bld-linux64-ec2-104",
-        "sourcestamp": "f83f96c52b11",
-        "stage_platform": "linux",
-        "symbolsUrl": "http://ftp.mozilla.org/pub/mozilla.org/firefox/tinderbox-builds/cedar-linux/1376504732/firefox-26.0a1.en-US.linux-i686.crashreporter-symbols.zip",
-        "testresults": [
-          [
-            "num_ctors",
-            "num_ctors",
-            168,
-            "168"
-          ]
-        ],
-        "testsUrl": "http://ftp.mozilla.org/pub/mozilla.org/firefox/tinderbox-builds/cedar-linux/1376504732/firefox-26.0a1.en-US.linux-i686.tests.zip",
-        "toolsdir": "/builds/slave/ced-lx-00000000000000000000000/tools"
-      },
-      "reason": "scheduler",
-      "request_ids": [
-        28007013
-      ],
-      "requesttime": 1376504732,
-      "result": 0,
-      "slave_id": 4797,
-      "starttime": 1376504737
-    },
-    {
-      "builder_id": 147016,
-      "buildnumber": 5,
-      "endtime": 1376513570,
-      "id": 26554367,
-      "master_id": 86,
-      "properties": {
-        "appName": "Firefox",
-        "appVersion": "26.0a1",
-        "basedir": "e:/builds/moz2_slave/ced-w32-d-00000000000000000000",
-        "branch": "test_treeherder",
-        "builddir": "ced-w32-d-00000000000000000000",
-        "buildername": "WINNT 5.2 cedar leak test build",
-        "buildid": "20130814112532",
-        "buildnumber": 5,
-        "builduid": "f3984a6bbba449f3b35ec692a783fd62",
-        "comments": "Merge m-c -> cedar",
-        "filepath": null,
-        "forced_clobber": false,
-        "got_revision": "f83f96c52b11",
-        "hashType": "sha512",
-        "installerFilename": "firefox-26.0a1.en-US.win32.installer.exe",
-        "installerHash": "f959819df5e6e82cb060b0b72a59b894d59e378ff0d373a142aafc46dd3c6516a8a9062c84150b94b08b1b7ebd280f47c8994d8421669fac18158294160d1dd5",
-        "installerSize": "29614544",
-        "jsshellUrl": "http://ftp.mozilla.org/pub/mozilla.org/firefox/tinderbox-builds/cedar-win32-debug/1376504732/jsshell-win32.zip",
-        "log_url": "http://ftp.mozilla.org/pub/mozilla.org/firefox/tinderbox-builds/cedar-win32-debug/1376504732/cedar-win32-debug-bm63-build1-build5.txt.gz",
-        "master": "http://buildbot-master63.srv.releng.use1.mozilla.com:8001/",
-        "packageFilename": "firefox-26.0a1.en-US.win32.zip",
-        "packageHash": "a643ff749bcd7c88254b65477cd664bffc1a1a6357ddb21b68a7a0fc2e575021c69685b344350f5b3b29e89a66a68c481868b44059529351c496b0e23b8d34c2",
-        "packageSize": "41337703",
-        "packageUrl": "http://ftp.mozilla.org/pub/mozilla.org/firefox/tinderbox-builds/cedar-win32-debug/1376504732/firefox-26.0a1.en-US.win32.zip",
-        "periodic_clobber": false,
-        "platform": "win32-debug",
-        "product": "firefox",
-        "project": "",
-        "purge_actual": "127.22GB",
-        "purge_target": "9GB",
-        "purged_clobber": true,
-        "repository": "",
-        "request_ids": [
-          28007020
-        ],
-        "request_times": {
-          "28007020": 1376504732
-        },
-        "revision": "f83f96c52b11820b317f21d138b5ce3452334082",
-        "scheduler": "cedar-firefox",
-        "slavebuilddir": "ced-w32-d-00000000000000000000",
-        "slavename": "w64-ix-slave123",
-        "stage_platform": "win32-debug",
-        "symbolsUrl": "http://ftp.mozilla.org/pub/mozilla.org/firefox/tinderbox-builds/cedar-win32-debug/1376504732/firefox-26.0a1.en-US.win32.crashreporter-symbols.zip",
-        "testsUrl": "http://ftp.mozilla.org/pub/mozilla.org/firefox/tinderbox-builds/cedar-win32-debug/1376504732/firefox-26.0a1.en-US.win32.tests.zip",
-        "toolsdir": "e:/builds/moz2_slave/ced-w32-d-00000000000000000000/tools"
-      },
-      "reason": "scheduler",
-      "request_ids": [
-        28007020
-      ],
-      "requesttime": 1376504732,
-      "result": 0,
-      "slave_id": 4462,
-      "starttime": 1376504738
-    },
-    {
-      "builder_id": 147816,
-      "buildnumber": 15,
-      "endtime": 1376512413,
-      "id": 26553778,
-      "master_id": 76,
-      "properties": {
-        "appName": "Firefox",
-        "appVersion": "26.0a1",
-        "basedir": "e:/builds/moz2_slave/ced-w32-0000000000000000000000",
-        "branch": "test_treeherder",
-        "builddir": "ced-w32-0000000000000000000000",
-        "buildername": "WINNT 5.2 cedar build",
-        "buildid": "20130814112532",
-        "buildnumber": 15,
-        "builduid": "f3984a6bbba449f3b35ec692a783fd62",
-        "comments": "Merge m-c -> cedar",
-        "filepath": null,
-        "forced_clobber": false,
-        "got_revision": "f83f96c52b11",
-        "hashType": "sha512",
-        "installerFilename": "firefox-26.0a1.en-US.win32.installer.exe",
-        "installerHash": "0aeb7316a501dc485789679b62011b0f2237c5e34c6ce741f60c6b2b856d586fc3efa9a0a1d965862fab92450a17a6518ecc162d20c7cf6eba81df3d4fcc7f4a",
-        "installerSize": "25631992",
-        "jsshellUrl": "http://ftp.mozilla.org/pub/mozilla.org/firefox/tinderbox-builds/cedar-win32/1376504732/jsshell-win32.zip",
-        "log_url": "http://ftp.mozilla.org/pub/mozilla.org/firefox/tinderbox-builds/cedar-win32/1376504732/cedar-win32-bm57-build1-build15.txt.gz",
-        "master": "http://buildbot-master57.srv.releng.use1.mozilla.com:8001/",
-        "packageFilename": "firefox-26.0a1.en-US.win32.zip",
-        "packageHash": "138d92fff69fa9c5210fd810e117b9d05f46a766a55307eb54e107ba726769328b9e99c1d0bae1aea9fe2f6f3900a1a291f7c9b44d30b46beaac0ca42b620f87",
-        "packageSize": "33612317",
-        "packageUrl": "http://ftp.mozilla.org/pub/mozilla.org/firefox/tinderbox-builds/cedar-win32/1376504732/firefox-26.0a1.en-US.win32.zip",
-        "periodic_clobber": false,
-        "platform": "win32",
-        "product": "firefox",
-        "project": "",
-        "purge_actual": "154.21GB",
-        "purge_target": "12GB",
-        "purged_clobber": true,
-        "repository": "",
-        "request_ids": [
-          28007019
-        ],
-        "request_times": {
-          "28007019": 1376504732
-        },
-        "revision": "f83f96c52b11820b317f21d138b5ce3452334082",
-        "scheduler": "cedar-firefox",
-        "slavebuilddir": "ced-w32-0000000000000000000000",
-        "slavename": "w64-ix-slave131",
-        "stage_platform": "win32",
-        "symbolsUrl": "http://ftp.mozilla.org/pub/mozilla.org/firefox/tinderbox-builds/cedar-win32/1376504732/firefox-26.0a1.en-US.win32.crashreporter-symbols.zip",
-        "testsUrl": "http://ftp.mozilla.org/pub/mozilla.org/firefox/tinderbox-builds/cedar-win32/1376504732/firefox-26.0a1.en-US.win32.tests.zip",
-        "toolsdir": "e:/builds/moz2_slave/ced-w32-0000000000000000000000/tools"
-      },
-      "reason": "scheduler",
-      "request_ids": [
-        28007019
-      ],
-      "requesttime": 1376504732,
-      "result": 0,
-      "slave_id": 4446,
-      "starttime": 1376504739
-    },
-    {
-      "builder_id": 127685,
-      "buildnumber": 11,
-      "endtime": 1376512724,
-      "id": 26553933,
-      "master_id": 86,
-      "properties": {
-        "appName": "Firefox",
-        "appVersion": "26.0a1",
-        "basedir": "/builds/slave/ced-l64-d-00000000000000000000",
-        "branch": "test_treeherder",
-        "builddir": "ced-l64-d-00000000000000000000",
-        "buildername": "Linux x86-64 cedar leak test build",
-        "buildid": "20130814112532",
-        "buildnumber": 11,
-        "builduid": "f3984a6bbba449f3b35ec692a783fd62",
-        "comments": "Merge m-c -> cedar",
-        "filepath": null,
-        "forced_clobber": false,
-        "got_revision": "f83f96c52b11",
-        "hashType": "sha512",
-        "jsshellUrl": "http://ftp.mozilla.org/pub/mozilla.org/firefox/tinderbox-builds/cedar-linux64-debug/1376504732/jsshell-linux-x86_64.zip",
-        "log_url": "http://ftp.mozilla.org/pub/mozilla.org/firefox/tinderbox-builds/cedar-linux64-debug/1376504732/cedar-linux64-debug-bm63-build1-build11.txt.gz",
-        "master": "http://buildbot-master63.srv.releng.use1.mozilla.com:8001/",
-        "packageFilename": "firefox-26.0a1.en-US.linux-x86_64.tar.bz2",
-        "packageHash": "77780a7e4774881d38da76ac3cada398cfe25242ae6392bef927cfbeadb10e45dcf27a0381b6214ddd1be4f28abf92cefd016b85b57820839e7604b450ef259d",
-        "packageSize": "36617804",
-        "packageUrl": "http://ftp.mozilla.org/pub/mozilla.org/firefox/tinderbox-builds/cedar-linux64-debug/1376504732/firefox-26.0a1.en-US.linux-x86_64.tar.bz2",
-        "periodic_clobber": false,
-        "platform": "linux64-debug",
-        "product": "firefox",
-        "project": "",
-        "purge_actual": "34.90GB",
-        "purge_target": "14GB",
-        "purged_clobber": true,
-        "repository": "",
-        "request_ids": [
-          28007016
-        ],
-        "request_times": {
-          "28007016": 1376504732
-        },
-        "revision": "f83f96c52b11820b317f21d138b5ce3452334082",
-        "scheduler": "cedar-firefox",
-        "slavebuilddir": "ced-l64-d-00000000000000000000",
-        "slavename": "bld-linux64-ec2-057",
-        "stage_platform": "linux64-debug",
-        "symbolsUrl": "http://ftp.mozilla.org/pub/mozilla.org/firefox/tinderbox-builds/cedar-linux64-debug/1376504732/firefox-26.0a1.en-US.linux-x86_64.crashreporter-symbols.zip",
-        "testsUrl": "http://ftp.mozilla.org/pub/mozilla.org/firefox/tinderbox-builds/cedar-linux64-debug/1376504732/firefox-26.0a1.en-US.linux-x86_64.tests.zip",
-        "toolsdir": "/builds/slave/ced-l64-d-00000000000000000000/tools"
-      },
-      "reason": "scheduler",
-      "request_ids": [
-        28007016
-      ],
-      "requesttime": 1376504732,
-      "result": 0,
-      "slave_id": 2484,
-      "starttime": 1376504739
-    },
-    {
-      "builder_id": 144145,
-      "buildnumber": 7,
-      "endtime": 1376512051,
-      "id": 26553612,
-      "master_id": 86,
-      "properties": {
-        "appName": "Firefox",
-        "appVersion": "26.0a1",
-        "basedir": "/builds/slave/ced-osx64-00000000000000000000",
-        "branch": "test_treeherder",
-        "builddir": "ced-osx64-00000000000000000000",
-        "buildername": "OS X 10.7 cedar build",
-        "buildid": "20130814112532",
-        "buildnumber": 7,
-        "builduid": "f3984a6bbba449f3b35ec692a783fd62",
-        "comments": "Merge m-c -> cedar",
-        "filepath": null,
-        "forced_clobber": false,
-        "got_revision": "f83f96c52b11",
-        "hashType": "sha512",
-        "jsshellUrl": "http://ftp.mozilla.org/pub/mozilla.org/firefox/tinderbox-builds/cedar-macosx64/1376504732/jsshell-mac.zip",
-        "log_url": "http://ftp.mozilla.org/pub/mozilla.org/firefox/tinderbox-builds/cedar-macosx64/1376504732/cedar-macosx64-bm63-build1-build7.txt.gz",
-        "master": "http://buildbot-master63.srv.releng.use1.mozilla.com:8001/",
-        "packageFilename": "firefox-26.0a1.en-US.mac.dmg",
-        "packageHash": "45decad16b3b34faa288866ba9b3c42087ef0f1aa1821450e6eb715867408d84c2b1f9b626849763aeced3a52b2bb8b3e61e7331e2f62f011737c824bfa8d0c6",
-        "packageSize": "69058279",
-        "packageUrl": "http://ftp.mozilla.org/pub/mozilla.org/firefox/tinderbox-builds/cedar-macosx64/1376504732/firefox-26.0a1.en-US.mac.dmg",
-        "periodic_clobber": false,
-        "platform": "macosx64",
-        "product": "firefox",
-        "project": "",
-        "purge_actual": "804.49GB",
-        "purge_target": "12GB",
-        "purged_clobber": true,
-        "repository": "",
-        "request_ids": [
-          28007017
-        ],
-        "request_times": {
-          "28007017": 1376504732
-        },
-        "revision": "f83f96c52b11820b317f21d138b5ce3452334082",
-        "scheduler": "cedar-firefox",
-        "slavebuilddir": "ced-osx64-00000000000000000000",
-        "slavename": "bld-lion-r5-012",
-        "stage_platform": "macosx64",
-        "symbolsUrl": "http://ftp.mozilla.org/pub/mozilla.org/firefox/tinderbox-builds/cedar-macosx64/1376504732/firefox-26.0a1.en-US.mac.crashreporter-symbols.zip",
-        "testsUrl": "http://ftp.mozilla.org/pub/mozilla.org/firefox/tinderbox-builds/cedar-macosx64/1376504732/firefox-26.0a1.en-US.mac.tests.zip",
-        "toolsdir": "/builds/slave/ced-osx64-00000000000000000000/tools"
-      },
-      "reason": "scheduler",
-      "request_ids": [
-        28007017
-      ],
-      "requesttime": 1376504732,
-      "result": 0,
-      "slave_id": 2022,
-      "starttime": 1376504739
-    },
-    {
-      "builder_id": 133897,
-      "buildnumber": 15,
-      "endtime": 1376511378,
-      "id": 26553284,
-      "master_id": 76,
-      "properties": {
-        "appName": "Firefox",
-        "appVersion": "26.0a1",
-        "basedir": "/builds/slave/ced-l64-0000000000000000000000",
-        "branch": "test_treeherder",
-        "builddir": "ced-l64-0000000000000000000000",
-        "buildername": "Linux x86-64 cedar build",
-        "buildid": "20130814112532",
-        "buildnumber": 15,
-        "builduid": "f3984a6bbba449f3b35ec692a783fd62",
-        "comments": "Merge m-c -> cedar",
-        "filepath": null,
-        "forced_clobber": false,
-        "got_revision": "f83f96c52b11",
-        "hashType": "sha512",
-        "jsshellUrl": "http://ftp.mozilla.org/pub/mozilla.org/firefox/tinderbox-builds/cedar-linux64/1376504732/jsshell-linux-x86_64.zip",
-        "log_url": "http://ftp.mozilla.org/pub/mozilla.org/firefox/tinderbox-builds/cedar-linux64/1376504732/cedar-linux64-bm57-build1-build15.txt.gz",
-        "master": "http://buildbot-master57.srv.releng.use1.mozilla.com:8001/",
-        "num_ctors": 168,
-        "packageFilename": "firefox-26.0a1.en-US.linux-x86_64.tar.bz2",
-        "packageHash": "47597e0fbf0e43fa9c64d5f5af0bc1145b6513d558390e189369f2b07a8b0425b842bd3026c09f286589092c43858b9cae36d5f7d6bcccde92a623aaa3e90b61",
-        "packageSize": "36967750",
-        "packageUrl": "http://ftp.mozilla.org/pub/mozilla.org/firefox/tinderbox-builds/cedar-linux64/1376504732/firefox-26.0a1.en-US.linux-x86_64.tar.bz2",
-        "periodic_clobber": false,
-        "platform": "linux64",
-        "product": "firefox",
-        "project": "",
-        "purge_actual": "91.38GB",
-        "purge_target": "14GB",
-        "purged_clobber": true,
-        "repository": "",
-        "request_ids": [
-          28007015
-        ],
-        "request_times": {
-          "28007015": 1376504732
-        },
-        "revision": "f83f96c52b11820b317f21d138b5ce3452334082",
-        "scheduler": "cedar-firefox",
-        "slavebuilddir": "ced-l64-0000000000000000000000",
-        "slavename": "bld-linux64-ec2-056",
-        "sourcestamp": "f83f96c52b11",
-        "stage_platform": "linux64",
-        "symbolsUrl": "http://ftp.mozilla.org/pub/mozilla.org/firefox/tinderbox-builds/cedar-linux64/1376504732/firefox-26.0a1.en-US.linux-x86_64.crashreporter-symbols.zip",
-        "testresults": [
-          [
-            "num_ctors",
-            "num_ctors",
-            168,
-            "168"
-          ]
-        ],
-        "testsUrl": "http://ftp.mozilla.org/pub/mozilla.org/firefox/tinderbox-builds/cedar-linux64/1376504732/firefox-26.0a1.en-US.linux-x86_64.tests.zip",
-        "toolsdir": "/builds/slave/ced-l64-0000000000000000000000/tools"
-      },
-      "reason": "scheduler",
-      "request_ids": [
-        28007015
-      ],
-      "requesttime": 1376504732,
-      "result": 0,
-      "slave_id": 2486,
-      "starttime": 1376504740
-    },
-    {
-      "builder_id": 173187,
-      "buildnumber": 9,
-      "endtime": 1376512539,
-      "id": 26553838,
-      "master_id": 76,
-      "properties": {
-        "basedir": "/builds/slave/b2g_ced_emulator_d_dep-0000000",
-        "branch": "test_treeherder",
-        "buildername": "b2g_cedar_emulator_debug_dep",
-        "buildid": "20130814112533",
-        "buildnumber": 9,
-        "builduid": "b437e1a7eaa442f993c8e74888b2b5db",
-        "compare_locales_revision": "1fc4e9bc8287",
-        "gaia_revision": "b0f83dd98f0068168e793e9d91b1be879812f5a1",
-        "gecko_revision": "f83f96c52b11",
-        "log_url": "http://ftp.mozilla.org/pub/mozilla.org/b2g/tinderbox-builds/cedar-emulator_debug/1376504733/b2g_cedar_emulator_debug_dep-bm57-build1-build9.txt.gz",
-        "master": "http://buildbot-master57.srv.releng.use1.mozilla.com:8001/",
-        "platform": "emulator_debug",
-        "product": "b2g",
-        "project": "",
-        "repo_path": "projects/cedar",
-        "repository": "",
-        "request_ids": [
-          28007022
-        ],
-        "request_times": {
-          "28007022": 1376504733
-        },
-        "revision": "f83f96c52b11",
-        "scheduler": "b2g_cedar-b2g",
-        "script_repo_revision": "3a78b256dced",
-        "slavename": "bld-linux64-ec2-044"
-      },
-      "reason": "scheduler",
-      "request_ids": [
-        28007022
-      ],
-      "requesttime": 1376504733,
-      "result": 0,
-      "slave_id": 2304,
-      "starttime": 1376504741
-    },
-    {
-      "builder_id": 142349,
-      "buildnumber": 11,
-      "endtime": 1376512126,
-      "id": 26553653,
-      "master_id": 86,
-      "properties": {
-        "appName": "B2G",
-        "appVersion": "26.0a1",
-        "basedir": "/builds/slave/ced-l64_g-00000000000000000000",
-        "branch": "test_treeherder",
-        "builddir": "ced-l64_g-00000000000000000000",
-        "buildername": "b2g_cedar_linux64_gecko build",
-        "buildid": "20130814112533",
-        "buildnumber": 11,
-        "builduid": "b437e1a7eaa442f993c8e74888b2b5db",
-        "comments": "Merge m-c -> cedar",
-        "forced_clobber": false,
-        "gaiaRepoPath": "/integration/gaia-central",
-        "gaiaRevision": "47daabb93939a38f1f8efa7fdb1cd22ef8445600",
-        "got_revision": "f83f96c52b11",
-        "hashType": "sha512",
-        "log_url": "http://ftp.mozilla.org/pub/mozilla.org/b2g/tinderbox-builds/cedar-linux64_gecko/1376504733/cedar-linux64_gecko-bm63-build1-build11.txt.gz",
-        "master": "http://buildbot-master63.srv.releng.use1.mozilla.com:8001/",
-        "packageUrl": "http://ftp.mozilla.org/pub/mozilla.org/b2g/tinderbox-builds/cedar-linux64_gecko/1376504733/b2g-26.0a1.multi.linux-x86_64.tar.bz2",
-        "periodic_clobber": false,
-        "platform": "linux64_gecko",
-        "product": "b2g",
-        "project": "",
-        "purge_actual": "56.64GB",
-        "purge_target": "13GB",
-        "purged_clobber": true,
-        "repository": "",
-        "request_ids": [
-          28007024
-        ],
-        "request_times": {
-          "28007024": 1376504733
-        },
-        "revision": "f83f96c52b11820b317f21d138b5ce3452334082",
-        "scheduler": "b2g_cedar-b2g",
-        "slavebuilddir": "ced-l64_g-00000000000000000000",
-        "slavename": "bld-linux64-ec2-094",
-        "stage_platform": "linux64_gecko",
-        "symbolsUrl": "http://ftp.mozilla.org/pub/mozilla.org/b2g/tinderbox-builds/cedar-linux64_gecko/1376504733/en-US/b2g-26.0a1.en-US.linux-x86_64.crashreporter-symbols.zip",
-        "testsUrl": "http://ftp.mozilla.org/pub/mozilla.org/b2g/tinderbox-builds/cedar-linux64_gecko/1376504733/b2g-26.0a1.multi.linux-x86_64.tests.zip",
-        "toolsdir": "/builds/slave/ced-l64_g-00000000000000000000/tools"
-      },
-      "reason": "scheduler",
-      "request_ids": [
-        28007024
-      ],
-      "requesttime": 1376504733,
-      "result": 0,
-      "slave_id": 2548,
-      "starttime": 1376504742
-    },
-    {
-      "builder_id": 127562,
-      "buildnumber": 11,
-      "endtime": 1376511382,
-      "id": 26553288,
-      "master_id": 86,
-      "properties": {
-        "appName": "B2G",
-        "appVersion": "26.0a1",
-        "basedir": "/builds/slave/ced-linux32_g-0000000000000000",
-        "branch": "test_treeherder",
-        "builddir": "ced-linux32_g-0000000000000000",
-        "buildername": "b2g_cedar_linux32_gecko build",
-        "buildid": "20130814112533",
-        "buildnumber": 11,
-        "builduid": "b437e1a7eaa442f993c8e74888b2b5db",
-        "comments": "Merge m-c -> cedar",
-        "forced_clobber": false,
-        "gaiaRepoPath": "/integration/gaia-central",
-        "gaiaRevision": "47daabb93939a38f1f8efa7fdb1cd22ef8445600",
-        "got_revision": "f83f96c52b11",
-        "hashType": "sha512",
-        "log_url": "http://ftp.mozilla.org/pub/mozilla.org/b2g/tinderbox-builds/cedar-linux32_gecko/1376504733/cedar-linux32_gecko-bm63-build1-build11.txt.gz",
-        "master": "http://buildbot-master63.srv.releng.use1.mozilla.com:8001/",
-        "packageUrl": "http://ftp.mozilla.org/pub/mozilla.org/b2g/tinderbox-builds/cedar-linux32_gecko/1376504733/b2g-26.0a1.multi.linux-i686.tar.bz2",
-        "periodic_clobber": false,
-        "platform": "linux32_gecko",
-        "product": "b2g",
-        "project": "",
-        "purge_actual": "69.44GB",
-        "purge_target": "13GB",
-        "purged_clobber": true,
-        "repository": "",
-        "request_ids": [
-          28007023
-        ],
-        "request_times": {
-          "28007023": 1376504733
-        },
-        "revision": "f83f96c52b11820b317f21d138b5ce3452334082",
-        "scheduler": "b2g_cedar-b2g",
-        "slavebuilddir": "ced-linux32_g-0000000000000000",
-        "slavename": "bld-linux64-ec2-038",
-        "stage_platform": "linux32_gecko",
-        "symbolsUrl": "http://ftp.mozilla.org/pub/mozilla.org/b2g/tinderbox-builds/cedar-linux32_gecko/1376504733/en-US/b2g-26.0a1.en-US.linux-i686.crashreporter-symbols.zip",
-        "testsUrl": "http://ftp.mozilla.org/pub/mozilla.org/b2g/tinderbox-builds/cedar-linux32_gecko/1376504733/b2g-26.0a1.multi.linux-i686.tests.zip",
-        "toolsdir": "/builds/slave/ced-linux32_g-0000000000000000/tools"
-      },
-      "reason": "scheduler",
-      "request_ids": [
-        28007023
-      ],
-      "requesttime": 1376504733,
-      "result": 0,
-      "slave_id": 2255,
-      "starttime": 1376504745
-    },
-    {
-      "builder_id": 137295,
-      "buildnumber": 9,
-      "endtime": 1376511182,
-      "id": 26553177,
-      "master_id": 77,
-      "properties": {
-        "basedir": "/builds/slave/b2g_ced_unagi_dep-000000000000",
-        "branch": "test_treeherder",
-        "buildername": "b2g_cedar_unagi_dep",
-        "buildid": "20130814112533",
-        "buildnumber": 9,
-        "builduid": "b437e1a7eaa442f993c8e74888b2b5db",
-        "compare_locales_revision": "1fc4e9bc8287",
-        "gaia_revision": "b0f83dd98f0068168e793e9d91b1be879812f5a1",
-        "gecko_revision": "f83f96c52b11",
-        "log_url": "http://ftp.mozilla.org/pub/mozilla.org/b2g/tinderbox-builds/cedar-unagi/1376504733/b2g_cedar_unagi_dep-bm58-build1-build9.txt.gz",
-        "master": "http://buildbot-master58.srv.releng.usw2.mozilla.com:8001/",
-        "platform": "unagi",
-        "product": "b2g",
-        "project": "",
-        "repo_path": "projects/cedar",
-        "repository": "",
-        "request_ids": [
-          28007026
-        ],
-        "request_times": {
-          "28007026": 1376504733
-        },
-        "revision": "f83f96c52b11",
-        "scheduler": "b2g_cedar-b2g",
-        "script_repo_revision": "3a78b256dced",
-        "slavename": "bld-centos6-hp-019"
-      },
-      "reason": "scheduler",
-      "request_ids": [
-        28007026
-      ],
-      "requesttime": 1376504733,
-      "result": 0,
-      "slave_id": 1936,
-      "starttime": 1376504752
-    },
-    {
-      "builder_id": 115926,
-      "buildnumber": 557,
-      "endtime": 1376512062,
-      "id": 26553618,
-      "master_id": 72,
-      "properties": {
-        "basedir": "/builds/slave/test",
-        "branch": "test_treeherder",
-        "build_url": "http://ftp.mozilla.org/pub/mozilla.org/firefox/try-builds/jandemooij@gmail.com-5974016d7c49/try-linux-debug/firefox-26.0a1.en-US.linux-i686.tar.bz2",
-        "buildername": "Ubuntu VM 12.04 try debug test reftest",
-        "buildid": "20130814095056",
-        "buildnumber": 557,
-        "builduid": "df39c7ff9efc416d93cd988e6c9503e8",
-        "log_url": "http://ftp.mozilla.org/pub/mozilla.org/firefox/try-builds/jandemooij@gmail.com-5974016d7c49/try-linux-debug/try_ubuntu32_vm-debug_test-reftest-bm52-tests1-linux-build557.txt.gz",
-        "master": "http://buildbot-master52.srv.releng.use1.mozilla.com:8201/",
-        "platform": "linux",
-        "product": "firefox",
-        "project": "",
-        "repo_path": "try",
-        "repository": "",
-        "request_ids": [
-          28006997
-        ],
-        "request_times": {
-          "28006997": 1376504656
-        },
-        "revision": "5974016d7c49",
-        "scheduler": "tests-try-ubuntu32_vm-debug-unittest",
-        "script_repo_revision": "2f85052176f2",
-        "slavebuilddir": "test",
-        "slavename": "tst-linux32-ec2-052",
-        "stage_platform": "linux",
-        "symbols_url": "http://ftp.mozilla.org/pub/mozilla.org/firefox/try-builds/jandemooij@gmail.com-5974016d7c49/try-linux-debug/firefox-26.0a1.en-US.linux-i686.crashreporter-symbols.zip"
-      },
-      "reason": "scheduler",
-      "request_ids": [
-        28006997
-      ],
-      "requesttime": 1376504656,
-      "result": 2,
-      "slave_id": 3387,
-      "starttime": 1376504752
-    },
-    {
-      "builder_id": 140859,
-      "buildnumber": 90,
-      "endtime": 1376510963,
-      "id": 26553073,
-      "master_id": 71,
-      "properties": {
-        "basedir": "/home/cltbld/talos-slave/test",
-        "branch": "test_treeherder",
-        "build_url": "http://ftp.mozilla.org/pub/mozilla.org/firefox/tinderbox-builds/fx-team-linux64-debug/1376499083/firefox-26.0a1.en-US.linux-x86_64.tar.bz2",
-        "buildername": "Rev3 Fedora 12x64 fx-team debug test mochitest-browser-chrome",
-        "buildid": "20130814095123",
-        "buildnumber": 90,
-        "builduid": "e8c4556f1c404e6786b4d1e975cdf9c5",
-        "log_url": "http://ftp.mozilla.org/pub/mozilla.org/firefox/tinderbox-builds/fx-team-linux64-debug/1376499083/fx-team_fedora64-debug_test-mochitest-browser-chrome-bm51-tests1-linux-build90.txt.gz",
-        "master": "http://buildbot-master51.srv.releng.use1.mozilla.com:8201/",
-        "pgo_build": "False",
-        "platform": "linux64",
-        "product": "firefox",
-        "project": "",
-        "repo_path": "integration/fx-team",
-        "repository": "",
-        "request_ids": [
-          28007286
-        ],
-        "request_times": {
-          "28007286": 1376505173
-        },
-        "revision": "15aaeae89597",
-        "scheduler": "tests-fx-team-fedora64-debug-unittest",
-        "script_repo_revision": "2f85052176f2",
-        "slavebuilddir": "test",
-        "slavename": "talos-r3-fed64-049",
-        "stage_platform": "linux64",
-        "symbols_url": "http://ftp.mozilla.org/pub/mozilla.org/firefox/tinderbox-builds/fx-team-linux64-debug/1376499083/firefox-26.0a1.en-US.linux-x86_64.crashreporter-symbols.zip"
-      },
-      "reason": "scheduler",
-      "request_ids": [
-        28007286
-      ],
-      "requesttime": 1376505173,
-      "result": 0,
-      "slave_id": 740,
-      "starttime": 1376505175
-    },
-    {
-      "builder_id": 140866,
-      "buildnumber": 566,
-      "endtime": 1376512331,
-      "id": 26553741,
-      "master_id": 74,
-      "properties": {
-        "basedir": "/home/cltbld/talos-slave/test",
-        "branch": "test_treeherder",
-        "build_url": "http://ftp.mozilla.org/pub/mozilla.org/firefox/try-builds/bhackett@mozilla.com-b565da72f936/try-linux-debug/firefox-26.0a1.en-US.linux-i686.tar.bz2",
-        "buildername": "Rev3 Fedora 12 try debug test mochitest-browser-chrome",
-        "buildid": "20130814095330",
-        "buildnumber": 566,
-        "builduid": "5cf2b4c25ed5401396ed4e95689680d5",
-        "log_url": "http://ftp.mozilla.org/pub/mozilla.org/firefox/try-builds/bhackett@mozilla.com-b565da72f936/try-linux-debug/try_fedora-debug_test-mochitest-browser-chrome-bm53-tests1-linux-build566.txt.gz",
-        "master": "http://buildbot-master53.srv.releng.usw2.mozilla.com:8201/",
-        "platform": "linux",
-        "product": "firefox",
-        "project": "",
-        "repo_path": "try",
-        "repository": "",
-        "request_ids": [
-          28007499
-        ],
-        "request_times": {
-          "28007499": 1376505250
-        },
-        "revision": "b565da72f936",
-        "scheduler": "tests-try-fedora-debug-unittest",
-        "script_repo_revision": "2f85052176f2",
-        "slavebuilddir": "test",
-        "slavename": "talos-r3-fed-052",
-        "stage_platform": "linux",
-        "symbols_url": "http://ftp.mozilla.org/pub/mozilla.org/firefox/try-builds/bhackett@mozilla.com-b565da72f936/try-linux-debug/firefox-26.0a1.en-US.linux-i686.crashreporter-symbols.zip"
-      },
-      "reason": "scheduler",
-      "request_ids": [
-        28007499
-      ],
-      "requesttime": 1376505250,
-      "result": 0,
-      "slave_id": 936,
-      "starttime": 1376505251
-    },
-    {
-      "builder_id": 141333,
-      "buildnumber": 285,
-      "endtime": 1376511095,
-      "id": 26553135,
-      "master_id": 72,
-      "properties": {
-        "basedir": "/home/cltbld/talos-slave/test",
-        "branch": "test_treeherder",
-        "build_url": "http://ftp.mozilla.org/pub/mozilla.org/firefox/tinderbox-builds/mozilla-inbound-linux64-debug/1376499321/firefox-26.0a1.en-US.linux-x86_64.tar.bz2",
-        "buildername": "Rev3 Fedora 12x64 mozilla-inbound debug test mochitest-browser-chrome",
-        "buildid": "20130814095521",
-        "buildnumber": 285,
-        "builduid": "02c611f9a35f4834812ecb7ad6306871",
-        "log_url": "http://ftp.mozilla.org/pub/mozilla.org/firefox/tinderbox-builds/mozilla-inbound-linux64-debug/1376499321/mozilla-inbound_fedora64-debug_test-mochitest-browser-chrome-bm52-tests1-linux-build285.txt.gz",
-        "master": "http://buildbot-master52.srv.releng.use1.mozilla.com:8201/",
-        "pgo_build": "False",
-        "platform": "linux64",
-        "product": "firefox",
-        "project": "",
-        "repo_path": "integration/mozilla-inbound",
-        "repository": "",
-        "request_ids": [
-          28007472
-        ],
-        "request_times": {
-          "28007472": 1376505245
-        },
-        "revision": "c218913f49b2",
-        "scheduler": "tests-mozilla-inbound-fedora64-debug-unittest",
-        "script_repo_revision": "2f85052176f2",
-        "slavebuilddir": "test",
-        "slavename": "talos-r3-fed64-046",
-        "stage_platform": "linux64",
-        "symbols_url": "http://ftp.mozilla.org/pub/mozilla.org/firefox/tinderbox-builds/mozilla-inbound-linux64-debug/1376499321/firefox-26.0a1.en-US.linux-x86_64.crashreporter-symbols.zip"
-      },
-      "reason": "scheduler",
-      "request_ids": [
-        28007472
-      ],
-      "requesttime": 1376505245,
-      "result": 0,
-      "slave_id": 741,
-      "starttime": 1376505275
-    },
-    {
-      "builder_id": 115926,
-      "buildnumber": 558,
-      "endtime": 1376512557,
-      "id": 26553856,
-      "master_id": 72,
-      "properties": {
-        "basedir": "/builds/slave/test",
-        "branch": "test_treeherder",
-        "build_url": "http://ftp.mozilla.org/pub/mozilla.org/firefox/try-builds/bhackett@mozilla.com-b565da72f936/try-linux-debug/firefox-26.0a1.en-US.linux-i686.tar.bz2",
-        "buildername": "Ubuntu VM 12.04 try debug test reftest",
-        "buildid": "20130814095330",
-        "buildnumber": 558,
-        "builduid": "5cf2b4c25ed5401396ed4e95689680d5",
-        "log_url": "http://ftp.mozilla.org/pub/mozilla.org/firefox/try-builds/bhackett@mozilla.com-b565da72f936/try-linux-debug/try_ubuntu32_vm-debug_test-reftest-bm52-tests1-linux-build558.txt.gz",
-        "master": "http://buildbot-master52.srv.releng.use1.mozilla.com:8201/",
-        "platform": "linux",
-        "product": "firefox",
-        "project": "",
-        "repo_path": "try",
-        "repository": "",
-        "request_ids": [
-          28007500
-        ],
-        "request_times": {
-          "28007500": 1376505250
-        },
-        "revision": "b565da72f936",
-        "scheduler": "tests-try-ubuntu32_vm-debug-unittest",
-        "script_repo_revision": "2f85052176f2",
-        "slavebuilddir": "test",
-        "slavename": "tst-linux32-ec2-071",
-        "stage_platform": "linux",
-        "symbols_url": "http://ftp.mozilla.org/pub/mozilla.org/firefox/try-builds/bhackett@mozilla.com-b565da72f936/try-linux-debug/firefox-26.0a1.en-US.linux-i686.crashreporter-symbols.zip"
-      },
-      "reason": "scheduler",
-      "request_ids": [
-        28007500
-      ],
-      "requesttime": 1376505250,
-      "result": 2,
-      "slave_id": 3467,
-      "starttime": 1376505276
-    },
-    {
-      "builder_id": 141013,
-      "buildnumber": 804,
-      "endtime": 1376511070,
-      "id": 26553121,
-      "master_id": 79,
-      "properties": {
-        "basedir": "/home/cltbld/talos-slave/test",
-        "branch": "test_treeherder",
-        "build_url": "http://ftp.mozilla.org/pub/mozilla.org/firefox/try-builds/bhackett@mozilla.com-b565da72f936/try-linux64-debug/firefox-26.0a1.en-US.linux-x86_64.tar.bz2",
-        "buildername": "Rev3 Fedora 12x64 try debug test mochitest-browser-chrome",
-        "buildid": "20130814095300",
-        "buildnumber": 804,
-        "builduid": "5cf2b4c25ed5401396ed4e95689680d5",
-        "log_url": "http://ftp.mozilla.org/pub/mozilla.org/firefox/try-builds/bhackett@mozilla.com-b565da72f936/try-linux64-debug/try_fedora64-debug_test-mochitest-browser-chrome-bm68-tests1-linux-build804.txt.gz",
-        "master": "http://buildbot-master68.srv.releng.usw2.mozilla.com:8201/",
-        "platform": "linux64",
-        "product": "firefox",
-        "project": "",
-        "repo_path": "try",
-        "repository": "",
-        "request_ids": [
-          28007512
-        ],
-        "request_times": {
-          "28007512": 1376505305
-        },
-        "revision": "b565da72f936",
-        "scheduler": "tests-try-fedora64-debug-unittest",
-        "script_repo_revision": "2f85052176f2",
-        "slavebuilddir": "test",
-        "slavename": "talos-r3-fed64-050",
-        "stage_platform": "linux64",
-        "symbols_url": "http://ftp.mozilla.org/pub/mozilla.org/firefox/try-builds/bhackett@mozilla.com-b565da72f936/try-linux64-debug/firefox-26.0a1.en-US.linux-x86_64.crashreporter-symbols.zip"
-      },
-      "reason": "scheduler",
-      "request_ids": [
-        28007512
-      ],
-      "requesttime": 1376505305,
-      "result": 0,
-      "slave_id": 758,
-      "starttime": 1376505308
-    },
-    {
-      "builder_id": 141013,
-      "buildnumber": 805,
-      "endtime": 1376511158,
-      "id": 26553163,
-      "master_id": 79,
-      "properties": {
-        "basedir": "/home/cltbld/talos-slave/test",
-        "branch": "test_treeherder",
-        "build_url": "http://ftp.mozilla.org/pub/mozilla.org/firefox/try-builds/jseward@mozilla.com-056d9a2ff795/try-linux64-debug/firefox-26.0a1.en-US.linux-x86_64.tar.bz2",
-        "buildername": "Rev3 Fedora 12x64 try debug test mochitest-browser-chrome",
-        "buildid": "20130814080622",
-        "buildnumber": 805,
-        "builduid": "0318bafdd6a044a6a863a2910755198a",
-        "log_url": "http://ftp.mozilla.org/pub/mozilla.org/firefox/try-builds/jseward@mozilla.com-056d9a2ff795/try-linux64-debug/try_fedora64-debug_test-mochitest-browser-chrome-bm68-tests1-linux-build805.txt.gz",
-        "master": "http://buildbot-master68.srv.releng.usw2.mozilla.com:8201/",
-        "platform": "linux64",
-        "product": "firefox",
-        "project": "",
-        "repo_path": "try",
-        "repository": "",
-        "request_ids": [
-          28007525
-        ],
-        "request_times": {
-          "28007525": 1376505356
-        },
-        "revision": "056d9a2ff795",
-        "scheduler": "tests-try-fedora64-debug-unittest",
-        "script_repo_revision": "2f85052176f2",
-        "slavebuilddir": "test",
-        "slavename": "talos-r3-fed64-063",
-        "stage_platform": "linux64",
-        "symbols_url": "http://ftp.mozilla.org/pub/mozilla.org/firefox/try-builds/jseward@mozilla.com-056d9a2ff795/try-linux64-debug/firefox-26.0a1.en-US.linux-x86_64.crashreporter-symbols.zip"
-      },
-      "reason": "Self-serve: Rebuilt by jseward@mozilla.com",
-      "request_ids": [
-        28007525
-      ],
-      "requesttime": 1376505356,
-      "result": 0,
-      "slave_id": 1752,
-      "starttime": 1376505360
-    },
-    {
-      "builder_id": 140862,
-      "buildnumber": 488,
-      "endtime": 1376511495,
-      "id": 26553352,
-      "master_id": 71,
-      "properties": {
-        "basedir": "/home/cltbld/talos-slave/test",
-        "branch": "test_treeherder",
-        "build_url": "http://ftp.mozilla.org/pub/mozilla.org/firefox/try-builds/jcoppeard@mozilla.com-d4a17013c2e7/try-linux64-debug/firefox-26.0a1.en-US.linux-x86_64.tar.bz2",
-        "buildername": "Rev3 Fedora 12x64 try debug test mochitest-browser-chrome",
-        "buildid": "20130814095747",
-        "buildnumber": 488,
-        "builduid": "61e0d19026e145fdbba02c16d7d4d8d6",
-        "log_url": "http://ftp.mozilla.org/pub/mozilla.org/firefox/try-builds/jcoppeard@mozilla.com-d4a17013c2e7/try-linux64-debug/try_fedora64-debug_test-mochitest-browser-chrome-bm51-tests1-linux-build488.txt.gz",
-        "master": "http://buildbot-master51.srv.releng.use1.mozilla.com:8201/",
-        "platform": "linux64",
-        "product": "firefox",
-        "project": "",
-        "repo_path": "try",
-        "repository": "",
-        "request_ids": [
-          28007564
-        ],
-        "request_times": {
-          "28007564": 1376505488
-        },
-        "revision": "d4a17013c2e7",
-        "scheduler": "tests-try-fedora64-debug-unittest",
-        "script_repo_revision": "2f85052176f2",
-        "slavebuilddir": "test",
-        "slavename": "talos-r3-fed64-047",
-        "stage_platform": "linux64",
-        "symbols_url": "http://ftp.mozilla.org/pub/mozilla.org/firefox/try-builds/jcoppeard@mozilla.com-d4a17013c2e7/try-linux64-debug/firefox-26.0a1.en-US.linux-x86_64.crashreporter-symbols.zip"
-      },
-      "reason": "scheduler",
-      "request_ids": [
-        28007564
-      ],
-      "requesttime": 1376505488,
-      "result": 0,
-      "slave_id": 742,
-      "starttime": 1376505495
-    },
-    {
-      "builder_id": 139902,
-      "buildnumber": 813,
-      "endtime": 1376512818,
-      "id": 26553967,
-      "master_id": 74,
-      "properties": {
-        "basedir": "/home/cltbld/talos-slave/test",
-        "branch": "test_treeherder",
-        "build_url": "http://ftp.mozilla.org/pub/mozilla.org/firefox/tinderbox-builds/mozilla-inbound-linux-debug/1376500461/firefox-26.0a1.en-US.linux-i686.tar.bz2",
-        "buildername": "Rev3 Fedora 12 mozilla-inbound debug test mochitest-browser-chrome",
-        "buildid": "20130814101421",
-        "buildnumber": 813,
-        "builduid": "5e66a9f140074af7b494fe2c7c1d54af",
-        "log_url": "http://ftp.mozilla.org/pub/mozilla.org/firefox/tinderbox-builds/mozilla-inbound-linux-debug/1376500461/mozilla-inbound_fedora-debug_test-mochitest-browser-chrome-bm53-tests1-linux-build813.txt.gz",
-        "master": "http://buildbot-master53.srv.releng.usw2.mozilla.com:8201/",
-        "pgo_build": "False",
-        "platform": "linux",
-        "product": "firefox",
-        "project": "",
-        "repo_path": "integration/mozilla-inbound",
-        "repository": "",
-        "request_ids": [
-          28007667
-        ],
-        "request_times": {
-          "28007667": 1376505595
-        },
-        "revision": "1fa1f28c6edf",
-        "scheduler": "tests-mozilla-inbound-fedora-debug-unittest",
-        "script_repo_revision": "2f85052176f2",
-        "slavebuilddir": "test",
-        "slavename": "talos-r3-fed-050",
-        "stage_platform": "linux",
-        "symbols_url": "http://ftp.mozilla.org/pub/mozilla.org/firefox/tinderbox-builds/mozilla-inbound-linux-debug/1376500461/firefox-26.0a1.en-US.linux-i686.crashreporter-symbols.zip"
-      },
-      "reason": "scheduler",
-      "request_ids": [
-        28007667
-      ],
-      "requesttime": 1376505595,
-      "result": 0,
-      "slave_id": 700,
-      "starttime": 1376505601
-    },
-    {
-      "builder_id": 115759,
-      "buildnumber": 750,
-      "endtime": 1376512431,
-      "id": 26553784,
-      "master_id": 73,
-      "properties": {
-        "basedir": "/builds/slave/test",
-        "branch": "test_treeherder",
-        "build_url": "http://ftp.mozilla.org/pub/mozilla.org/firefox/tinderbox-builds/mozilla-inbound-linux-debug/1376500461/firefox-26.0a1.en-US.linux-i686.tar.bz2",
-        "buildername": "Ubuntu VM 12.04 mozilla-inbound debug test reftest",
-        "buildid": "20130814101421",
-        "buildnumber": 750,
-        "builduid": "5e66a9f140074af7b494fe2c7c1d54af",
-        "log_url": "http://ftp.mozilla.org/pub/mozilla.org/firefox/tinderbox-builds/mozilla-inbound-linux-debug/1376500461/mozilla-inbound_ubuntu32_vm-debug_test-reftest-bm54-tests1-linux-build750.txt.gz",
-        "master": "http://buildbot-master54.srv.releng.usw2.mozilla.com:8201/",
-        "pgo_build": "False",
-        "platform": "linux",
-        "product": "firefox",
-        "project": "",
-        "repo_path": "integration/mozilla-inbound",
-        "repository": "",
-        "request_ids": [
-          28007674
-        ],
-        "request_times": {
-          "28007674": 1376505597
-        },
-        "revision": "1fa1f28c6edf",
-        "scheduler": "tests-mozilla-inbound-ubuntu32_vm-debug-unittest",
-        "script_repo_revision": "2f85052176f2",
-        "slavebuilddir": "test",
-        "slavename": "tst-linux32-ec2-322",
-        "stage_platform": "linux",
-        "symbols_url": "http://ftp.mozilla.org/pub/mozilla.org/firefox/tinderbox-builds/mozilla-inbound-linux-debug/1376500461/firefox-26.0a1.en-US.linux-i686.crashreporter-symbols.zip"
-      },
-      "reason": "scheduler",
-      "request_ids": [
-        28007674
-      ],
-      "requesttime": 1376505597,
-      "result": 0,
-      "slave_id": 3598,
-      "starttime": 1376505602
-    },
-    {
-      "builder_id": 140866,
-      "buildnumber": 567,
-      "endtime": 1376512827,
-      "id": 26553972,
-      "master_id": 74,
-      "properties": {
-        "basedir": "/home/cltbld/talos-slave/test",
-        "branch": "test_treeherder",
-        "build_url": "http://ftp.mozilla.org/pub/mozilla.org/firefox/try-builds/jcoppeard@mozilla.com-d4a17013c2e7/try-linux-debug/firefox-26.0a1.en-US.linux-i686.tar.bz2",
-        "buildername": "Rev3 Fedora 12 try debug test mochitest-browser-chrome",
-        "buildid": "20130814100114",
-        "buildnumber": 567,
-        "builduid": "61e0d19026e145fdbba02c16d7d4d8d6",
-        "log_url": "http://ftp.mozilla.org/pub/mozilla.org/firefox/try-builds/jcoppeard@mozilla.com-d4a17013c2e7/try-linux-debug/try_fedora-debug_test-mochitest-browser-chrome-bm53-tests1-linux-build567.txt.gz",
-        "master": "http://buildbot-master53.srv.releng.usw2.mozilla.com:8201/",
-        "platform": "linux",
-        "product": "firefox",
-        "project": "",
-        "repo_path": "try",
-        "repository": "",
-        "request_ids": [
-          28007711
-        ],
-        "request_times": {
-          "28007711": 1376505605
-        },
-        "revision": "d4a17013c2e7",
-        "scheduler": "tests-try-fedora-debug-unittest",
-        "script_repo_revision": "2f85052176f2",
-        "slavebuilddir": "test",
-        "slavename": "talos-r3-fed-097",
-        "stage_platform": "linux",
-        "symbols_url": "http://ftp.mozilla.org/pub/mozilla.org/firefox/try-builds/jcoppeard@mozilla.com-d4a17013c2e7/try-linux-debug/firefox-26.0a1.en-US.linux-i686.crashreporter-symbols.zip"
-      },
-      "reason": "scheduler",
-      "request_ids": [
-        28007711
-      ],
-      "requesttime": 1376505605,
-      "result": 0,
-      "slave_id": 3651,
-      "starttime": 1376505607
-    },
-    {
-      "builder_id": 121375,
-      "buildnumber": 575,
-      "endtime": 1376512669,
-      "id": 26553910,
-      "master_id": 80,
-      "properties": {
-        "basedir": "/builds/slave/test",
-        "branch": "test_treeherder",
-        "build_url": "http://ftp.mozilla.org/pub/mozilla.org/firefox/try-builds/jcoppeard@mozilla.com-d4a17013c2e7/try-linux-debug/firefox-26.0a1.en-US.linux-i686.tar.bz2",
-        "buildername": "Ubuntu VM 12.04 try debug test reftest",
-        "buildid": "20130814100114",
-        "buildnumber": 575,
-        "builduid": "61e0d19026e145fdbba02c16d7d4d8d6",
-        "log_url": "http://ftp.mozilla.org/pub/mozilla.org/firefox/try-builds/jcoppeard@mozilla.com-d4a17013c2e7/try-linux-debug/try_ubuntu32_vm-debug_test-reftest-bm67-tests1-linux-build575.txt.gz",
-        "master": "http://buildbot-master67.srv.releng.use1.mozilla.com:8201/",
-        "platform": "linux",
-        "product": "firefox",
-        "project": "",
-        "repo_path": "try",
-        "repository": "",
-        "request_ids": [
-          28007712
-        ],
-        "request_times": {
-          "28007712": 1376505606
-        },
-        "revision": "d4a17013c2e7",
-        "scheduler": "tests-try-ubuntu32_vm-debug-unittest",
-        "script_repo_revision": "2f85052176f2",
-        "slavebuilddir": "test",
-        "slavename": "tst-linux32-ec2-006",
-        "stage_platform": "linux",
-        "symbols_url": "http://ftp.mozilla.org/pub/mozilla.org/firefox/try-builds/jcoppeard@mozilla.com-d4a17013c2e7/try-linux-debug/firefox-26.0a1.en-US.linux-i686.crashreporter-symbols.zip"
-      },
-      "reason": "scheduler",
-      "request_ids": [
-        28007712
-      ],
-      "requesttime": 1376505606,
-      "result": 0,
-      "slave_id": 3352,
-      "starttime": 1376505616
-    },
-    {
-      "builder_id": 143336,
-      "buildnumber": 1796,
-      "endtime": 1376513223,
-      "id": 26554142,
-      "master_id": 78,
-      "properties": {
-        "appName": "Firefox",
-        "appVersion": "26.0a1",
-        "basedir": "/builds/slave/try-osx64-00000000000000000000",
-        "branch": "test_treeherder",
-        "builddir": "try-osx64-00000000000000000000",
-        "buildername": "OS X 10.7 try build",
-        "buildid": "20130814114039",
-        "buildnumber": 1796,
-        "builduid": "39570642c2fe4206bc6ac9ecdb37d740",
-        "comments": "try: -b do -p linux64,macosx64,win32 -u xpcshell,mochitests -t none",
-        "filepath": null,
-        "forced_clobber": false,
-        "got_revision": "2be7ca1d1d99",
-        "hashType": "sha512",
-        "jsshellUrl": "http://ftp.mozilla.org/pub/mozilla.org/firefox/try-builds/mixedpuppy@gmail.com-2be7ca1d1d99/try-macosx64/jsshell-mac.zip",
-        "log_url": "http://ftp.mozilla.org/pub/mozilla.org/firefox/try-builds/mixedpuppy@gmail.com-2be7ca1d1d99/try-macosx64/try-macosx64-bm56-try1-build1796.txt.gz",
-        "master": "http://buildbot-master56.srv.releng.usw2.mozilla.com:8101/",
-        "packageFilename": "firefox-26.0a1.en-US.mac.dmg",
-        "packageHash": "dd2d3dcd5dd35b0c3ca29e501d9352ad76445a5457c074b2b9b7b196ca6f20fef6b8122a33af380208d8fa93394b755fc88840ccacb4d55d4c2c94e4c24bee7f",
-        "packageSize": "62721199",
-        "packageUrl": "http://ftp.mozilla.org/pub/mozilla.org/firefox/try-builds/mixedpuppy@gmail.com-2be7ca1d1d99/try-macosx64/firefox-26.0a1.en-US.mac.dmg",
-        "periodic_clobber": false,
-        "platform": "macosx64",
-        "product": "firefox",
-        "project": "",
-        "purge_actual": "876.10GB",
-        "purge_target": "12GB",
-        "purged_clobber": false,
-        "repository": "",
-        "request_ids": [
-          28007726
-        ],
-        "request_times": {
-          "28007726": 1376505622
-        },
-        "revision": "2be7ca1d1d9985374a8dedc8b0e1d5ee3d2cea00",
-        "scheduler": "try-firefox",
-        "slavebuilddir": "try-osx64-00000000000000000000",
-        "slavename": "bld-lion-r5-035",
-        "stage_platform": "macosx64",
-        "symbolsUrl": "http://ftp.mozilla.org/pub/mozilla.org/firefox/try-builds/mixedpuppy@gmail.com-2be7ca1d1d99/try-macosx64/firefox-26.0a1.en-US.mac.crashreporter-symbols.zip",
-        "testsUrl": "http://ftp.mozilla.org/pub/mozilla.org/firefox/try-builds/mixedpuppy@gmail.com-2be7ca1d1d99/try-macosx64/firefox-26.0a1.en-US.mac.tests.zip",
-        "toolsdir": "/builds/slave/try-osx64-00000000000000000000/tools",
-        "who": "mixedpuppy@gmail.com"
-      },
-      "reason": "scheduler",
-      "request_ids": [
-        28007726
-      ],
-      "requesttime": 1376505622,
-      "result": 0,
-      "slave_id": 2032,
-      "starttime": 1376505625
-    },
-    {
-      "builder_id": 143762,
-      "buildnumber": 2080,
-      "endtime": 1376514323,
-      "id": 26554806,
-      "master_id": 78,
-      "properties": {
-        "appName": "Firefox",
-        "appVersion": "26.0a1",
-        "basedir": "e:/builds/moz2_slave/try-w32-d-00000000000000000000",
-        "branch": "test_treeherder",
-        "builddir": "try-w32-d-00000000000000000000",
-        "buildername": "WINNT 5.2 try leak test build",
-        "buildid": "20130814114050",
-        "buildnumber": 2080,
-        "builduid": "39570642c2fe4206bc6ac9ecdb37d740",
-        "comments": "try: -b do -p linux64,macosx64,win32 -u xpcshell,mochitests -t none",
-        "filepath": null,
-        "forced_clobber": false,
-        "got_revision": "2be7ca1d1d99",
-        "hashType": "sha512",
-        "installerFilename": "firefox-26.0a1.en-US.win32.installer.exe",
-        "installerHash": "89eb6593bad2ea4cae311e15b22b09c1c2f59077dc4b2c3f967450d3fbb8579a5a01e78df7df5ec02d1eac9727168b97ac93b04375a174bab1f437d77655136c",
-        "installerSize": "27515632",
-        "jsshellUrl": "http://ftp.mozilla.org/pub/mozilla.org/firefox/try-builds/mixedpuppy@gmail.com-2be7ca1d1d99/try-win32-debug/jsshell-win32.zip",
-        "log_url": "http://ftp.mozilla.org/pub/mozilla.org/firefox/try-builds/mixedpuppy@gmail.com-2be7ca1d1d99/try-win32-debug/try-win32-debug-bm56-try1-build2080.txt.gz",
-        "master": "http://buildbot-master56.srv.releng.usw2.mozilla.com:8101/",
-        "packageFilename": "firefox-26.0a1.en-US.win32.zip",
-        "packageHash": "d778554a02afbfa59788b8a9c6639f3932c6be0cfa349025382a2d01d4b3ccc630ed0c046b00f7e3caed238e7cbf917e4b68f317597ae912bd0d4e5e58e40729",
-        "packageSize": "38070440",
-        "packageUrl": "http://ftp.mozilla.org/pub/mozilla.org/firefox/try-builds/mixedpuppy@gmail.com-2be7ca1d1d99/try-win32-debug/firefox-26.0a1.en-US.win32.zip",
-        "periodic_clobber": false,
-        "platform": "win32-debug",
-        "product": "firefox",
-        "project": "",
-        "purge_actual": "157.16GB",
-        "purge_target": "9GB",
-        "purged_clobber": false,
-        "repository": "",
-        "request_ids": [
-          28007724
-        ],
-        "request_times": {
-          "28007724": 1376505622
-        },
-        "revision": "2be7ca1d1d9985374a8dedc8b0e1d5ee3d2cea00",
-        "scheduler": "try-firefox",
-        "slavebuilddir": "try-w32-d-00000000000000000000",
-        "slavename": "w64-ix-slave48",
-        "stage_platform": "win32-debug",
-        "symbolsUrl": "http://ftp.mozilla.org/pub/mozilla.org/firefox/try-builds/mixedpuppy@gmail.com-2be7ca1d1d99/try-win32-debug/firefox-26.0a1.en-US.win32.crashreporter-symbols.zip",
-        "testsUrl": "http://ftp.mozilla.org/pub/mozilla.org/firefox/try-builds/mixedpuppy@gmail.com-2be7ca1d1d99/try-win32-debug/firefox-26.0a1.en-US.win32.tests.zip",
-        "toolsdir": "e:/builds/moz2_slave/try-w32-d-00000000000000000000/tools",
-        "who": "mixedpuppy@gmail.com"
-      },
-      "reason": "scheduler",
-      "request_ids": [
-        28007724
-      ],
-      "requesttime": 1376505622,
-      "result": 0,
-      "slave_id": 2110,
-      "starttime": 1376505625
-    },
-    {
-      "builder_id": 143599,
-      "buildnumber": 2392,
-      "endtime": 1376513060,
-      "id": 26554073,
-      "master_id": 78,
-      "properties": {
-        "appName": "Firefox",
-        "appVersion": "26.0a1",
-        "basedir": "e:/builds/moz2_slave/try-w32-0000000000000000000000",
-        "branch": "test_treeherder",
-        "builddir": "try-w32-0000000000000000000000",
-        "buildername": "WINNT 5.2 try build",
-        "buildid": "20130814114051",
-        "buildnumber": 2392,
-        "builduid": "39570642c2fe4206bc6ac9ecdb37d740",
-        "comments": "try: -b do -p linux64,macosx64,win32 -u xpcshell,mochitests -t none",
-        "filepath": null,
-        "forced_clobber": false,
-        "got_revision": "2be7ca1d1d99",
-        "hashType": "sha512",
-        "installerFilename": "firefox-26.0a1.en-US.win32.installer.exe",
-        "installerHash": "2012282044dd5cc152c153c766ab86a23183c76dd372dc4cd98532e7e09542be069221255227f184f94e633f12af932fa510d67f2972cb5038989e613930b0bb",
-        "installerSize": "23616072",
-        "jsshellUrl": "http://ftp.mozilla.org/pub/mozilla.org/firefox/try-builds/mixedpuppy@gmail.com-2be7ca1d1d99/try-win32/jsshell-win32.zip",
-        "log_url": "http://ftp.mozilla.org/pub/mozilla.org/firefox/try-builds/mixedpuppy@gmail.com-2be7ca1d1d99/try-win32/try-win32-bm56-try1-build2392.txt.gz",
-        "master": "http://buildbot-master56.srv.releng.usw2.mozilla.com:8101/",
-        "packageFilename": "firefox-26.0a1.en-US.win32.zip",
-        "packageHash": "c945250cfc9ab7ee5a7ae66302451a50aa14a49c0f6b28e8fcb0412d786d63cd620427b7782655327ad31e1d0087e1e722f8723c9d57a3c96930a7beb00b05b6",
-        "packageSize": "30475556",
-        "packageUrl": "http://ftp.mozilla.org/pub/mozilla.org/firefox/try-builds/mixedpuppy@gmail.com-2be7ca1d1d99/try-win32/firefox-26.0a1.en-US.win32.zip",
-        "periodic_clobber": false,
-        "platform": "win32",
-        "product": "firefox",
-        "project": "",
-        "purge_actual": "152.40GB",
-        "purge_target": "12GB",
-        "purged_clobber": false,
-        "repository": "",
-        "request_ids": [
-          28007729
-        ],
-        "request_times": {
-          "28007729": 1376505622
-        },
-        "revision": "2be7ca1d1d9985374a8dedc8b0e1d5ee3d2cea00",
-        "scheduler": "try-firefox",
-        "slavebuilddir": "try-w32-0000000000000000000000",
-        "slavename": "w64-ix-slave62",
-        "stage_platform": "win32",
-        "symbolsUrl": "http://ftp.mozilla.org/pub/mozilla.org/firefox/try-builds/mixedpuppy@gmail.com-2be7ca1d1d99/try-win32/firefox-26.0a1.en-US.win32.crashreporter-symbols.zip",
-        "testsUrl": "http://ftp.mozilla.org/pub/mozilla.org/firefox/try-builds/mixedpuppy@gmail.com-2be7ca1d1d99/try-win32/firefox-26.0a1.en-US.win32.tests.zip",
-        "toolsdir": "e:/builds/moz2_slave/try-w32-0000000000000000000000/tools",
-        "who": "mixedpuppy@gmail.com"
-      },
-      "reason": "scheduler",
-      "request_ids": [
-        28007729
-      ],
-      "requesttime": 1376505622,
-      "result": 0,
-      "slave_id": 2138,
-      "starttime": 1376505626
-    },
-    {
-      "builder_id": 117700,
-      "buildnumber": 1308,
-      "endtime": 1376512239,
-      "id": 26553691,
-      "master_id": 78,
-      "properties": {
-        "appName": "Firefox",
-        "appVersion": "26.0a1",
-        "basedir": "/builds/slave/try-l64-d-00000000000000000000",
-        "branch": "test_treeherder",
-        "builddir": "try-l64-d-00000000000000000000",
-        "buildername": "Linux x86-64 try leak test build",
-        "buildid": "20130814114726",
-        "buildnumber": 1308,
-        "builduid": "39570642c2fe4206bc6ac9ecdb37d740",
-        "comments": "try: -b do -p linux64,macosx64,win32 -u xpcshell,mochitests -t none",
-        "filepath": null,
-        "forced_clobber": false,
-        "got_revision": "2be7ca1d1d99",
-        "hashType": "sha512",
-        "jsshellUrl": "http://ftp.mozilla.org/pub/mozilla.org/firefox/try-builds/mixedpuppy@gmail.com-2be7ca1d1d99/try-linux64-debug/jsshell-linux-x86_64.zip",
-        "log_url": "http://ftp.mozilla.org/pub/mozilla.org/firefox/try-builds/mixedpuppy@gmail.com-2be7ca1d1d99/try-linux64-debug/try-linux64-debug-bm56-try1-build1308.txt.gz",
-        "master": "http://buildbot-master56.srv.releng.usw2.mozilla.com:8101/",
-        "packageFilename": "firefox-26.0a1.en-US.linux-x86_64.tar.bz2",
-        "packageHash": "5bdfc05aaa33aa04d6ab9dcbd6c758fab246243f50ccec54690d84835ad369c8a66bb40e667a583ef96baaf9424d7ccd22d34d020fb04548e5bfddd1ddc80c9d",
-        "packageSize": "33466717",
-        "packageUrl": "http://ftp.mozilla.org/pub/mozilla.org/firefox/try-builds/mixedpuppy@gmail.com-2be7ca1d1d99/try-linux64-debug/firefox-26.0a1.en-US.linux-x86_64.tar.bz2",
-        "periodic_clobber": true,
-        "platform": "linux64-debug",
-        "product": "firefox",
-        "project": "",
-        "purge_actual": "40.38GB",
-        "purge_target": "14GB",
-        "purged_clobber": false,
-        "repository": "",
-        "request_ids": [
-          28007727
-        ],
-        "request_times": {
-          "28007727": 1376505622
-        },
-        "revision": "2be7ca1d1d9985374a8dedc8b0e1d5ee3d2cea00",
-        "scheduler": "try-firefox",
-        "slavebuilddir": "try-l64-d-00000000000000000000",
-        "slavename": "bld-linux64-ix-053",
-        "stage_platform": "linux64-debug",
-        "symbolsUrl": "http://ftp.mozilla.org/pub/mozilla.org/firefox/try-builds/mixedpuppy@gmail.com-2be7ca1d1d99/try-linux64-debug/firefox-26.0a1.en-US.linux-x86_64.crashreporter-symbols.zip",
-        "testsUrl": "http://ftp.mozilla.org/pub/mozilla.org/firefox/try-builds/mixedpuppy@gmail.com-2be7ca1d1d99/try-linux64-debug/firefox-26.0a1.en-US.linux-x86_64.tests.zip",
-        "toolsdir": "/builds/slave/try-l64-d-00000000000000000000/tools",
-        "who": "mixedpuppy@gmail.com"
-      },
-      "reason": "scheduler",
-      "request_ids": [
-        28007727
-      ],
-      "requesttime": 1376505622,
-      "result": 0,
-      "slave_id": 3954,
-      "starttime": 1376505626
-    },
-    {
-      "builder_id": 120555,
-      "buildnumber": 1885,
-      "endtime": 1376511597,
-      "id": 26553394,
-      "master_id": 93,
-      "properties": {
-        "basedir": "/builds/slave/talos-slave/test",
-        "branch": "test_treeherder",
-        "build_url": "http://ftp.mozilla.org/pub/mozilla.org/firefox/tinderbox-builds/mozilla-inbound-macosx64-debug/1376502081/firefox-26.0a1.en-US.mac64.dmg",
-        "buildername": "Rev5 MacOSX Mountain Lion 10.8 mozilla-inbound debug test mochitest-browser-chrome",
-        "buildid": "20130814104121",
-        "buildnumber": 1885,
-        "builduid": "56d1370353394fc487c082b6c96af781",
-        "log_url": "http://ftp.mozilla.org/pub/mozilla.org/firefox/tinderbox-builds/mozilla-inbound-macosx64-debug/1376502081/mozilla-inbound_mountainlion-debug_test-mochitest-browser-chrome-bm80-tests1-macosx-build1885.txt.gz",
-        "master": "http://buildbot-master80.srv.releng.usw2.mozilla.com:8201/",
-        "pgo_build": "False",
-        "platform": "macosx64",
-        "product": "firefox",
-        "project": "",
-        "repo_path": "integration/mozilla-inbound",
-        "repository": "",
-        "request_ids": [
-          28007464,
-          28007873
-        ],
-        "request_times": {
-          "28007464": 1376505242,
-          "28007873": 1376505721
-        },
-        "revision": "9df04b16a655",
-        "scheduler": "tests-mozilla-inbound-mountainlion-debug-unittest",
-        "script_repo_revision": "2f85052176f2",
-        "slavebuilddir": "test",
-        "slavename": "talos-mtnlion-r5-077",
-        "stage_platform": "macosx64",
-        "symbols_url": "http://ftp.mozilla.org/pub/mozilla.org/firefox/tinderbox-builds/mozilla-inbound-macosx64-debug/1376502081/firefox-26.0a1.en-US.mac64.crashreporter-symbols.zip"
-      },
-      "reason": "scheduler",
-      "request_ids": [
-        28007464,
-        28007873
-      ],
-      "requesttime": 1376505242,
-      "result": 0,
-      "slave_id": 2431,
-      "starttime": 1376505765
-    },
-    {
-      "builder_id": 119896,
-      "buildnumber": 1101,
-      "endtime": 1376513297,
-      "id": 26554204,
-      "master_id": 82,
-      "properties": {
-        "basedir": "/Users/cltbld/talos-slave/test",
-        "branch": "test_treeherder",
-        "build_url": "http://ftp.mozilla.org/pub/mozilla.org/firefox/tinderbox-builds/mozilla-inbound-macosx64-debug/1376502081/firefox-26.0a1.en-US.mac64.dmg",
-        "buildername": "Rev4 MacOSX Snow Leopard 10.6 mozilla-inbound debug test mochitest-browser-chrome",
-        "buildid": "20130814104121",
-        "buildnumber": 1101,
-        "builduid": "56d1370353394fc487c082b6c96af781",
-        "log_url": "http://ftp.mozilla.org/pub/mozilla.org/firefox/tinderbox-builds/mozilla-inbound-macosx64-debug/1376502081/mozilla-inbound_snowleopard-debug_test-mochitest-browser-chrome-bm75-tests1-macosx-build1101.txt.gz",
-        "master": "http://buildbot-master75.srv.releng.use1.mozilla.com:8201/",
-        "pgo_build": "False",
-        "platform": "macosx64",
-        "product": "firefox",
-        "project": "",
-        "repo_path": "integration/mozilla-inbound",
-        "repository": "",
-        "request_ids": [
-          28007478,
-          28007795
-        ],
-        "request_times": {
-          "28007478": 1376505247,
-          "28007795": 1376505667
-        },
-        "revision": "9df04b16a655",
-        "scheduler": "tests-mozilla-inbound-snowleopard-debug-unittest",
-        "script_repo_revision": "2f85052176f2",
-        "slavebuilddir": "test",
-        "slavename": "talos-r4-snow-047",
-        "stage_platform": "macosx64",
-        "symbols_url": "http://ftp.mozilla.org/pub/mozilla.org/firefox/tinderbox-builds/mozilla-inbound-macosx64-debug/1376502081/firefox-26.0a1.en-US.mac64.crashreporter-symbols.zip"
-      },
-      "reason": "scheduler",
-      "request_ids": [
-        28007478,
-        28007795
-      ],
-      "requesttime": 1376505247,
-      "result": 0,
-      "slave_id": 1404,
-      "starttime": 1376505819
-    },
-    {
-      "builder_id": 140996,
-      "buildnumber": 770,
-      "endtime": 1376511687,
-      "id": 26553427,
-      "master_id": 79,
-      "properties": {
-        "basedir": "/home/cltbld/talos-slave/test",
-        "branch": "test_treeherder",
-        "build_url": "http://ftp.mozilla.org/pub/mozilla.org/firefox/tinderbox-builds/mozilla-inbound-linux64-debug/1376501362/firefox-26.0a1.en-US.linux-x86_64.tar.bz2",
-        "buildername": "Rev3 Fedora 12x64 mozilla-inbound debug test mochitest-browser-chrome",
-        "buildid": "20130814102922",
-        "buildnumber": 770,
-        "builduid": "426fd52b759f42dcac6099a2e6adf392",
-        "log_url": "http://ftp.mozilla.org/pub/mozilla.org/firefox/tinderbox-builds/mozilla-inbound-linux64-debug/1376501362/mozilla-inbound_fedora64-debug_test-mochitest-browser-chrome-bm68-tests1-linux-build770.txt.gz",
-        "master": "http://buildbot-master68.srv.releng.usw2.mozilla.com:8201/",
-        "pgo_build": "False",
-        "platform": "linux64",
-        "product": "firefox",
-        "project": "",
-        "repo_path": "integration/mozilla-inbound",
-        "repository": "",
-        "request_ids": [
-          28008054
-        ],
-        "request_times": {
-          "28008054": 1376505841
-        },
-        "revision": "503e2b1df1da",
-        "scheduler": "tests-mozilla-inbound-fedora64-debug-unittest",
-        "script_repo_revision": "2f85052176f2",
-        "slavebuilddir": "test",
-        "slavename": "talos-r3-fed64-057",
-        "stage_platform": "linux64",
-        "symbols_url": "http://ftp.mozilla.org/pub/mozilla.org/firefox/tinderbox-builds/mozilla-inbound-linux64-debug/1376501362/firefox-26.0a1.en-US.linux-x86_64.crashreporter-symbols.zip"
-      },
-      "reason": "scheduler",
-      "request_ids": [
-        28008054
-      ],
-      "requesttime": 1376505841,
-      "result": 0,
-      "slave_id": 1295,
-      "starttime": 1376505845
-    },
-    {
-      "builder_id": 120591,
-      "buildnumber": 1024,
-      "endtime": 1376513634,
-      "id": 26554399,
-      "master_id": 90,
-      "properties": {
-        "basedir": "/builds/slave/talos-slave/test",
-        "branch": "test_treeherder",
-        "build_url": "http://ftp.mozilla.org/pub/mozilla.org/firefox/tinderbox-builds/mozilla-inbound-macosx64-debug/1376502081/firefox-26.0a1.en-US.mac64.dmg",
-        "buildername": "Rev4 MacOSX Lion 10.7 mozilla-inbound debug test mochitest-browser-chrome",
-        "buildid": "20130814104121",
-        "buildnumber": 1024,
-        "builduid": "56d1370353394fc487c082b6c96af781",
-        "log_url": "http://ftp.mozilla.org/pub/mozilla.org/firefox/tinderbox-builds/mozilla-inbound-macosx64-debug/1376502081/mozilla-inbound_lion-debug_test-mochitest-browser-chrome-bm79-tests1-macosx-build1024.txt.gz",
-        "master": "http://buildbot-master79.srv.releng.usw2.mozilla.com:8201/",
-        "pgo_build": "False",
-        "platform": "macosx64",
-        "product": "firefox",
-        "project": "",
-        "repo_path": "integration/mozilla-inbound",
-        "repository": "",
-        "request_ids": [
-          28007491,
-          28007808
-        ],
-        "request_times": {
-          "28007491": 1376505248,
-          "28007808": 1376505667
-        },
-        "revision": "9df04b16a655",
-        "scheduler": "tests-mozilla-inbound-lion-debug-unittest",
-        "script_repo_revision": "2f85052176f2",
-        "slavebuilddir": "test",
-        "slavename": "talos-r4-lion-023",
-        "stage_platform": "macosx64",
-        "symbols_url": "http://ftp.mozilla.org/pub/mozilla.org/firefox/tinderbox-builds/mozilla-inbound-macosx64-debug/1376502081/firefox-26.0a1.en-US.mac64.crashreporter-symbols.zip"
-      },
-      "reason": "scheduler",
-      "request_ids": [
-        28007491,
-        28007808
-      ],
-      "requesttime": 1376505248,
-      "result": 0,
-      "slave_id": 1606,
-      "starttime": 1376505928
-    },
-    {
-      "builder_id": 172471,
-      "buildnumber": 279,
-      "endtime": 1376512687,
-      "id": 26553918,
-      "master_id": 99,
-      "properties": {
-        "basedir": "/builds/slave/b2g_m-in_emulator_d_dep-000000",
-        "branch": "test_treeherder",
-        "buildername": "b2g_mozilla-inbound_emulator_debug_dep",
-        "buildid": "20130814101422",
-        "buildnumber": 279,
-        "builduid": "a0cfc29052ce4714a0ffb0463e73f9ce",
-        "compare_locales_revision": "1fc4e9bc8287",
-        "gaia_revision": "79700f6708d0b2b5ec6be0d07b63616dceb193ea",
-        "gecko_revision": "1fa1f28c6edf",
-        "log_url": "http://ftp.mozilla.org/pub/mozilla.org/b2g/tinderbox-builds/mozilla-inbound-emulator_debug/1376500462/b2g_mozilla-inbound_emulator_debug_dep-bm62-build1-build279.txt.gz",
-        "master": "http://buildbot-master62.srv.releng.use1.mozilla.com:8001/",
-        "platform": "emulator_debug",
-        "product": "b2g",
-        "project": "",
-        "repo_path": "integration/mozilla-inbound",
-        "repository": "",
-        "request_ids": [
-          28008111
-        ],
-        "request_times": {
-          "28008111": 1376505962
-        },
-        "revision": "1fa1f28c6edf",
-        "scheduler": "b2g_mozilla-inbound-b2g",
-        "script_repo_revision": "2f85052176f2",
-        "slavename": "bld-centos6-hp-008"
-      },
-      "reason": "Self-serve: Rebuilt by rvandermeulen@mozilla.com",
-      "request_ids": [
-        28008111
-      ],
-      "requesttime": 1376505962,
-      "result": 0,
-      "slave_id": 1947,
-      "starttime": 1376505964
-    },
-    {
-      "builder_id": 124166,
-      "buildnumber": 212,
-      "endtime": 1376511391,
-      "id": 26553289,
-      "master_id": 93,
-      "properties": {
-        "basedir": "/builds/slave/talos-slave/test",
-        "branch": "test_treeherder",
-        "build_url": "http://ftp.mozilla.org/pub/mozilla.org/firefox/tinderbox-builds/mozilla-aurora-macosx64-debug/1376502381/firefox-25.0a2.en-US.mac64.dmg",
-        "buildername": "Rev5 MacOSX Mountain Lion 10.8 mozilla-aurora debug test mochitest-browser-chrome",
-        "buildid": "20130814104621",
-        "buildnumber": 212,
-        "builduid": "f377570e7b214e2185ad6a995fde8820",
-        "log_url": "http://ftp.mozilla.org/pub/mozilla.org/firefox/tinderbox-builds/mozilla-aurora-macosx64-debug/1376502381/mozilla-aurora_mountainlion-debug_test-mochitest-browser-chrome-bm80-tests1-macosx-build212.txt.gz",
-        "master": "http://buildbot-master80.srv.releng.usw2.mozilla.com:8201/",
-        "pgo_build": "False",
-        "platform": "macosx64",
-        "product": "firefox",
-        "project": "",
-        "repo_path": "releases/mozilla-aurora",
-        "repository": "",
-        "request_ids": [
-          28008294
-        ],
-        "request_times": {
-          "28008294": 1376506086
-        },
-        "revision": "932868752f82",
-        "scheduler": "tests-mozilla-aurora-mountainlion-debug-unittest",
-        "script_repo_revision": "2f85052176f2",
-        "slavebuilddir": "test",
-        "slavename": "talos-mtnlion-r5-026",
-        "stage_platform": "macosx64",
-        "symbols_url": "http://ftp.mozilla.org/pub/mozilla.org/firefox/tinderbox-builds/mozilla-aurora-macosx64-debug/1376502381/firefox-25.0a2.en-US.mac64.crashreporter-symbols.zip"
-      },
-      "reason": "scheduler",
-      "request_ids": [
-        28008294
-      ],
-      "requesttime": 1376506086,
-      "result": 0,
-      "slave_id": 2339,
-      "starttime": 1376506151
-    },
-    {
-      "builder_id": 127008,
-      "buildnumber": 130,
-      "endtime": 1376513451,
-      "id": 26554302,
-      "master_id": 82,
-      "properties": {
-        "basedir": "/Users/cltbld/talos-slave/test",
-        "branch": "test_treeherder",
-        "build_url": "http://ftp.mozilla.org/pub/mozilla.org/firefox/tinderbox-builds/mozilla-aurora-macosx64-debug/1376502381/firefox-25.0a2.en-US.mac64.dmg",
-        "buildername": "Rev4 MacOSX Lion 10.7 mozilla-aurora debug test mochitest-browser-chrome",
-        "buildid": "20130814104621",
-        "buildnumber": 130,
-        "builduid": "f377570e7b214e2185ad6a995fde8820",
-        "log_url": "http://ftp.mozilla.org/pub/mozilla.org/firefox/tinderbox-builds/mozilla-aurora-macosx64-debug/1376502381/mozilla-aurora_lion-debug_test-mochitest-browser-chrome-bm75-tests1-macosx-build130.txt.gz",
-        "master": "http://buildbot-master75.srv.releng.use1.mozilla.com:8201/",
-        "pgo_build": "False",
-        "platform": "macosx64",
-        "product": "firefox",
-        "project": "",
-        "repo_path": "releases/mozilla-aurora",
-        "repository": "",
-        "request_ids": [
-          28008405
-        ],
-        "request_times": {
-          "28008405": 1376506135
-        },
-        "revision": "932868752f82",
-        "scheduler": "tests-mozilla-aurora-lion-debug-unittest",
-        "script_repo_revision": "2f85052176f2",
-        "slavebuilddir": "test",
-        "slavename": "talos-r4-lion-087",
-        "stage_platform": "macosx64",
-        "symbols_url": "http://ftp.mozilla.org/pub/mozilla.org/firefox/tinderbox-builds/mozilla-aurora-macosx64-debug/1376502381/firefox-25.0a2.en-US.mac64.crashreporter-symbols.zip"
-      },
-      "reason": "scheduler",
-      "request_ids": [
-        28008405
-      ],
-      "requesttime": 1376506135,
-      "result": 0,
-      "slave_id": 4128,
-      "starttime": 1376506208
-    },
-    {
-      "builder_id": 120452,
-      "buildnumber": 63,
-      "endtime": 1376511408,
-      "id": 26553305,
-      "master_id": 99,
-      "properties": {
-        "appName": "Fennec",
-        "appVersion": "26.0a1",
-        "basedir": "/builds/slave/fx-team-and-a6-000000000000000",
-        "branch": "test_treeherder",
-        "builddir": "fx-team-and-a6-000000000000000",
-        "buildername": "Android Armv6 fx-team build",
-        "buildid": "20130814115224",
-        "buildnumber": 63,
-        "builduid": "52d2fd23d8da4748b92df17ed1dbde5c",
-        "comments": "Bug 689291 test-tabs.test window focus changes active tab | Test output exceeded timeout (60s). r=Mossop",
-        "filepath": null,
-        "forced_clobber": false,
-        "got_revision": "919119a88080",
-        "hashType": "sha512",
-        "jsshellUrl": "http://ftp.mozilla.org/pub/mozilla.org/mobile/tinderbox-builds/fx-team-android-armv6/1376506344/jsshell-android-arm-armv6.zip",
-        "log_url": "http://ftp.mozilla.org/pub/mozilla.org/mobile/tinderbox-builds/fx-team-android-armv6/1376506344/fx-team-android-armv6-bm62-build1-build63.txt.gz",
-        "master": "http://buildbot-master62.srv.releng.use1.mozilla.com:8001/",
-        "packageFilename": "fennec-26.0a1.en-US.android-arm-armv6.apk",
-        "packageHash": "476a9f842c5c3c8b701e3d05b29517eaab56f6053f422a27d4b25da431332c6c77c765677357b7edd80cb72ad57d4f26fb9ef97f61fede507bbb187c3ded632c",
-        "packageSize": "25862748",
-        "packageUrl": "http://ftp.mozilla.org/pub/mozilla.org/mobile/tinderbox-builds/fx-team-android-armv6/1376506344/fennec-26.0a1.en-US.android-arm-armv6.apk",
-        "periodic_clobber": false,
-        "platform": "android-armv6",
-        "product": "mobile",
-        "project": "",
-        "purge_actual": "101.09GB",
-        "purge_target": "12GB",
-        "purged_clobber": false,
-        "repository": "",
-        "request_ids": [
-          28008490
-        ],
-        "request_times": {
-          "28008490": 1376506344
-        },
-        "revision": "919119a88080a7642cdc6f3609766cb84b642e1a",
-        "robocopApkUrl": "http://ftp.mozilla.org/pub/mozilla.org/mobile/tinderbox-builds/fx-team-android-armv6/1376506344/robocop.apk",
-        "scheduler": "fx-team-mobile",
-        "slavebuilddir": "fx-team-and-a6-000000000000000",
-        "slavename": "bld-linux64-ec2-170",
-        "stage_platform": "android-armv6",
-        "symbolsUrl": "http://ftp.mozilla.org/pub/mozilla.org/mobile/tinderbox-builds/fx-team-android-armv6/1376506344/fennec-26.0a1.en-US.android-arm-armv6.crashreporter-symbols.zip",
-        "testsUrl": "http://ftp.mozilla.org/pub/mozilla.org/mobile/tinderbox-builds/fx-team-android-armv6/1376506344/fennec-26.0a1.en-US.android-arm-armv6.tests.zip",
-        "toolsdir": "/builds/slave/fx-team-and-a6-000000000000000/tools",
-        "unsignedApkUrl": "http://ftp.mozilla.org/pub/mozilla.org/mobile/tinderbox-builds/fx-team-android-armv6/1376506344/gecko-unsigned-unaligned.apk"
-      },
-      "reason": "scheduler",
-      "request_ids": [
-        28008490
-      ],
-      "requesttime": 1376506344,
-      "result": 0,
-      "slave_id": 4881,
-      "starttime": 1376506352
-    },
-    {
-      "builder_id": 121827,
-      "buildnumber": 79,
-      "endtime": 1376512318,
-      "id": 26553737,
-      "master_id": 99,
-      "properties": {
-        "appName": "B2G",
-        "appVersion": "26.0a1",
-        "basedir": "/builds/slave/fx-team-linux32_g-000000000000",
-        "branch": "test_treeherder",
-        "builddir": "fx-team-linux32_g-000000000000",
-        "buildername": "b2g_fx-team_linux32_gecko build",
-        "buildid": "20130814115223",
-        "buildnumber": 79,
-        "builduid": "5114acb64a934005a3dca5605cee6605",
-        "comments": "Bug 689291 test-tabs.test window focus changes active tab | Test output exceeded timeout (60s). r=Mossop",
-        "forced_clobber": false,
-        "gaiaRepoPath": "/integration/gaia-central",
-        "gaiaRevision": "47daabb93939a38f1f8efa7fdb1cd22ef8445600",
-        "got_revision": "919119a88080",
-        "hashType": "sha512",
-        "log_url": "http://ftp.mozilla.org/pub/mozilla.org/b2g/tinderbox-builds/fx-team-linux32_gecko/1376506343/fx-team-linux32_gecko-bm62-build1-build79.txt.gz",
-        "master": "http://buildbot-master62.srv.releng.use1.mozilla.com:8001/",
-        "packageUrl": "http://ftp.mozilla.org/pub/mozilla.org/b2g/tinderbox-builds/fx-team-linux32_gecko/1376506343/b2g-26.0a1.multi.linux-i686.tar.bz2",
-        "periodic_clobber": false,
-        "platform": "linux32_gecko",
-        "product": "b2g",
-        "project": "",
-        "purge_actual": "67.51GB",
-        "purge_target": "13GB",
-        "purged_clobber": false,
-        "repository": "",
-        "request_ids": [
-          28008485
-        ],
-        "request_times": {
-          "28008485": 1376506344
-        },
-        "revision": "919119a88080a7642cdc6f3609766cb84b642e1a",
-        "scheduler": "b2g_fx-team-b2g",
-        "slavebuilddir": "fx-team-linux32_g-000000000000",
-        "slavename": "bld-linux64-ec2-032",
-        "stage_platform": "linux32_gecko",
-        "symbolsUrl": "http://ftp.mozilla.org/pub/mozilla.org/b2g/tinderbox-builds/fx-team-linux32_gecko/1376506343/en-US/b2g-26.0a1.en-US.linux-i686.crashreporter-symbols.zip",
-        "testsUrl": "http://ftp.mozilla.org/pub/mozilla.org/b2g/tinderbox-builds/fx-team-linux32_gecko/1376506343/b2g-26.0a1.multi.linux-i686.tests.zip",
-        "toolsdir": "/builds/slave/fx-team-linux32_g-000000000000/tools"
-      },
-      "reason": "scheduler",
-      "request_ids": [
-        28008485
-      ],
-      "requesttime": 1376506344,
-      "result": 0,
-      "slave_id": 2263,
-      "starttime": 1376506353
-    },
-    {
-      "builder_id": 146705,
-      "buildnumber": 96,
-      "endtime": 1376514346,
-      "id": 26554832,
-      "master_id": 99,
-      "properties": {
-        "appName": "Firefox",
-        "appVersion": "26.0a1",
-        "basedir": "/builds/slave/fx-team-osx64-0000000000000000",
-        "branch": "test_treeherder",
-        "builddir": "fx-team-osx64-0000000000000000",
-        "buildername": "OS X 10.7 fx-team build",
-        "buildid": "20130814115223",
-        "buildnumber": 96,
-        "builduid": "a2a7488ff85242688aea1521436b5224",
-        "comments": "Bug 689291 test-tabs.test window focus changes active tab | Test output exceeded timeout (60s). r=Mossop",
-        "filepath": null,
-        "forced_clobber": false,
-        "got_revision": "919119a88080",
-        "hashType": "sha512",
-        "jsshellUrl": "http://ftp.mozilla.org/pub/mozilla.org/firefox/tinderbox-builds/fx-team-macosx64/1376506343/jsshell-mac.zip",
-        "log_url": "http://ftp.mozilla.org/pub/mozilla.org/firefox/tinderbox-builds/fx-team-macosx64/1376506343/fx-team-macosx64-bm62-build1-build96.txt.gz",
-        "master": "http://buildbot-master62.srv.releng.use1.mozilla.com:8001/",
-        "packageFilename": "firefox-26.0a1.en-US.mac.dmg",
-        "packageHash": "ce86f218d48c99276eeebd5476b420a73bdc74aa23be2e4ca5dbbff45afb024eaf21062fcbe897111be380cdf4e67281b8a193400f536f9fdb0117148ef3d68f",
-        "packageSize": "69054080",
-        "packageUrl": "http://ftp.mozilla.org/pub/mozilla.org/firefox/tinderbox-builds/fx-team-macosx64/1376506343/firefox-26.0a1.en-US.mac.dmg",
-        "periodic_clobber": false,
-        "platform": "macosx64",
-        "product": "firefox",
-        "project": "",
-        "purge_actual": "810.37GB",
-        "purge_target": "12GB",
-        "purged_clobber": false,
-        "repository": "",
-        "request_ids": [
-          28008479
-        ],
-        "request_times": {
-          "28008479": 1376506344
-        },
-        "revision": "919119a88080a7642cdc6f3609766cb84b642e1a",
-        "scheduler": "fx-team-firefox",
-        "slavebuilddir": "fx-team-osx64-0000000000000000",
-        "slavename": "bld-lion-r5-049",
-        "stage_platform": "macosx64",
-        "symbolsUrl": "http://ftp.mozilla.org/pub/mozilla.org/firefox/tinderbox-builds/fx-team-macosx64/1376506343/firefox-26.0a1.en-US.mac.crashreporter-symbols.zip",
-        "testsUrl": "http://ftp.mozilla.org/pub/mozilla.org/firefox/tinderbox-builds/fx-team-macosx64/1376506343/firefox-26.0a1.en-US.mac.tests.zip",
-        "toolsdir": "/builds/slave/fx-team-osx64-0000000000000000/tools"
-      },
-      "reason": "scheduler",
-      "request_ids": [
-        28008479
-      ],
-      "requesttime": 1376506344,
-      "result": 0,
-      "slave_id": 2040,
-      "starttime": 1376506353
-    },
-    {
-      "builder_id": 126453,
-      "buildnumber": 58,
-      "endtime": 1376512144,
-      "id": 26553659,
-      "master_id": 99,
-      "properties": {
-        "appName": "Fennec",
-        "appVersion": "26.0a1",
-        "basedir": "/builds/slave/fx-team-and-000000000000000000",
-        "branch": "test_treeherder",
-        "builddir": "fx-team-and-000000000000000000",
-        "buildername": "Android fx-team build",
-        "buildid": "20130814115224",
-        "buildnumber": 58,
-        "builduid": "52d2fd23d8da4748b92df17ed1dbde5c",
-        "comments": "Bug 689291 test-tabs.test window focus changes active tab | Test output exceeded timeout (60s). r=Mossop",
-        "filepath": null,
-        "forced_clobber": false,
-        "got_revision": "919119a88080",
-        "hashType": "sha512",
-        "jsshellUrl": "http://ftp.mozilla.org/pub/mozilla.org/mobile/tinderbox-builds/fx-team-android/1376506344/jsshell-android-arm.zip",
-        "log_url": "http://ftp.mozilla.org/pub/mozilla.org/mobile/tinderbox-builds/fx-team-android/1376506344/fx-team-android-bm62-build1-build58.txt.gz",
-        "master": "http://buildbot-master62.srv.releng.use1.mozilla.com:8001/",
-        "packageFilename": "fennec-26.0a1.en-US.android-arm.apk",
-        "packageHash": "57a9a0ce91aadbb1934dc917c16738557af135c1a44b767d1a87273b2ca5669d73e06e594a6b338e1f65b9a3aa601898cc7b3d2b9600129e69abb9fc32a916e1",
-        "packageSize": "30007620",
-        "packageUrl": "http://ftp.mozilla.org/pub/mozilla.org/mobile/tinderbox-builds/fx-team-android/1376506344/fennec-26.0a1.en-US.android-arm.apk",
-        "periodic_clobber": false,
-        "platform": "android",
-        "product": "mobile",
-        "project": "",
-        "purge_actual": "154.40GB",
-        "purge_target": "12GB",
-        "purged_clobber": true,
-        "repository": "",
-        "request_ids": [
-          28008489
-        ],
-        "request_times": {
-          "28008489": 1376506344
-        },
-        "revision": "919119a88080a7642cdc6f3609766cb84b642e1a",
-        "robocopApkUrl": "http://ftp.mozilla.org/pub/mozilla.org/mobile/tinderbox-builds/fx-team-android/1376506344/robocop.apk",
-        "scheduler": "fx-team-mobile",
-        "slavebuilddir": "fx-team-and-000000000000000000",
-        "slavename": "bld-linux64-ec2-062",
-        "stage_platform": "android",
-        "symbolsUrl": "http://ftp.mozilla.org/pub/mozilla.org/mobile/tinderbox-builds/fx-team-android/1376506344/fennec-26.0a1.en-US.android-arm.crashreporter-symbols.zip",
-        "testsUrl": "http://ftp.mozilla.org/pub/mozilla.org/mobile/tinderbox-builds/fx-team-android/1376506344/fennec-26.0a1.en-US.android-arm.tests.zip",
-        "toolsdir": "/builds/slave/fx-team-and-000000000000000000/tools",
-        "unsignedApkUrl": "http://ftp.mozilla.org/pub/mozilla.org/mobile/tinderbox-builds/fx-team-android/1376506344/gecko-unsigned-unaligned.apk"
-      },
-      "reason": "scheduler",
-      "request_ids": [
-        28008489
-      ],
-      "requesttime": 1376506344,
-      "result": 0,
-      "slave_id": 2500,
-      "starttime": 1376506354
-    },
-    {
-      "builder_id": 155220,
-      "buildnumber": 48,
-      "endtime": 1376513156,
-      "id": 26554110,
-      "master_id": 99,
-      "properties": {
-        "basedir": "/builds/slave/b2g_fx-team_emulator_dep-00000",
-        "branch": "test_treeherder",
-        "buildername": "b2g_fx-team_emulator_dep",
-        "buildid": "20130814115223",
-        "buildnumber": 48,
-        "builduid": "5114acb64a934005a3dca5605cee6605",
-        "compare_locales_revision": "1fc4e9bc8287",
-        "gaia_revision": "b0f83dd98f0068168e793e9d91b1be879812f5a1",
-        "gecko_revision": "919119a88080",
-        "log_url": "http://ftp.mozilla.org/pub/mozilla.org/b2g/tinderbox-builds/fx-team-emulator/1376506343/b2g_fx-team_emulator_dep-bm62-build1-build48.txt.gz",
-        "master": "http://buildbot-master62.srv.releng.use1.mozilla.com:8001/",
-        "platform": "emulator",
-        "product": "b2g",
-        "project": "",
-        "repo_path": "integration/fx-team",
-        "repository": "",
-        "request_ids": [
-          28008483
-        ],
-        "request_times": {
-          "28008483": 1376506344
-        },
-        "revision": "919119a88080",
-        "scheduler": "b2g_fx-team-b2g",
-        "script_repo_revision": "2f85052176f2",
-        "slavename": "bld-linux64-ec2-178"
-      },
-      "reason": "scheduler",
-      "request_ids": [
-        28008483
-      ],
-      "requesttime": 1376506344,
-      "result": 0,
-      "slave_id": 4877,
-      "starttime": 1376506354
-    },
-    {
-      "builder_id": 131157,
-      "buildnumber": 97,
-      "endtime": 1376514612,
-      "id": 26555000,
-      "master_id": 99,
-      "properties": {
-        "appName": "Firefox",
-        "appVersion": "26.0a1",
-        "basedir": "/builds/slave/fx-team-l64-d-0000000000000000",
-        "branch": "test_treeherder",
-        "builddir": "fx-team-l64-d-0000000000000000",
-        "buildername": "Linux x86-64 fx-team leak test build",
-        "buildid": "20130814115223",
-        "buildnumber": 97,
-        "builduid": "a2a7488ff85242688aea1521436b5224",
-        "comments": "Bug 689291 test-tabs.test window focus changes active tab | Test output exceeded timeout (60s). r=Mossop",
-        "filepath": null,
-        "forced_clobber": false,
-        "got_revision": "919119a88080",
-        "hashType": "sha512",
-        "jsshellUrl": "http://ftp.mozilla.org/pub/mozilla.org/firefox/tinderbox-builds/fx-team-linux64-debug/1376506343/jsshell-linux-x86_64.zip",
-        "log_url": "http://ftp.mozilla.org/pub/mozilla.org/firefox/tinderbox-builds/fx-team-linux64-debug/1376506343/fx-team-linux64-debug-bm62-build1-build97.txt.gz",
-        "master": "http://buildbot-master62.srv.releng.use1.mozilla.com:8001/",
-        "packageFilename": "firefox-26.0a1.en-US.linux-x86_64.tar.bz2",
-        "packageHash": "ea974946b27eb32c34a2a344bd88abaf5a36241c8f8d999234a4e634be4c854b96d9190ffa7705136a695c835eeb7eff5bdf5f39da892846c25e5a71cb0503fa",
-        "packageSize": "36617227",
-        "packageUrl": "http://ftp.mozilla.org/pub/mozilla.org/firefox/tinderbox-builds/fx-team-linux64-debug/1376506343/firefox-26.0a1.en-US.linux-x86_64.tar.bz2",
-        "periodic_clobber": false,
-        "platform": "linux64-debug",
-        "product": "firefox",
-        "project": "",
-        "purge_actual": "95.85GB",
-        "purge_target": "14GB",
-        "purged_clobber": false,
-        "repository": "",
-        "request_ids": [
-          28008478
-        ],
-        "request_times": {
-          "28008478": 1376506344
-        },
-        "revision": "919119a88080a7642cdc6f3609766cb84b642e1a",
-        "scheduler": "fx-team-firefox",
-        "slavebuilddir": "fx-team-l64-d-0000000000000000",
-        "slavename": "bld-linux64-ec2-151",
-        "stage_platform": "linux64-debug",
-        "symbolsUrl": "http://ftp.mozilla.org/pub/mozilla.org/firefox/tinderbox-builds/fx-team-linux64-debug/1376506343/firefox-26.0a1.en-US.linux-x86_64.crashreporter-symbols.zip",
-        "testsUrl": "http://ftp.mozilla.org/pub/mozilla.org/firefox/tinderbox-builds/fx-team-linux64-debug/1376506343/firefox-26.0a1.en-US.linux-x86_64.tests.zip",
-        "toolsdir": "/builds/slave/fx-team-l64-d-0000000000000000/tools"
-      },
-      "reason": "scheduler",
-      "request_ids": [
-        28008478
-      ],
-      "requesttime": 1376506344,
-      "result": 0,
-      "slave_id": 4848,
-      "starttime": 1376506354
-    },
-    {
-      "builder_id": 146086,
-      "buildnumber": 116,
-      "endtime": 1376514559,
-      "id": 26554965,
-      "master_id": 99,
-      "properties": {
-        "appName": "Firefox",
-        "appVersion": "26.0a1",
-        "basedir": "e:/builds/moz2_slave/fx-team-w32-d-0000000000000000",
-        "branch": "test_treeherder",
-        "builddir": "fx-team-w32-d-0000000000000000",
-        "buildername": "WINNT 5.2 fx-team leak test build",
-        "buildid": "20130814115223",
-        "buildnumber": 116,
-        "builduid": "a2a7488ff85242688aea1521436b5224",
-        "comments": "Bug 689291 test-tabs.test window focus changes active tab | Test output exceeded timeout (60s). r=Mossop",
-        "filepath": null,
-        "forced_clobber": false,
-        "got_revision": "919119a88080",
-        "hashType": "sha512",
-        "installerFilename": "firefox-26.0a1.en-US.win32.installer.exe",
-        "installerHash": "680e619f9649243c55cd3caa75f7ae297b9a775e4f7b2400d785c661596a9ce08785cd9f24ee43e666161eed7cbbfab768d5fe3bd0d07cdbdfb650882998ad34",
-        "installerSize": "29614736",
-        "jsshellUrl": "http://ftp.mozilla.org/pub/mozilla.org/firefox/tinderbox-builds/fx-team-win32-debug/1376506343/jsshell-win32.zip",
-        "log_url": "http://ftp.mozilla.org/pub/mozilla.org/firefox/tinderbox-builds/fx-team-win32-debug/1376506343/fx-team-win32-debug-bm62-build1-build116.txt.gz",
-        "master": "http://buildbot-master62.srv.releng.use1.mozilla.com:8001/",
-        "packageFilename": "firefox-26.0a1.en-US.win32.zip",
-        "packageHash": "3743feb0486d53dc134305ce142ad90e989d7a7dffb9f0269886b2e3faf4eeb628dbd9de700df6d3aa6b4a925bac773db9d41d74de641319a24869f3e5a817ca",
-        "packageSize": "41339183",
-        "packageUrl": "http://ftp.mozilla.org/pub/mozilla.org/firefox/tinderbox-builds/fx-team-win32-debug/1376506343/firefox-26.0a1.en-US.win32.zip",
-        "periodic_clobber": false,
-        "platform": "win32-debug",
-        "product": "firefox",
-        "project": "",
-        "purge_actual": "103.01GB",
-        "purge_target": "9GB",
-        "purged_clobber": false,
-        "repository": "",
-        "request_ids": [
-          28008482
-        ],
-        "request_times": {
-          "28008482": 1376506344
-        },
-        "revision": "919119a88080a7642cdc6f3609766cb84b642e1a",
-        "scheduler": "fx-team-firefox",
-        "slavebuilddir": "fx-team-w32-d-0000000000000000",
-        "slavename": "w64-ix-slave109",
-        "stage_platform": "win32-debug",
-        "symbolsUrl": "http://ftp.mozilla.org/pub/mozilla.org/firefox/tinderbox-builds/fx-team-win32-debug/1376506343/firefox-26.0a1.en-US.win32.crashreporter-symbols.zip",
-        "testsUrl": "http://ftp.mozilla.org/pub/mozilla.org/firefox/tinderbox-builds/fx-team-win32-debug/1376506343/firefox-26.0a1.en-US.win32.tests.zip",
-        "toolsdir": "e:/builds/moz2_slave/fx-team-w32-d-0000000000000000/tools"
-      },
-      "reason": "scheduler",
-      "request_ids": [
-        28008482
-      ],
-      "requesttime": 1376506344,
-      "result": 0,
-      "slave_id": 2319,
-      "starttime": 1376506356
-    },
-    {
-      "builder_id": 117347,
-      "buildnumber": 73,
-      "endtime": 1376511344,
-      "id": 26553263,
-      "master_id": 76,
-      "properties": {
-        "appName": "Fennec",
-        "appVersion": "26.0a1",
-        "basedir": "/builds/slave/fx-team-and-x86-00000000000000",
-        "branch": "test_treeherder",
-        "builddir": "fx-team-and-x86-00000000000000",
-        "buildername": "Android X86 fx-team build",
-        "buildid": "20130814115224",
-        "buildnumber": 73,
-        "builduid": "52d2fd23d8da4748b92df17ed1dbde5c",
-        "comments": "Bug 689291 test-tabs.test window focus changes active tab | Test output exceeded timeout (60s). r=Mossop",
-        "filepath": null,
-        "forced_clobber": false,
-        "got_revision": "919119a88080",
-        "hashType": "sha512",
-        "jsshellUrl": "http://ftp.mozilla.org/pub/mozilla.org/mobile/tinderbox-builds/fx-team-android-x86/1376506344/jsshell-android-i386.zip",
-        "log_url": "http://ftp.mozilla.org/pub/mozilla.org/mobile/tinderbox-builds/fx-team-android-x86/1376506344/fx-team-android-x86-bm57-build1-build73.txt.gz",
-        "master": "http://buildbot-master57.srv.releng.use1.mozilla.com:8001/",
-        "packageFilename": "fennec-26.0a1.en-US.android-i386.apk",
-        "packageHash": "a31c78403f67751d4a3e68585a5f12fbdc493fa01d2079776433b3900d50ed76c2e4c5cad3400bdbd33f3c5dbc838742286300bef25603e5b0f27a36e38f3da2",
-        "packageSize": "32603364",
-        "packageUrl": "http://ftp.mozilla.org/pub/mozilla.org/mobile/tinderbox-builds/fx-team-android-x86/1376506344/fennec-26.0a1.en-US.android-i386.apk",
-        "periodic_clobber": false,
-        "platform": "android-x86",
-        "product": "mobile",
-        "project": "",
-        "purge_actual": "89.94GB",
-        "purge_target": "12GB",
-        "purged_clobber": false,
-        "repository": "",
-        "request_ids": [
-          28008492
-        ],
-        "request_times": {
-          "28008492": 1376506344
-        },
-        "revision": "919119a88080a7642cdc6f3609766cb84b642e1a",
-        "robocopApkUrl": "http://ftp.mozilla.org/pub/mozilla.org/mobile/tinderbox-builds/fx-team-android-x86/1376506344/robocop.apk",
-        "scheduler": "fx-team-mobile",
-        "slavebuilddir": "fx-team-and-x86-00000000000000",
-        "slavename": "bld-linux64-ec2-102",
-        "stage_platform": "android-x86",
-        "symbolsUrl": "http://ftp.mozilla.org/pub/mozilla.org/mobile/tinderbox-builds/fx-team-android-x86/1376506344/fennec-26.0a1.en-US.android-i386.crashreporter-symbols.zip",
-        "testsUrl": "http://ftp.mozilla.org/pub/mozilla.org/mobile/tinderbox-builds/fx-team-android-x86/1376506344/fennec-26.0a1.en-US.android-i386.tests.zip",
-        "toolsdir": "/builds/slave/fx-team-and-x86-00000000000000/tools",
-        "unsignedApkUrl": "http://ftp.mozilla.org/pub/mozilla.org/mobile/tinderbox-builds/fx-team-android-x86/1376506344/gecko-unsigned-unaligned.apk"
-      },
-      "reason": "scheduler",
-      "request_ids": [
-        28008492
-      ],
-      "requesttime": 1376506344,
-      "result": 0,
-      "slave_id": 4824,
-      "starttime": 1376506357
-    },
-    {
-      "builder_id": 117381,
-      "buildnumber": 103,
-      "endtime": 1376512196,
-      "id": 26553678,
-      "master_id": 76,
-      "properties": {
-        "appName": "Firefox",
-        "appVersion": "26.0a1",
-        "basedir": "/builds/slave/fx-team-lx-d-00000000000000000",
-        "branch": "test_treeherder",
-        "builddir": "fx-team-lx-d-00000000000000000",
-        "buildername": "Linux fx-team leak test build",
-        "buildid": "20130814115223",
-        "buildnumber": 103,
-        "builduid": "a2a7488ff85242688aea1521436b5224",
-        "comments": "Bug 689291 test-tabs.test window focus changes active tab | Test output exceeded timeout (60s). r=Mossop",
-        "filepath": null,
-        "forced_clobber": false,
-        "got_revision": "919119a88080",
-        "hashType": "sha512",
-        "jsshellUrl": "http://ftp.mozilla.org/pub/mozilla.org/firefox/tinderbox-builds/fx-team-linux-debug/1376506343/jsshell-linux-i686.zip",
-        "log_url": "http://ftp.mozilla.org/pub/mozilla.org/firefox/tinderbox-builds/fx-team-linux-debug/1376506343/fx-team-linux-debug-bm57-build1-build103.txt.gz",
-        "master": "http://buildbot-master57.srv.releng.use1.mozilla.com:8001/",
-        "packageFilename": "firefox-26.0a1.en-US.linux-i686.tar.bz2",
-        "packageHash": "de2782b34976b599fdcead5cf06e081222284b968bf2005be8807ec2f3ebae2af141d573468f76d131c6edfb8a308dec4992e2aa076b064fd66ea9494000384f",
-        "packageSize": "36528207",
-        "packageUrl": "http://ftp.mozilla.org/pub/mozilla.org/firefox/tinderbox-builds/fx-team-linux-debug/1376506343/firefox-26.0a1.en-US.linux-i686.tar.bz2",
-        "periodic_clobber": false,
-        "platform": "linux-debug",
-        "product": "firefox",
-        "project": "",
-        "purge_actual": "19.69GB",
-        "purge_target": "14GB",
-        "purged_clobber": true,
-        "repository": "",
-        "request_ids": [
-          28008475
-        ],
-        "request_times": {
-          "28008475": 1376506344
-        },
-        "revision": "919119a88080a7642cdc6f3609766cb84b642e1a",
-        "scheduler": "fx-team-firefox",
-        "slavebuilddir": "fx-team-lx-d-00000000000000000",
-        "slavename": "bld-linux64-ix-033",
-        "stage_platform": "linux-debug",
-        "symbolsUrl": "http://ftp.mozilla.org/pub/mozilla.org/firefox/tinderbox-builds/fx-team-linux-debug/1376506343/firefox-26.0a1.en-US.linux-i686.crashreporter-symbols.zip",
-        "testsUrl": "http://ftp.mozilla.org/pub/mozilla.org/firefox/tinderbox-builds/fx-team-linux-debug/1376506343/firefox-26.0a1.en-US.linux-i686.tests.zip",
-        "toolsdir": "/builds/slave/fx-team-lx-d-00000000000000000/tools"
-      },
-      "reason": "scheduler",
-      "request_ids": [
-        28008475
-      ],
-      "requesttime": 1376506344,
-      "result": 0,
-      "slave_id": 3801,
-      "starttime": 1376506357
-    },
-    {
-      "builder_id": 120426,
-      "buildnumber": 75,
-      "endtime": 1376512507,
-      "id": 26553826,
-      "master_id": 86,
-      "properties": {
-        "appName": "B2G",
-        "appVersion": "26.0a1",
-        "basedir": "/builds/slave/fx-team-l64_g-0000000000000000",
-        "branch": "test_treeherder",
-        "builddir": "fx-team-l64_g-0000000000000000",
-        "buildername": "b2g_fx-team_linux64_gecko build",
-        "buildid": "20130814115223",
-        "buildnumber": 75,
-        "builduid": "5114acb64a934005a3dca5605cee6605",
-        "comments": "Bug 689291 test-tabs.test window focus changes active tab | Test output exceeded timeout (60s). r=Mossop",
-        "forced_clobber": false,
-        "gaiaRepoPath": "/integration/gaia-central",
-        "gaiaRevision": "47daabb93939a38f1f8efa7fdb1cd22ef8445600",
-        "got_revision": "919119a88080",
-        "hashType": "sha512",
-        "log_url": "http://ftp.mozilla.org/pub/mozilla.org/b2g/tinderbox-builds/fx-team-linux64_gecko/1376506343/fx-team-linux64_gecko-bm63-build1-build75.txt.gz",
-        "master": "http://buildbot-master63.srv.releng.use1.mozilla.com:8001/",
-        "packageUrl": "http://ftp.mozilla.org/pub/mozilla.org/b2g/tinderbox-builds/fx-team-linux64_gecko/1376506343/b2g-26.0a1.multi.linux-x86_64.tar.bz2",
-        "periodic_clobber": false,
-        "platform": "linux64_gecko",
-        "product": "b2g",
-        "project": "",
-        "purge_actual": "17.30GB",
-        "purge_target": "13GB",
-        "purged_clobber": true,
-        "repository": "",
-        "request_ids": [
-          28008486
-        ],
-        "request_times": {
-          "28008486": 1376506344
-        },
-        "revision": "919119a88080a7642cdc6f3609766cb84b642e1a",
-        "scheduler": "b2g_fx-team-b2g",
-        "slavebuilddir": "fx-team-l64_g-0000000000000000",
-        "slavename": "bld-linux64-ec2-123",
-        "stage_platform": "linux64_gecko",
-        "symbolsUrl": "http://ftp.mozilla.org/pub/mozilla.org/b2g/tinderbox-builds/fx-team-linux64_gecko/1376506343/en-US/b2g-26.0a1.en-US.linux-x86_64.crashreporter-symbols.zip",
-        "testsUrl": "http://ftp.mozilla.org/pub/mozilla.org/b2g/tinderbox-builds/fx-team-linux64_gecko/1376506343/b2g-26.0a1.multi.linux-x86_64.tests.zip",
-        "toolsdir": "/builds/slave/fx-team-l64_g-0000000000000000/tools"
-      },
-      "reason": "scheduler",
-      "request_ids": [
-        28008486
-      ],
-      "requesttime": 1376506344,
-      "result": 0,
-      "slave_id": 4814,
-      "starttime": 1376506358
-    },
-    {
-      "builder_id": 172630,
-      "buildnumber": 28,
-      "endtime": 1376513849,
-      "id": 26554534,
-      "master_id": 76,
-      "properties": {
-        "basedir": "/builds/slave/b2g_fx-team_emulator_d_dep-000",
-        "branch": "test_treeherder",
-        "buildername": "b2g_fx-team_emulator_debug_dep",
-        "buildid": "20130814115223",
-        "buildnumber": 28,
-        "builduid": "5114acb64a934005a3dca5605cee6605",
-        "compare_locales_revision": "1fc4e9bc8287",
-        "gaia_revision": "b0f83dd98f0068168e793e9d91b1be879812f5a1",
-        "gecko_revision": "919119a88080",
-        "log_url": "http://ftp.mozilla.org/pub/mozilla.org/b2g/tinderbox-builds/fx-team-emulator_debug/1376506343/b2g_fx-team_emulator_debug_dep-bm57-build1-build28.txt.gz",
-        "master": "http://buildbot-master57.srv.releng.use1.mozilla.com:8001/",
-        "platform": "emulator_debug",
-        "product": "b2g",
-        "project": "",
-        "repo_path": "integration/fx-team",
-        "repository": "",
-        "request_ids": [
-          28008484
-        ],
-        "request_times": {
-          "28008484": 1376506344
-        },
-        "revision": "919119a88080",
-        "scheduler": "b2g_fx-team-b2g",
-        "script_repo_revision": "2f85052176f2",
-        "slavename": "bld-linux64-ec2-025"
-      },
-      "reason": "scheduler",
-      "request_ids": [
-        28008484
-      ],
-      "requesttime": 1376506344,
-      "result": 0,
-      "slave_id": 2187,
-      "starttime": 1376506358
-    },
-    {
-      "builder_id": 120237,
-      "buildnumber": 59,
-      "endtime": 1376512843,
-      "id": 26553980,
-      "master_id": 86,
-      "properties": {
-        "basedir": "/builds/slave/b2g_fx-team_unagi_dep-00000000",
-        "branch": "test_treeherder",
-        "buildername": "b2g_fx-team_unagi_dep",
-        "buildid": "20130814115223",
-        "buildnumber": 59,
-        "builduid": "5114acb64a934005a3dca5605cee6605",
-        "compare_locales_revision": "1fc4e9bc8287",
-        "gaia_revision": "b0f83dd98f0068168e793e9d91b1be879812f5a1",
-        "gecko_revision": "919119a88080",
-        "log_url": "http://ftp.mozilla.org/pub/mozilla.org/b2g/tinderbox-builds/fx-team-unagi/1376506343/b2g_fx-team_unagi_dep-bm63-build1-build59.txt.gz",
-        "master": "http://buildbot-master63.srv.releng.use1.mozilla.com:8001/",
-        "platform": "unagi",
-        "product": "b2g",
-        "project": "",
-        "repo_path": "integration/fx-team",
-        "repository": "",
-        "request_ids": [
-          28008488
-        ],
-        "request_times": {
-          "28008488": 1376506344
-        },
-        "revision": "919119a88080",
-        "scheduler": "b2g_fx-team-b2g",
-        "script_repo_revision": "2f85052176f2",
-        "slavename": "bld-linux64-ec2-148"
-      },
-      "reason": "scheduler",
-      "request_ids": [
-        28008488
-      ],
-      "requesttime": 1376506344,
-      "result": 0,
-      "slave_id": 4911,
-      "starttime": 1376506358
-    },
-    {
-      "builder_id": 124441,
-      "buildnumber": 95,
-      "endtime": 1376513277,
-      "id": 26554185,
-      "master_id": 86,
-      "properties": {
-        "appName": "Firefox",
-        "appVersion": "26.0a1",
-        "basedir": "/builds/slave/fx-team-l64-000000000000000000",
-        "branch": "test_treeherder",
-        "builddir": "fx-team-l64-000000000000000000",
-        "buildername": "Linux x86-64 fx-team build",
-        "buildid": "20130814115223",
-        "buildnumber": 95,
-        "builduid": "a2a7488ff85242688aea1521436b5224",
-        "comments": "Bug 689291 test-tabs.test window focus changes active tab | Test output exceeded timeout (60s). r=Mossop",
-        "filepath": null,
-        "forced_clobber": false,
-        "got_revision": "919119a88080",
-        "hashType": "sha512",
-        "jsshellUrl": "http://ftp.mozilla.org/pub/mozilla.org/firefox/tinderbox-builds/fx-team-linux64/1376506343/jsshell-linux-x86_64.zip",
-        "log_url": "http://ftp.mozilla.org/pub/mozilla.org/firefox/tinderbox-builds/fx-team-linux64/1376506343/fx-team-linux64-bm63-build1-build95.txt.gz",
-        "master": "http://buildbot-master63.srv.releng.use1.mozilla.com:8001/",
-        "num_ctors": 168,
-        "packageFilename": "firefox-26.0a1.en-US.linux-x86_64.tar.bz2",
-        "packageHash": "2faaac830aa2999fd33cee9e902d84f92fedbe35d8b3492192ca229bdf4c69aa9159b699893fd4ca6d4f18c72f616d23d5d8dd04e153f2318b4444c69f4fd214",
-        "packageSize": "36971724",
-        "packageUrl": "http://ftp.mozilla.org/pub/mozilla.org/firefox/tinderbox-builds/fx-team-linux64/1376506343/firefox-26.0a1.en-US.linux-x86_64.tar.bz2",
-        "periodic_clobber": false,
-        "platform": "linux64",
-        "product": "firefox",
-        "project": "",
-        "purge_actual": "126.94GB",
-        "purge_target": "14GB",
-        "purged_clobber": true,
-        "repository": "",
-        "request_ids": [
-          28008476
-        ],
-        "request_times": {
-          "28008476": 1376506344
-        },
-        "revision": "919119a88080a7642cdc6f3609766cb84b642e1a",
-        "scheduler": "fx-team-firefox",
-        "slavebuilddir": "fx-team-l64-000000000000000000",
-        "slavename": "bld-linux64-ec2-131",
-        "sourcestamp": "919119a88080",
-        "stage_platform": "linux64",
-        "symbolsUrl": "http://ftp.mozilla.org/pub/mozilla.org/firefox/tinderbox-builds/fx-team-linux64/1376506343/firefox-26.0a1.en-US.linux-x86_64.crashreporter-symbols.zip",
-        "testresults": [
-          [
-            "num_ctors",
-            "num_ctors",
-            168,
-            "168"
-          ]
-        ],
-        "testsUrl": "http://ftp.mozilla.org/pub/mozilla.org/firefox/tinderbox-builds/fx-team-linux64/1376506343/firefox-26.0a1.en-US.linux-x86_64.tests.zip",
-        "toolsdir": "/builds/slave/fx-team-l64-000000000000000000/tools"
-      },
-      "reason": "scheduler",
-      "request_ids": [
-        28008476
-      ],
-      "requesttime": 1376506344,
-      "result": 0,
-      "slave_id": 4906,
-      "starttime": 1376506358
-    },
-    {
-      "builder_id": 125869,
-      "buildnumber": 83,
-      "endtime": 1376513165,
-      "id": 26554118,
-      "master_id": 86,
-      "properties": {
-        "appName": "Firefox",
-        "appVersion": "26.0a1",
-        "basedir": "/builds/slave/fx-team-lx-0000000000000000000",
-        "branch": "test_treeherder",
-        "builddir": "fx-team-lx-0000000000000000000",
-        "buildername": "Linux fx-team build",
-        "buildid": "20130814115223",
-        "buildnumber": 83,
-        "builduid": "a2a7488ff85242688aea1521436b5224",
-        "comments": "Bug 689291 test-tabs.test window focus changes active tab | Test output exceeded timeout (60s). r=Mossop",
-        "filepath": null,
-        "forced_clobber": false,
-        "got_revision": "919119a88080",
-        "hashType": "sha512",
-        "jsshellUrl": "http://ftp.mozilla.org/pub/mozilla.org/firefox/tinderbox-builds/fx-team-linux/1376506343/jsshell-linux-i686.zip",
-        "log_url": "http://ftp.mozilla.org/pub/mozilla.org/firefox/tinderbox-builds/fx-team-linux/1376506343/fx-team-linux-bm63-build1-build83.txt.gz",
-        "master": "http://buildbot-master63.srv.releng.use1.mozilla.com:8001/",
-        "num_ctors": 168,
-        "packageFilename": "firefox-26.0a1.en-US.linux-i686.tar.bz2",
-        "packageHash": "8ef0ce8a569373d591b7d35edc0517065eee02a579da3bd6ba4056c474c0a4c1a9055fc848dcf87649380e5b18935bc158f8777acf61cbc0e97ef91886e13212",
-        "packageSize": "36580672",
-        "packageUrl": "http://ftp.mozilla.org/pub/mozilla.org/firefox/tinderbox-builds/fx-team-linux/1376506343/firefox-26.0a1.en-US.linux-i686.tar.bz2",
-        "periodic_clobber": false,
-        "platform": "linux",
-        "product": "firefox",
-        "project": "",
-        "purge_actual": "95.11GB",
-        "purge_target": "12GB",
-        "purged_clobber": false,
-        "repository": "",
-        "request_ids": [
-          28008474
-        ],
-        "request_times": {
-          "28008474": 1376506344
-        },
-        "revision": "919119a88080a7642cdc6f3609766cb84b642e1a",
-        "scheduler": "fx-team-firefox",
-        "slavebuilddir": "fx-team-lx-0000000000000000000",
-        "slavename": "bld-linux64-ec2-039",
-        "sourcestamp": "919119a88080",
-        "stage_platform": "linux",
-        "symbolsUrl": "http://ftp.mozilla.org/pub/mozilla.org/firefox/tinderbox-builds/fx-team-linux/1376506343/firefox-26.0a1.en-US.linux-i686.crashreporter-symbols.zip",
-        "testresults": [
-          [
-            "num_ctors",
-            "num_ctors",
-            168,
-            "168"
-          ]
-        ],
-        "testsUrl": "http://ftp.mozilla.org/pub/mozilla.org/firefox/tinderbox-builds/fx-team-linux/1376506343/firefox-26.0a1.en-US.linux-i686.tests.zip",
-        "toolsdir": "/builds/slave/fx-team-lx-0000000000000000000/tools"
-      },
-      "reason": "scheduler",
-      "request_ids": [
-        28008474
-      ],
-      "requesttime": 1376506344,
-      "result": 0,
-      "slave_id": 2260,
-      "starttime": 1376506360
-    },
-    {
-      "builder_id": 126001,
-      "buildnumber": 83,
-      "endtime": 1376512038,
-      "id": 26553604,
-      "master_id": 86,
-      "properties": {
-        "appName": "Fennec",
-        "appVersion": "26.0a1",
-        "basedir": "/builds/slave/fx-team-and-d-0000000000000000",
-        "branch": "test_treeherder",
-        "builddir": "fx-team-and-d-0000000000000000",
-        "buildername": "Android Debug fx-team build",
-        "buildid": "20130814115224",
-        "buildnumber": 83,
-        "builduid": "52d2fd23d8da4748b92df17ed1dbde5c",
-        "comments": "Bug 689291 test-tabs.test window focus changes active tab | Test output exceeded timeout (60s). r=Mossop",
-        "filepath": null,
-        "forced_clobber": false,
-        "got_revision": "919119a88080",
-        "hashType": "sha512",
-        "jsshellUrl": "http://ftp.mozilla.org/pub/mozilla.org/mobile/tinderbox-builds/fx-team-android-debug/1376506344/jsshell-android-arm.zip",
-        "log_url": "http://ftp.mozilla.org/pub/mozilla.org/mobile/tinderbox-builds/fx-team-android-debug/1376506344/fx-team-android-debug-bm63-build1-build83.txt.gz",
-        "master": "http://buildbot-master63.srv.releng.use1.mozilla.com:8001/",
-        "packageFilename": "fennec-26.0a1.en-US.android-arm.apk",
-        "packageHash": "4695217eb4dffd24607459545c62d08aac22c8dceb0b05c99fd3d73f45a5447cd74bf102c4dbeb2fa086271e3242195426994a8aa00f4304e7d31a825f5a314d",
-        "packageSize": "30460960",
-        "packageUrl": "http://ftp.mozilla.org/pub/mozilla.org/mobile/tinderbox-builds/fx-team-android-debug/1376506344/fennec-26.0a1.en-US.android-arm.apk",
-        "periodic_clobber": false,
-        "platform": "android-debug",
-        "product": "mobile",
-        "project": "",
-        "purge_actual": "123.78GB",
-        "purge_target": "14GB",
-        "purged_clobber": true,
-        "repository": "",
-        "request_ids": [
-          28008491
-        ],
-        "request_times": {
-          "28008491": 1376506344
-        },
-        "revision": "919119a88080a7642cdc6f3609766cb84b642e1a",
-        "robocopApkUrl": "http://ftp.mozilla.org/pub/mozilla.org/mobile/tinderbox-builds/fx-team-android-debug/1376506344/robocop.apk",
-        "scheduler": "fx-team-mobile",
-        "slavebuilddir": "fx-team-and-d-0000000000000000",
-        "slavename": "bld-linux64-ec2-153",
-        "stage_platform": "android-debug",
-        "symbolsUrl": "http://ftp.mozilla.org/pub/mozilla.org/mobile/tinderbox-builds/fx-team-android-debug/1376506344/fennec-26.0a1.en-US.android-arm.crashreporter-symbols.zip",
-        "testsUrl": "http://ftp.mozilla.org/pub/mozilla.org/mobile/tinderbox-builds/fx-team-android-debug/1376506344/fennec-26.0a1.en-US.android-arm.tests.zip",
-        "toolsdir": "/builds/slave/fx-team-and-d-0000000000000000/tools",
-        "unsignedApkUrl": "http://ftp.mozilla.org/pub/mozilla.org/mobile/tinderbox-builds/fx-team-android-debug/1376506344/gecko-unsigned-unaligned.apk"
-      },
-      "reason": "scheduler",
-      "request_ids": [
-        28008491
-      ],
-      "requesttime": 1376506344,
-      "result": 0,
-      "slave_id": 4842,
-      "starttime": 1376506361
-    },
-    {
-      "builder_id": 120835,
-      "buildnumber": 810,
-      "endtime": 1376511670,
-      "id": 26553420,
-      "master_id": 88,
-      "properties": {
-        "basedir": "/builds/slave/talos-slave/test",
-        "branch": "test_treeherder",
-        "build_url": "http://ftp.mozilla.org/pub/mozilla.org/firefox/tinderbox-builds/mozilla-inbound-macosx64/1376494882/firefox-26.0a1.en-US.mac.dmg",
-        "builddir": "mozilla-inbound_lion_test-svgr",
-        "buildername": "Rev4 MacOSX Lion 10.7 mozilla-inbound talos svgr",
-        "buildid": "20130814084122",
-        "buildnumber": 810,
-        "builduid": "f4a9d75c6bdc4f5a9fb99f067f6252d9",
-        "log_url": "http://ftp.mozilla.org/pub/mozilla.org/firefox/tinderbox-builds/mozilla-inbound-macosx64/1376494882/mozilla-inbound_lion_test-svgr-bm76-tests1-macosx-build810.txt.gz",
-        "master": "http://buildbot-master76.srv.releng.use1.mozilla.com:8201/",
-        "pgo_build": "False",
-        "platform": "lion",
-        "product": "firefox",
-        "project": "",
-        "repository": "",
-        "request_ids": [
-          28007789
-        ],
-        "request_times": {
-          "28007789": 1376505664
-        },
-        "revision": "e74f694f31c4",
-        "scheduler": "tests-mozilla-inbound-macosx64-talos",
-        "script_repo_revision": "2f85052176f2",
-        "slavebuilddir": "test",
-        "slavename": "talos-r4-lion-017",
-        "stage_platform": "macosx64"
-      },
-      "reason": "Self-serve: Rebuilt by philringnalda@gmail.com",
-      "request_ids": [
-        28007789
-      ],
-      "requesttime": 1376505664,
-      "result": 2,
-      "slave_id": 1569,
-      "starttime": 1376506436
-    },
-    {
-      "builder_id": 125627,
-      "buildnumber": 109,
-      "endtime": 1376513183,
-      "id": 26554128,
-      "master_id": 82,
-      "properties": {
-        "basedir": "/Users/cltbld/talos-slave/test",
-        "branch": "test_treeherder",
-        "build_url": "http://ftp.mozilla.org/pub/mozilla.org/firefox/tinderbox-builds/mozilla-aurora-macosx64-debug/1376502381/firefox-25.0a2.en-US.mac64.dmg",
-        "buildername": "Rev4 MacOSX Snow Leopard 10.6 mozilla-aurora debug test mochitest-browser-chrome",
-        "buildid": "20130814104621",
-        "buildnumber": 109,
-        "builduid": "f377570e7b214e2185ad6a995fde8820",
-        "log_url": "http://ftp.mozilla.org/pub/mozilla.org/firefox/tinderbox-builds/mozilla-aurora-macosx64-debug/1376502381/mozilla-aurora_snowleopard-debug_test-mochitest-browser-chrome-bm75-tests1-macosx-build109.txt.gz",
-        "master": "http://buildbot-master75.srv.releng.use1.mozilla.com:8201/",
-        "pgo_build": "False",
-        "platform": "macosx64",
-        "product": "firefox",
-        "project": "",
-        "repo_path": "releases/mozilla-aurora",
-        "repository": "",
-        "request_ids": [
-          28008392
-        ],
-        "request_times": {
-          "28008392": 1376506135
-        },
-        "revision": "932868752f82",
-        "scheduler": "tests-mozilla-aurora-snowleopard-debug-unittest",
-        "script_repo_revision": "2f85052176f2",
-        "slavebuilddir": "test",
-        "slavename": "talos-r4-snow-049",
-        "stage_platform": "macosx64",
-        "symbols_url": "http://ftp.mozilla.org/pub/mozilla.org/firefox/tinderbox-builds/mozilla-aurora-macosx64-debug/1376502381/firefox-25.0a2.en-US.mac64.crashreporter-symbols.zip"
-      },
-      "reason": "scheduler",
-      "request_ids": [
-        28008392
-      ],
-      "requesttime": 1376506135,
-      "result": 0,
-      "slave_id": 1398,
-      "starttime": 1376506451
-    },
-    {
-      "builder_id": 123187,
-      "buildnumber": 1011,
-      "endtime": 1376511713,
-      "id": 26553441,
-      "master_id": 87,
-      "properties": {
-        "basedir": "/builds/slave/talos-slave/test",
-        "branch": "test_treeherder",
-        "build_url": "http://ftp.mozilla.org/pub/mozilla.org/firefox/tinderbox-builds/mozilla-inbound-macosx64/1376493503/firefox-26.0a1.en-US.mac.dmg",
-        "builddir": "mozilla-inbound_lion_test-svgr",
-        "buildername": "Rev4 MacOSX Lion 10.7 mozilla-inbound talos svgr",
-        "buildid": "20130814081823",
-        "buildnumber": 1011,
-        "builduid": "f6f693c834704035b029967c6cbd7c92",
-        "log_url": "http://ftp.mozilla.org/pub/mozilla.org/firefox/tinderbox-builds/mozilla-inbound-macosx64/1376493503/mozilla-inbound_lion_test-svgr-bm78-tests1-macosx-build1011.txt.gz",
-        "master": "http://buildbot-master78.srv.releng.usw2.mozilla.com:8201/",
-        "pgo_build": "False",
-        "platform": "lion",
-        "product": "firefox",
-        "project": "",
-        "repository": "",
-        "request_ids": [
-          28007838
-        ],
-        "request_times": {
-          "28007838": 1376505674
-        },
-        "revision": "e91392042137",
-        "scheduler": "tests-mozilla-inbound-macosx64-talos",
-        "script_repo_revision": "2f85052176f2",
-        "slavebuilddir": "test",
-        "slavename": "talos-r4-lion-057",
-        "stage_platform": "macosx64"
-      },
-      "reason": "Self-serve: Rebuilt by philringnalda@gmail.com",
-      "request_ids": [
-        28007838
-      ],
-      "requesttime": 1376505674,
-      "result": 2,
-      "slave_id": 1628,
-      "starttime": 1376506485
-    },
-    {
-      "builder_id": 143422,
-      "buildnumber": 1579,
-      "endtime": 1376513913,
-      "id": 26554573,
-      "master_id": 75,
-      "properties": {
-        "appName": "Firefox",
-        "appVersion": "26.0a1",
-        "basedir": "/builds/slave/try-osx64-00000000000000000000",
-        "branch": "test_treeherder",
-        "builddir": "try-osx64-00000000000000000000",
-        "buildername": "OS X 10.7 try build",
-        "buildid": "20130814115549",
-        "buildnumber": 1579,
-        "builduid": "9280004e244143ab9e323b9ec63c4db9",
-        "comments": "try: -b do -p linux64,macosx64,win32 -u xpcshell,mochitests -t none",
-        "filepath": null,
-        "forced_clobber": false,
-        "got_revision": "13f8f04eb4de",
-        "hashType": "sha512",
-        "jsshellUrl": "http://ftp.mozilla.org/pub/mozilla.org/firefox/try-builds/mixedpuppy@gmail.com-13f8f04eb4de/try-macosx64/jsshell-mac.zip",
-        "log_url": "http://ftp.mozilla.org/pub/mozilla.org/firefox/try-builds/mixedpuppy@gmail.com-13f8f04eb4de/try-macosx64/try-macosx64-bm55-try1-build1579.txt.gz",
-        "master": "http://buildbot-master55.srv.releng.use1.mozilla.com:8101/",
-        "packageFilename": "firefox-26.0a1.en-US.mac.dmg",
-        "packageHash": "9c07e193b33505ccc837d0c86a03e95309917a806b815d7a17c04f9b1bc1d79cc6c596c9f2d32845d4f4bdf2b7b398ecde1db2f92d371aad33f1b8ab239bfb66",
-        "packageSize": "62721463",
-        "packageUrl": "http://ftp.mozilla.org/pub/mozilla.org/firefox/try-builds/mixedpuppy@gmail.com-13f8f04eb4de/try-macosx64/firefox-26.0a1.en-US.mac.dmg",
-        "periodic_clobber": false,
-        "platform": "macosx64",
-        "product": "firefox",
-        "project": "",
-        "purge_actual": "867.85GB",
-        "purge_target": "12GB",
-        "purged_clobber": false,
-        "repository": "",
-        "request_ids": [
-          28008562
-        ],
-        "request_times": {
-          "28008562": 1376506522
-        },
-        "revision": "13f8f04eb4de715c3ff92f496e95031a52fccce5",
-        "scheduler": "try-firefox",
-        "slavebuilddir": "try-osx64-00000000000000000000",
-        "slavename": "bld-lion-r5-029",
-        "stage_platform": "macosx64",
-        "symbolsUrl": "http://ftp.mozilla.org/pub/mozilla.org/firefox/try-builds/mixedpuppy@gmail.com-13f8f04eb4de/try-macosx64/firefox-26.0a1.en-US.mac.crashreporter-symbols.zip",
-        "testsUrl": "http://ftp.mozilla.org/pub/mozilla.org/firefox/try-builds/mixedpuppy@gmail.com-13f8f04eb4de/try-macosx64/firefox-26.0a1.en-US.mac.tests.zip",
-        "toolsdir": "/builds/slave/try-osx64-00000000000000000000/tools",
-        "who": "mixedpuppy@gmail.com"
-      },
-      "reason": "scheduler",
-      "request_ids": [
-        28008562
-      ],
-      "requesttime": 1376506522,
-      "result": 1,
-      "slave_id": 2029,
-      "starttime": 1376506534
-    },
-    {
-      "builder_id": 116906,
-      "buildnumber": 1858,
-      "endtime": 1376512670,
-      "id": 26553911,
-      "master_id": 75,
-      "properties": {
-        "appName": "Firefox",
-        "appVersion": "26.0a1",
-        "basedir": "/builds/slave/try-l64-d-00000000000000000000",
-        "branch": "test_treeherder",
-        "builddir": "try-l64-d-00000000000000000000",
-        "buildername": "Linux x86-64 try leak test build",
-        "buildid": "20130814120021",
-        "buildnumber": 1858,
-        "builduid": "9280004e244143ab9e323b9ec63c4db9",
-        "comments": "try: -b do -p linux64,macosx64,win32 -u xpcshell,mochitests -t none",
-        "filepath": null,
-        "forced_clobber": false,
-        "got_revision": "13f8f04eb4de",
-        "hashType": "sha512",
-        "jsshellUrl": "http://ftp.mozilla.org/pub/mozilla.org/firefox/try-builds/mixedpuppy@gmail.com-13f8f04eb4de/try-linux64-debug/jsshell-linux-x86_64.zip",
-        "log_url": "http://ftp.mozilla.org/pub/mozilla.org/firefox/try-builds/mixedpuppy@gmail.com-13f8f04eb4de/try-linux64-debug/try-linux64-debug-bm55-try1-build1858.txt.gz",
-        "master": "http://buildbot-master55.srv.releng.use1.mozilla.com:8101/",
-        "packageFilename": "firefox-26.0a1.en-US.linux-x86_64.tar.bz2",
-        "packageHash": "47bb7a9f6de5c24484596125ed68b2b34d88fe600f617d3df6e7f7a21b6c1bb34d0d7e19b8b6f86b8b25e9062c7e0f8178410ced3ecaf055a17bc082e1ad08e7",
-        "packageSize": "33452119",
-        "packageUrl": "http://ftp.mozilla.org/pub/mozilla.org/firefox/try-builds/mixedpuppy@gmail.com-13f8f04eb4de/try-linux64-debug/firefox-26.0a1.en-US.linux-x86_64.tar.bz2",
-        "periodic_clobber": false,
-        "platform": "linux64-debug",
-        "product": "firefox",
-        "project": "",
-        "purge_actual": "18.55GB",
-        "purge_target": "14GB",
-        "purged_clobber": false,
-        "repository": "",
-        "request_ids": [
-          28008563
-        ],
-        "request_times": {
-          "28008563": 1376506522
-        },
-        "revision": "13f8f04eb4de715c3ff92f496e95031a52fccce5",
-        "scheduler": "try-firefox",
-        "slavebuilddir": "try-l64-d-00000000000000000000",
-        "slavename": "bld-linux64-ix-050",
-        "stage_platform": "linux64-debug",
-        "symbolsUrl": "http://ftp.mozilla.org/pub/mozilla.org/firefox/try-builds/mixedpuppy@gmail.com-13f8f04eb4de/try-linux64-debug/firefox-26.0a1.en-US.linux-x86_64.crashreporter-symbols.zip",
-        "testsUrl": "http://ftp.mozilla.org/pub/mozilla.org/firefox/try-builds/mixedpuppy@gmail.com-13f8f04eb4de/try-linux64-debug/firefox-26.0a1.en-US.linux-x86_64.tests.zip",
-        "toolsdir": "/builds/slave/try-l64-d-00000000000000000000/tools",
-        "who": "mixedpuppy@gmail.com"
-      },
-      "reason": "scheduler",
-      "request_ids": [
-        28008563
-      ],
-      "requesttime": 1376506522,
-      "result": 0,
-      "slave_id": 3814,
-      "starttime": 1376506535
-    },
-    {
-      "builder_id": 144677,
-      "buildnumber": 1315,
-      "endtime": 1376515408,
-      "id": 26555433,
-      "master_id": 75,
-      "properties": {
-        "appName": "Firefox",
-        "appVersion": "26.0a1",
-        "basedir": "e:/builds/moz2_slave/try-w32-d-00000000000000000000",
-        "branch": "test_treeherder",
-        "builddir": "try-w32-d-00000000000000000000",
-        "buildername": "WINNT 5.2 try leak test build",
-        "buildid": "20130814115558",
-        "buildnumber": 1315,
-        "builduid": "9280004e244143ab9e323b9ec63c4db9",
-        "comments": "try: -b do -p linux64,macosx64,win32 -u xpcshell,mochitests -t none",
-        "filepath": null,
-        "forced_clobber": false,
-        "got_revision": "13f8f04eb4de",
-        "hashType": "sha512",
-        "installerFilename": "firefox-26.0a1.en-US.win32.installer.exe",
-        "installerHash": "0ffaae6b5792bf1fc5e8ed43de88718a8a034794b91abeda15fd46cc3aaec0917be826f6acb96beabb30e453e64b26c87e6dd19f7c55d561ae21be0c242f1d6b",
-        "installerSize": "27516424",
-        "jsshellUrl": "http://ftp.mozilla.org/pub/mozilla.org/firefox/try-builds/mixedpuppy@gmail.com-13f8f04eb4de/try-win32-debug/jsshell-win32.zip",
-        "log_url": "http://ftp.mozilla.org/pub/mozilla.org/firefox/try-builds/mixedpuppy@gmail.com-13f8f04eb4de/try-win32-debug/try-win32-debug-bm55-try1-build1315.txt.gz",
-        "master": "http://buildbot-master55.srv.releng.use1.mozilla.com:8101/",
-        "packageFilename": "firefox-26.0a1.en-US.win32.zip",
-        "packageHash": "4119a32d0cebc2abe423134155bae9e7bc2f8de85a7e118e32e5585d7ed03f2ef62da6ce8061af6e191f5a7c748a8bed06f31536b8239a5284d116b50660f8e0",
-        "packageSize": "38071077",
-        "packageUrl": "http://ftp.mozilla.org/pub/mozilla.org/firefox/try-builds/mixedpuppy@gmail.com-13f8f04eb4de/try-win32-debug/firefox-26.0a1.en-US.win32.zip",
-        "periodic_clobber": false,
-        "platform": "win32-debug",
-        "product": "firefox",
-        "project": "",
-        "purge_actual": "156.42GB",
-        "purge_target": "9GB",
-        "purged_clobber": false,
-        "repository": "",
-        "request_ids": [
-          28008560
-        ],
-        "request_times": {
-          "28008560": 1376506522
-        },
-        "revision": "13f8f04eb4de715c3ff92f496e95031a52fccce5",
-        "scheduler": "try-firefox",
-        "slavebuilddir": "try-w32-d-00000000000000000000",
-        "slavename": "w64-ix-slave26",
-        "stage_platform": "win32-debug",
-        "symbolsUrl": "http://ftp.mozilla.org/pub/mozilla.org/firefox/try-builds/mixedpuppy@gmail.com-13f8f04eb4de/try-win32-debug/firefox-26.0a1.en-US.win32.crashreporter-symbols.zip",
-        "testsUrl": "http://ftp.mozilla.org/pub/mozilla.org/firefox/try-builds/mixedpuppy@gmail.com-13f8f04eb4de/try-win32-debug/firefox-26.0a1.en-US.win32.tests.zip",
-        "toolsdir": "e:/builds/moz2_slave/try-w32-d-00000000000000000000/tools",
-        "who": "mixedpuppy@gmail.com"
-      },
-      "reason": "scheduler",
-      "request_ids": [
-        28008560
-      ],
-      "requesttime": 1376506522,
-      "result": 0,
-      "slave_id": 1327,
-      "starttime": 1376506536
-    },
-    {
-      "builder_id": 119562,
-      "buildnumber": 1210,
-      "endtime": 1376512585,
-      "id": 26553863,
-      "master_id": 91,
-      "properties": {
-        "appName": "Firefox",
-        "appVersion": "26.0a1",
-        "basedir": "/builds/slave/try-l64-0000000000000000000000",
-        "branch": "test_treeherder",
-        "builddir": "try-l64-0000000000000000000000",
-        "buildername": "Linux x86-64 try build",
-        "buildid": "20130814120120",
-        "buildnumber": 1210,
-        "builduid": "9280004e244143ab9e323b9ec63c4db9",
-        "comments": "try: -b do -p linux64,macosx64,win32 -u xpcshell,mochitests -t none",
-        "filepath": null,
-        "forced_clobber": false,
-        "got_revision": "13f8f04eb4de",
-        "hashType": "sha512",
-        "jsshellUrl": "http://ftp.mozilla.org/pub/mozilla.org/firefox/try-builds/mixedpuppy@gmail.com-13f8f04eb4de/try-linux64/jsshell-linux-x86_64.zip",
-        "log_url": "http://ftp.mozilla.org/pub/mozilla.org/firefox/try-builds/mixedpuppy@gmail.com-13f8f04eb4de/try-linux64/try-linux64-bm60-try1-build1210.txt.gz",
-        "master": "http://buildbot-master60.srv.releng.usw2.mozilla.com:8101/",
-        "num_ctors": 167,
-        "packageFilename": "firefox-26.0a1.en-US.linux-x86_64.tar.bz2",
-        "packageHash": "80a3eec267b4cbcf1a9703baa3ccfef759676e934286562a9cfbc735467351d262b5fdcaa9fd3c4013b5d4e03f1ec58cfba978dc934749908b849d77c14bef11",
-        "packageSize": "33694270",
-        "packageUrl": "http://ftp.mozilla.org/pub/mozilla.org/firefox/try-builds/mixedpuppy@gmail.com-13f8f04eb4de/try-linux64/firefox-26.0a1.en-US.linux-x86_64.tar.bz2",
-        "periodic_clobber": false,
-        "platform": "linux64",
-        "product": "firefox",
-        "project": "",
-        "purge_actual": "17.29GB",
-        "purge_target": "14GB",
-        "purged_clobber": false,
-        "repository": "",
-        "request_ids": [
-          28008561
-        ],
-        "request_times": {
-          "28008561": 1376506522
-        },
-        "revision": "13f8f04eb4de715c3ff92f496e95031a52fccce5",
-        "scheduler": "try-firefox",
-        "slavebuilddir": "try-l64-0000000000000000000000",
-        "slavename": "bld-linux64-ix-051",
-        "sourcestamp": "13f8f04eb4de",
-        "stage_platform": "linux64",
-        "symbolsUrl": "http://ftp.mozilla.org/pub/mozilla.org/firefox/try-builds/mixedpuppy@gmail.com-13f8f04eb4de/try-linux64/firefox-26.0a1.en-US.linux-x86_64.crashreporter-symbols.zip",
-        "testresults": [
-          [
-            "num_ctors",
-            "num_ctors",
-            167,
-            "167"
-          ]
-        ],
-        "testsUrl": "http://ftp.mozilla.org/pub/mozilla.org/firefox/try-builds/mixedpuppy@gmail.com-13f8f04eb4de/try-linux64/firefox-26.0a1.en-US.linux-x86_64.tests.zip",
-        "toolsdir": "/builds/slave/try-l64-0000000000000000000000/tools",
-        "who": "mixedpuppy@gmail.com"
-      },
-      "reason": "scheduler",
-      "request_ids": [
-        28008561
-      ],
-      "requesttime": 1376506522,
-      "result": 0,
-      "slave_id": 3956,
-      "starttime": 1376506538
-    },
-    {
-      "builder_id": 144336,
-      "buildnumber": 1617,
-      "endtime": 1376513909,
-      "id": 26554566,
-      "master_id": 91,
-      "properties": {
-        "appName": "Firefox",
-        "appVersion": "26.0a1",
-        "basedir": "e:/builds/moz2_slave/try-w32-0000000000000000000000",
-        "branch": "test_treeherder",
-        "builddir": "try-w32-0000000000000000000000",
-        "buildername": "WINNT 5.2 try build",
-        "buildid": "20130814115603",
-        "buildnumber": 1617,
-        "builduid": "9280004e244143ab9e323b9ec63c4db9",
-        "comments": "try: -b do -p linux64,macosx64,win32 -u xpcshell,mochitests -t none",
-        "filepath": null,
-        "forced_clobber": false,
-        "got_revision": "13f8f04eb4de",
-        "hashType": "sha512",
-        "installerFilename": "firefox-26.0a1.en-US.win32.installer.exe",
-        "installerHash": "a3e749721a0467ca56755473b52b62a8a26791bc2b5fce69bf8fe4d74ae5ce12087e1642e77e25285b74b88b474a3ab11c2e34cd44560981196112d0390a38fe",
-        "installerSize": "23615872",
-        "jsshellUrl": "http://ftp.mozilla.org/pub/mozilla.org/firefox/try-builds/mixedpuppy@gmail.com-13f8f04eb4de/try-win32/jsshell-win32.zip",
-        "log_url": "http://ftp.mozilla.org/pub/mozilla.org/firefox/try-builds/mixedpuppy@gmail.com-13f8f04eb4de/try-win32/try-win32-bm60-try1-build1617.txt.gz",
-        "master": "http://buildbot-master60.srv.releng.usw2.mozilla.com:8101/",
-        "packageFilename": "firefox-26.0a1.en-US.win32.zip",
-        "packageHash": "34dab89060ff7e9e49d1ac06643480c966218653788b198c776d1845274dfdfdc75c5b09be4934c8fa973d60b06cf7626df1ab15d76376edfa9b9b5c9f8c18ab",
-        "packageSize": "30477390",
-        "packageUrl": "http://ftp.mozilla.org/pub/mozilla.org/firefox/try-builds/mixedpuppy@gmail.com-13f8f04eb4de/try-win32/firefox-26.0a1.en-US.win32.zip",
-        "periodic_clobber": false,
-        "platform": "win32",
-        "product": "firefox",
-        "project": "",
-        "purge_actual": "162.13GB",
-        "purge_target": "12GB",
-        "purged_clobber": false,
-        "repository": "",
-        "request_ids": [
-          28008565
-        ],
-        "request_times": {
-          "28008565": 1376506522
-        },
-        "revision": "13f8f04eb4de715c3ff92f496e95031a52fccce5",
-        "scheduler": "try-firefox",
-        "slavebuilddir": "try-w32-0000000000000000000000",
-        "slavename": "w64-ix-slave52",
-        "stage_platform": "win32",
-        "symbolsUrl": "http://ftp.mozilla.org/pub/mozilla.org/firefox/try-builds/mixedpuppy@gmail.com-13f8f04eb4de/try-win32/firefox-26.0a1.en-US.win32.crashreporter-symbols.zip",
-        "testsUrl": "http://ftp.mozilla.org/pub/mozilla.org/firefox/try-builds/mixedpuppy@gmail.com-13f8f04eb4de/try-win32/firefox-26.0a1.en-US.win32.tests.zip",
-        "toolsdir": "e:/builds/moz2_slave/try-w32-0000000000000000000000/tools",
-        "who": "mixedpuppy@gmail.com"
-      },
-      "reason": "scheduler",
-      "request_ids": [
-        28008565
-      ],
-      "requesttime": 1376506522,
-      "result": 0,
-      "slave_id": 2125,
-      "starttime": 1376506538
-    },
-    {
-      "builder_id": 123081,
-      "buildnumber": 144,
-      "endtime": 1376511766,
-      "id": 26553472,
-      "master_id": 88,
-      "properties": {
-        "basedir": "/builds/slave/talos-slave/test",
-        "branch": "test_treeherder",
-        "build_url": "http://ftp.mozilla.org/pub/mozilla.org/firefox/tinderbox-builds/fx-team-macosx64/1376499083/firefox-26.0a1.en-US.mac.dmg",
-        "builddir": "fx-team_lion_test-svgr",
-        "buildername": "Rev4 MacOSX Lion 10.7 fx-team talos svgr",
-        "buildid": "20130814095123",
-        "buildnumber": 144,
-        "builduid": "e8c4556f1c404e6786b4d1e975cdf9c5",
-        "log_url": "http://ftp.mozilla.org/pub/mozilla.org/firefox/tinderbox-builds/fx-team-macosx64/1376499083/fx-team_lion_test-svgr-bm76-tests1-macosx-build144.txt.gz",
-        "master": "http://buildbot-master76.srv.releng.use1.mozilla.com:8201/",
-        "pgo_build": "False",
-        "platform": "lion",
-        "product": "firefox",
-        "project": "",
-        "repository": "",
-        "request_ids": [
-          28007904
-        ],
-        "request_times": {
-          "28007904": 1376505775
-        },
-        "revision": "15aaeae89597",
-        "scheduler": "tests-fx-team-macosx64-talos",
-        "script_repo_revision": "2f85052176f2",
-        "slavebuilddir": "test",
-        "slavename": "talos-r4-lion-053",
-        "stage_platform": "macosx64"
-      },
-      "reason": "scheduler",
-      "request_ids": [
-        28007904
-      ],
-      "requesttime": 1376505775,
-      "result": 2,
-      "slave_id": 1638,
-      "starttime": 1376506546
-    },
-    {
-      "builder_id": 119023,
-      "buildnumber": 121,
-      "endtime": 1376511494,
-      "id": 26553349,
-      "master_id": 88,
-      "properties": {
-        "basedir": "/Users/cltbld/talos-slave/test",
-        "branch": "test_treeherder",
-        "build_url": "http://ftp.mozilla.org/pub/mozilla.org/firefox/tinderbox-builds/fx-team-macosx64/1376499083/firefox-26.0a1.en-US.mac.dmg",
-        "builddir": "fx-team_lion_test-tp5o",
-        "buildername": "Rev4 MacOSX Lion 10.7 fx-team talos tp5o",
-        "buildid": "20130814095123",
-        "buildnumber": 121,
-        "builduid": "e8c4556f1c404e6786b4d1e975cdf9c5",
-        "log_url": "http://ftp.mozilla.org/pub/mozilla.org/firefox/tinderbox-builds/fx-team-macosx64/1376499083/fx-team_lion_test-tp5o-bm76-tests1-macosx-build121.txt.gz",
-        "master": "http://buildbot-master76.srv.releng.use1.mozilla.com:8201/",
-        "pgo_build": "False",
-        "platform": "lion",
-        "product": "firefox",
-        "project": "",
-        "repository": "",
-        "request_ids": [
-          28007907
-        ],
-        "request_times": {
-          "28007907": 1376505775
-        },
-        "revision": "15aaeae89597",
-        "scheduler": "tests-fx-team-macosx64-talos",
-        "script_repo_revision": "2f85052176f2",
-        "slavebuilddir": "test",
-        "slavename": "talos-r4-lion-088",
-        "stage_platform": "macosx64"
-      },
-      "reason": "scheduler",
-      "request_ids": [
-        28007907
-      ],
-      "requesttime": 1376505775,
-      "result": 2,
-      "slave_id": 4132,
-      "starttime": 1376506584
-    },
-    {
-      "builder_id": 119562,
-      "buildnumber": 1211,
-      "endtime": 1376513120,
-      "id": 26554092,
-      "master_id": 91,
-      "properties": {
-        "appName": "Firefox",
-        "appVersion": "26.0a1",
-        "basedir": "/builds/slave/try-l64-0000000000000000000000",
-        "branch": "test_treeherder",
-        "builddir": "try-l64-0000000000000000000000",
-        "buildername": "Linux x86-64 try build",
-        "buildid": "20130814120010",
-        "buildnumber": 1211,
-        "builduid": "3bec12d78d22411f84cfc1d4c7f266b9",
-        "comments": "try: try: -b do -p linux64 -u all -t none Bug 903212 - Remove explicit compartment GC machinery. v1 r=smaug",
-        "filepath": null,
-        "forced_clobber": false,
-        "got_revision": "f638d36d5a74",
-        "hashType": "sha512",
-        "jsshellUrl": "http://ftp.mozilla.org/pub/mozilla.org/firefox/try-builds/bobbyholley@gmail.com-f638d36d5a74/try-linux64/jsshell-linux-x86_64.zip",
-        "log_url": "http://ftp.mozilla.org/pub/mozilla.org/firefox/try-builds/bobbyholley@gmail.com-f638d36d5a74/try-linux64/try-linux64-bm60-try1-build1211.txt.gz",
-        "master": "http://buildbot-master60.srv.releng.usw2.mozilla.com:8101/",
-        "num_ctors": 168,
-        "packageFilename": "firefox-26.0a1.en-US.linux-x86_64.tar.bz2",
-        "packageHash": "e6c92d9a83142597c294dfcb8491cb45cb487345e1483bb493469630bb7d255d068c71f07061d227349b44b0e5a90f1f75af87a2568fdc40bb0e1b35c0163d47",
-        "packageSize": "37052544",
-        "packageUrl": "http://ftp.mozilla.org/pub/mozilla.org/firefox/try-builds/bobbyholley@gmail.com-f638d36d5a74/try-linux64/firefox-26.0a1.en-US.linux-x86_64.tar.bz2",
-        "periodic_clobber": false,
-        "platform": "linux64",
-        "product": "firefox",
-        "project": "",
-        "purge_actual": "20.74GB",
-        "purge_target": "14GB",
-        "purged_clobber": false,
-        "repository": "",
-        "request_ids": [
-          28008566
-        ],
-        "request_times": {
-          "28008566": 1376506582
-        },
-        "revision": "f638d36d5a7434d34fb6f4dea6943da04e6a069e",
-        "scheduler": "try-firefox",
-        "slavebuilddir": "try-l64-0000000000000000000000",
-        "slavename": "try-linux64-ec2-313",
-        "sourcestamp": "f638d36d5a74",
-        "stage_platform": "linux64",
-        "symbolsUrl": "http://ftp.mozilla.org/pub/mozilla.org/firefox/try-builds/bobbyholley@gmail.com-f638d36d5a74/try-linux64/firefox-26.0a1.en-US.linux-x86_64.crashreporter-symbols.zip",
-        "testresults": [
-          [
-            "num_ctors",
-            "num_ctors",
-            168,
-            "168"
-          ]
-        ],
-        "testsUrl": "http://ftp.mozilla.org/pub/mozilla.org/firefox/try-builds/bobbyholley@gmail.com-f638d36d5a74/try-linux64/firefox-26.0a1.en-US.linux-x86_64.tests.zip",
-        "toolsdir": "/builds/slave/try-l64-0000000000000000000000/tools",
-        "who": "bobbyholley@gmail.com"
-      },
-      "reason": "scheduler",
-      "request_ids": [
-        28008566
-      ],
-      "requesttime": 1376506582,
-      "result": 0,
-      "slave_id": 2613,
-      "starttime": 1376506598
-    },
-    {
-      "builder_id": 123064,
-      "buildnumber": 1069,
-      "endtime": 1376512672,
-      "id": 26553908,
-      "master_id": 91,
-      "properties": {
-        "appName": "Firefox",
-        "appVersion": "26.0a1",
-        "basedir": "/builds/slave/try-l64-d-00000000000000000000",
-        "branch": "test_treeherder",
-        "builddir": "try-l64-d-00000000000000000000",
-        "buildername": "Linux x86-64 try leak test build",
-        "buildid": "20130814120211",
-        "buildnumber": 1069,
-        "builduid": "3bec12d78d22411f84cfc1d4c7f266b9",
-        "comments": "try: try: -b do -p linux64 -u all -t none Bug 903212 - Remove explicit compartment GC machinery. v1 r=smaug",
-        "filepath": null,
-        "forced_clobber": false,
-        "got_revision": "f638d36d5a74",
-        "hashType": "sha512",
-        "jsshellUrl": "http://ftp.mozilla.org/pub/mozilla.org/firefox/try-builds/bobbyholley@gmail.com-f638d36d5a74/try-linux64-debug/jsshell-linux-x86_64.zip",
-        "log_url": "http://ftp.mozilla.org/pub/mozilla.org/firefox/try-builds/bobbyholley@gmail.com-f638d36d5a74/try-linux64-debug/try-linux64-debug-bm60-try1-build1069.txt.gz",
-        "master": "http://buildbot-master60.srv.releng.usw2.mozilla.com:8101/",
-        "packageFilename": "firefox-26.0a1.en-US.linux-x86_64.tar.bz2",
-        "packageHash": "29c31a61e0460644ae272129db7d68a195189dec73f86eb46ea8e7ca96043938adf0e86d7755dd1a0c41d613341315efaf376af68088909a46e45196d13e399c",
-        "packageSize": "36740398",
-        "packageUrl": "http://ftp.mozilla.org/pub/mozilla.org/firefox/try-builds/bobbyholley@gmail.com-f638d36d5a74/try-linux64-debug/firefox-26.0a1.en-US.linux-x86_64.tar.bz2",
-        "periodic_clobber": false,
-        "platform": "linux64-debug",
-        "product": "firefox",
-        "project": "",
-        "purge_actual": "23.18GB",
-        "purge_target": "14GB",
-        "purged_clobber": false,
-        "repository": "",
-        "request_ids": [
-          28008567
-        ],
-        "request_times": {
-          "28008567": 1376506582
-        },
-        "revision": "f638d36d5a7434d34fb6f4dea6943da04e6a069e",
-        "scheduler": "try-firefox",
-        "slavebuilddir": "try-l64-d-00000000000000000000",
-        "slavename": "bld-linux64-ix-049",
-        "stage_platform": "linux64-debug",
-        "symbolsUrl": "http://ftp.mozilla.org/pub/mozilla.org/firefox/try-builds/bobbyholley@gmail.com-f638d36d5a74/try-linux64-debug/firefox-26.0a1.en-US.linux-x86_64.crashreporter-symbols.zip",
-        "testsUrl": "http://ftp.mozilla.org/pub/mozilla.org/firefox/try-builds/bobbyholley@gmail.com-f638d36d5a74/try-linux64-debug/firefox-26.0a1.en-US.linux-x86_64.tests.zip",
-        "toolsdir": "/builds/slave/try-l64-d-00000000000000000000/tools",
-        "who": "bobbyholley@gmail.com"
-      },
-      "reason": "scheduler",
-      "request_ids": [
-        28008567
-      ],
-      "requesttime": 1376506582,
-      "result": 0,
-      "slave_id": 3818,
-      "starttime": 1376506598
-    },
-    {
-      "builder_id": 159054,
-      "buildnumber": 811,
-      "endtime": 1376511207,
-      "id": 26553189,
-      "master_id": 94,
-      "properties": {
-        "basedir": "C:\\slave\\test",
-        "branch": "test_treeherder",
-        "build_url": "http://ftp.mozilla.org/pub/mozilla.org/firefox/tinderbox-builds/mozilla-inbound-win32-debug/1376500942/firefox-26.0a1.en-US.win32.zip",
-        "buildername": "Windows XP 32-bit mozilla-inbound debug test mochitest-browser-chrome",
-        "buildid": "20130814102222",
-        "buildnumber": 811,
-        "builduid": "1b94f824176a48a59aeaf61252f53666",
-        "log_url": "http://ftp.mozilla.org/pub/mozilla.org/firefox/tinderbox-builds/mozilla-inbound-win32-debug/1376500942/mozilla-inbound_xp-ix-debug_test-mochitest-browser-chrome-bm69-tests1-windows-build811.txt.gz",
-        "master": "http://buildbot-master69.srv.releng.use1.mozilla.com:8201/",
-        "pgo_build": "False",
-        "platform": "win32",
-        "product": "firefox",
-        "project": "",
-        "repo_path": "integration/mozilla-inbound",
-        "repository": "",
-        "request_ids": [
-          28008611,
-          28008623
-        ],
-        "request_times": {
-          "28008611": 1376506672,
-          "28008623": 1376506672
-        },
-        "revision": "355dcff578b6",
-        "scheduler": "tests-mozilla-inbound-xp-ix-debug-unittest",
-        "script_repo_revision": "2f85052176f2",
-        "slavebuilddir": "test",
-        "slavename": "t-xp32-ix-017",
-        "stage_platform": "win32",
-        "symbols_url": "http://ftp.mozilla.org/pub/mozilla.org/firefox/tinderbox-builds/mozilla-inbound-win32-debug/1376500942/firefox-26.0a1.en-US.win32.crashreporter-symbols.zip"
-      },
-      "reason": "scheduler",
-      "request_ids": [
-        28008611,
-        28008623
-      ],
-      "requesttime": 1376506672,
-      "result": 0,
-      "slave_id": 4620,
-      "starttime": 1376506680
-    },
-    {
-      "builder_id": 120940,
-      "buildnumber": 751,
-      "endtime": 1376511927,
-      "id": 26553543,
-      "master_id": 94,
-      "properties": {
-        "basedir": "C:\\slave\\test",
-        "branch": "test_treeherder",
-        "build_url": "http://ftp.mozilla.org/pub/mozilla.org/firefox/tinderbox-builds/mozilla-inbound-win32-debug/1376500942/firefox-26.0a1.en-US.win32.zip",
-        "buildername": "WINNT 6.2 mozilla-inbound debug test mochitest-browser-chrome",
-        "buildid": "20130814102222",
-        "buildnumber": 751,
-        "builduid": "1b94f824176a48a59aeaf61252f53666",
-        "log_url": "http://ftp.mozilla.org/pub/mozilla.org/firefox/tinderbox-builds/mozilla-inbound-win32-debug/1376500942/mozilla-inbound_win8-debug_test-mochitest-browser-chrome-bm69-tests1-windows-build751.txt.gz",
-        "master": "http://buildbot-master69.srv.releng.use1.mozilla.com:8201/",
-        "pgo_build": "False",
-        "platform": "win32",
-        "product": "firefox",
-        "project": "",
-        "repo_path": "integration/mozilla-inbound",
-        "repository": "",
-        "request_ids": [
-          28008587,
-          28008599
-        ],
-        "request_times": {
-          "28008587": 1376506671,
-          "28008599": 1376506671
-        },
-        "revision": "355dcff578b6",
-        "scheduler": "tests-mozilla-inbound-win8-debug-unittest",
-        "script_repo_revision": "2f85052176f2",
-        "slavebuilddir": "test",
-        "slavename": "t-w864-ix-072",
-        "stage_platform": "win32",
-        "symbols_url": "http://ftp.mozilla.org/pub/mozilla.org/firefox/tinderbox-builds/mozilla-inbound-win32-debug/1376500942/firefox-26.0a1.en-US.win32.crashreporter-symbols.zip"
-      },
-      "reason": "scheduler",
-      "request_ids": [
-        28008587,
-        28008599
-      ],
-      "requesttime": 1376506671,
-      "result": 0,
-      "slave_id": 3909,
-      "starttime": 1376506687
-    }
-  ]
-}
