+++ tests/sample_data/builds-running-missing1.json
@@ -1,0 +1,36 @@
+{
+    "running": {
+        "test_treeherder": {
+            "45f8637cb9f7": [
+                {
+                    "submitted_at": 1369231311,
+                    "buildername": "WINNT 5.2 profiling build",
+                    "start_time": 1369231311,
+                    "number": 3,
+                    "claimed_by_name": "buildbot-master66.srv.releng.usw2.mozilla.com:/builds/buildbot/build1/master",
+                    "request_ids": [
+                        24526180
+                    ],
+                    "last_heartbeat": 1369231939,
+                    "id": 24767134,
+                    "revision": "45f8637cb9f7"
+                }
+            ],
+            "222222222222": [
+                {
+                    "submitted_at": 1369231312,
+                    "buildername": "WINNT 5.2 profiling build",
+                    "start_time": 1369231312,
+                    "number": 3,
+                    "claimed_by_name": "buildbot-master66.srv.releng.usw2.mozilla.com:/builds/buildbot/build1/master",
+                    "request_ids": [
+                        24526181
+                    ],
+                    "last_heartbeat": 1369231940,
+                    "id": 24767134,
+                    "revision": "222222222222aa71f463867b15d956a972ae7574"
+                }
+            ]
+        }
+    }
+}
