+++ treeherder/etl/tasks/__init__.py
@@ -1,0 +1,4 @@
+from .buildapi_tasks import *
+from .cleanup_tasks import *
+from .tbpl_tasks import *
+from .tasks import *
